### 2019-02-17 #1258 @whedon
[PRE REVIEW]: `tqdm`: A Fast, Extensible Progress Meter for Python and CLI
**Submitting author:** @casperdcl (<a href="http://orcid.org/0000-0002-7211-1557">da Costa-Luis, Casper O.</a>)
**Repository:** <a href="https://github.com/tqdm/tqdm" target ="_blank">https://github.com/tqdm/tqdm</a>
**Version:** v4.31.1
