### 2019-02-17 #1258 @labarba
@casperdcl — The `paper.md` file does not conform to the JOSS requirements. Please re-read and follow the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html) to prepare a `paper.md` with the needed metadata (the bibliography needs to go in a .bib file).

Very cool project… but for the purposes of a JOSS review, we will need your paper to address the research application of the software. We call this a "Statement of Need,"

> The authors should clearly state what problems the software is designed to solve and who the target audience is.

From the [Submission Requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):

> The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)
