### 2019-02-17 #1258 @casperdcl
@labarba thanks for your comments. I did indeed read the submission guidelines thoroughly and expected a discussion.

1. Does `paper.md` and/or `paper.bib` have to be located in the root of the repository? I am very hesitant to clutter the root (which is why I put the bib in-place. The guidelines did not explicitly forbid this). Ideally I can put all materials in a `paper` subfolder. Even better, I'd prefer to place them in a different repository (e.g. the documentation repo)
2. Does the `affiliations` filed have to be filled in? This work was not supported by any institution and thus there are no applicable affiliations. If this is required, perhaps "Independent" or "Non-affiliated" would suffice?
3. Browsing other accepted papers, I noticed a "statement of need" was not always explicitly given, but rather inferred from the summary. If you believe this paper needs it, I am happy to add one - do let me know.
4. If @whedon needs to be able to auto-complie this paper, could you provide details as to how I can replicate the paper build environment locally for testing? Alternatively I can provide my own build instructions.
