### 2019-02-17 #1258 @labarba
1. A folder is fine. A branch is fine. A different repo, no. (But you can delete after archive and publication.)
2. Need to put something in affiliation. Up to you what.
3. Statement of Need is not a required _heading_, but the content is. Your paper lacks the content.
4. The auto-build by `whedon` here is a requisite for publication. JOSS is able to be free thanks to many volunteers and extensive automation. You can attempt to build after pushing changes with the command `@whedon generate pdf` in a comment here.
