### 2019-02-17 b7dbddc
paper.md: split into separate paper.bib, move to examples/ subfolder
