+++ paper.md
@@ -1,502 +1,0 @@
----
-title: '`tqdm`: A Fast, Extensible Progress Meter for Python and CLI'
-tags:
- - progressbar
- - progressmeter
- - progress-bar
- - meter
- - rate
- - eta
- - console
- - terminal
- - time
- - progress
- - bar
- - gui
- - python
- - parallel
- - cli
- - utilities
-authors:
- - name: Casper O da Costa-Luis
-   orcid: 0000-0002-7211-1557
-date: 16 February 2019
-references:
- - id: tqdm-ar
-   type: book
-   author:
-   - family: Alī Ġūl
-     given: Maḥmūd
-   title: 'Early Southern Arabian Languages and Classical Arabic Sources: A Critical Examination of Literary and Lexicographical Sources by Comparison with the Inscriptions'
-   publisher: Yarmouk University
-   publisher-place: Irbid
-   issued:
-   - year: '1993'
- - id: tqdm-es
-   type: webpage
-   title: '¿Lenguaje sms que significa esto?'
-   URL: https://es.answers.yahoo.com/question/index?qid=20090405052137AAF2YBo&guccounter=1
-   author:
-   - literal: Yahoo Answers
- - id: pypi
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: Python Package Index (PyPI)
-   publisher: Python Software Foundation
-   title: tqdm
-   URL: https://pypi.org/project/tqdm/
- - id: conda
-   type: webpage
-   author:
-   - literal: Anaconda
-   issued:
-   - year: '2019'
-   title: '`tqdm` :: Anaconda Cloud'
-   URL: https://anaconda.org/conda-forge/tqdm
- - id: snapcraft
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: Snapcraft
-   title: 'Installing `tqdm` for Linux using the Snap Store'
-   URL: https://snapcraft.io/tqdm
- - id: zenodo
-   type: article
-   issued:
-   - year: '2019'
-   author:
-   - family: da Costa-Luis
-     given: Casper O.
-   - literal: '`tqdm` developers'
-   title: '`tqdm` stable'
-   publisher: Zenodo
-   DOI: 10.5281/zenodo.595120
- - id: stdout
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: Stack Overflow
-   title: Why is printing to stdout so slow? Can it be sped up?
-   URL: https://stackoverflow.com/questions/3857052/why-is-printing-to-stdout-so-slow-can-it-be-sped-up
- - id: pypi-downloads
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: Python Packaging Authority (PyPA)
-   publisher: Python Software Foundation
-   title: 'Analyzing PyPI package downloads — Python Packaging User Guide'
-   URL: https://packaging.python.org/guides/analyzing-pypi-package-downloads/
- - id: 'keras'
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - given: Ben
-   title: 'Keras integration with `tqdm` progress bars'
-   URL: https://github.com/bstriner/keras-tqdm
- - id: tqdm-results
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: GitHub
-   title: '`tqdm` Code Results'
-   URL: https://github.com/search?q=tqdm&type=Code
- - id: tqdm-dependents
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: GitHub
-   title: '`tqdm` dependents'
-   URL: https://github.com/tqdm/tqdm/network/dependents
- - id: lib-io
-   type: webpage
-   issued:
-   - year: '2019'
-   author:
-   - literal: Libraries.io
[…]
