+++ examples/paper.bib
@@ -1,0 +1,210 @@
+@book{tqdm-ar,
+ author="Maḥmūd Alī Ġūl",
+ title="Early Southern Arabian Languages and Classical Arabic Sources: A Critical Examination of Literary and Lexicographical Sources by Comparison with the Inscriptions",
+ publisher="Yarmouk University",
+ address="Irbid",
+ year="1993"
+}
+@misc{tqdm-es,
+ title="¿Lenguaje sms que significa esto?",
+ url="https://es.answers.yahoo.com/question/index?qid=20090405052137AAF2YBo&guccounter=1",
+ author="Yahoo Answers"
+}
+@misc{pypi,
+ year="2019",
+ author="{Python Package Index ({PyPI})}",
+ publisher="Python Software Foundation",
+ title="{tqdm}",
+ url="https://pypi.org/project/tqdm/"
+}
+@misc{conda,
+ author="Anaconda",
+ year="2019",
+ title="{tqdm} :: Anaconda Cloud",
+ url="https://anaconda.org/conda-forge/tqdm"
+}
+@misc{snapcraft,
+ year="2019",
+ author="Snapcraft",
+ title="Installing {tqdm} for Linux using the Snap Store",
+ url="https://snapcraft.io/tqdm"
+}
+@article{zenodo,
+ year="2019",
+ author="Casper O. {da Costa-Luis} and {{tqdm} developers}",
+ title="{tqdm} stable",
+ publisher="Zenodo",
+ doi="10.5281/zenodo.595120"
+}
+@misc{stdout,
+ year="2019",
+ author="{Stack Overflow}",
+ title="Why is printing to stdout so slow? Can it be sped up?",
+ url="https://stackoverflow.com/questions/3857052/why-is-printing-to-stdout-so-slow-can-it-be-sped-up"
+}
+@misc{pypi-downloads,
+ year="2019",
+ author="{Python Packaging Authority ({PyPA})}",
+ publisher="Python Software Foundation",
+ title="Analyzing {PyPI} package downloads -- Python Packaging User Guide",
+ url="https://packaging.python.org/guides/analyzing-pypi-package-downloads/"
+}
+@misc{keras,
+ year="2019",
+ author="Ben",
+ title="Keras integration with {tqdm} progress bars",
+ url="https://github.com/bstriner/keras-tqdm"
+}
+@misc{tqdm-results,
+ year="2019",
+ author="GitHub",
+ title="{tqdm} Code Results",
+ url="https://github.com/search?q=tqdm&type=Code"
+}
+@misc{tqdm-dependents,
+ year="2019",
+ author="GitHub",
+ title="{tqdm} dependents",
+ url="https://github.com/tqdm/tqdm/network/dependents"
+}
+@misc{lib-io,
+ year="2019",
+ author="Libraries.io",
+ title="{tqdm} on {PyPI}",
+ url="https://libraries.io/pypi/tqdm"
+}
+@misc{sourcerank,
+ year="2019",
+ author="Libraries.io",
+ title="SourceRank Breakdown for {tqdm}",
+ url="https://libraries.io/pypi/tqdm/sourcerank"
+}
+@misc{sourcerank-descending,
+ year="2019",
+ author="Libraries.io",
+ title="Libraries - The Open Source Discovery Service",
+ url="https://libraries.io/search?order=desc&platforms=PyPI&sort=rank"
+}
+@misc{stars,
+ year="2019",
+ author="GitHub",
+ title="{tqdm} Stargazers",
+ url="https://github.com/tqdm/tqdm/stargazers"
+}
+@misc{stars-hist,
+ year="2019",
+ author="{timqian}",
+ title="Star history",
+ url="https://timqian.com/star-history/#tqdm/tqdm"
+}
+@misc{trend-hist,
+ year="2018",
[…]

[…]
