### 2019-02-17 ec0b1f1
add affilitations key, tags, remove unicode, add missing year
