+++ examples/paper.bib
@@ -6,9 +6,10 @@
  year="1993"
 }
 @misc{tqdm-es,
+ year="2009",
  title="¿Lenguaje sms que significa esto?",
  url="https://es.answers.yahoo.com/question/index?qid=20090405052137AAF2YBo&guccounter=1",
- author="Yahoo Answers"
+ author="{Yahoo Answers}"
 }
 @misc{pypi,
  year="2019",
