### 2019-02-17 #1258 @casperdcl
- [x] affilitiations
- [ ] statement of need
- [x] fix generate issues
