+++ examples/paper.bib
@@ -184,6 +184,30 @@
  url="https://arxiv.org/abs/1604.00896",
  issn="2049-9647"
 }
+@article{moriwaki,
+ title={Mordred: a molecular descriptor calculator},
+ author={Moriwaki, Hirotomo and Tian, Yu-Shi and Kawashita, Norihito and Takagi, Tatsuya},
+ doi={10.1186/s13321-018-0258-y},
+ number={1},
+ volume={10},
+ month={February},
+ year={2018},
+ journal={Journal of cheminformatics},
+ issn={1758-2946},
+ pages={4}
+}
+@article{jackson,
+ title={3D for the people: multi-camera motion capture in the field with consumer-grade cameras and open source software},
+ author={Jackson, Brandon E and Evangelista, Dennis J and Ray, Dylan D and hedrick, Tyson L},
+ doi={10.1242/bio.018713},
+ number={9},
+ volume={5},
+ month={September},
+ year={2016},
+ journal={Biology open},
+ issn={2046-6390},
+ pages={1334--1342}
+}
 @misc{travis,
  year="2019",
  author="{Travis {CI}}",
