### 2019-02-17 #1258 @cMadan
@casperdcl, thank you for submitting tdqm for publication in JOSS. I have used tdqm myself and recommended it to others and think it's a great package. That said, I'm not convinced that it's within the scope for publication in JOSS, but perhaps you can make the relevance more clear.

From the submission requirements (https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):
> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar.

I am unsure if tdqm fits this criteria. It's definitely a useful utility in developing research software, but do you think it fits this criteria? For instance, do you think others will cite their use of tdqm in research papers?
