### 2019-02-17 #1258 @casperdcl
Thanks @cMadan - it's on the first page out of >100k packages on https://libraries.io/search?languages=Python&order=desc&sort=rank, and as per JOSS requirements I've listed some peer-reviewed publications which explicitly list `tqdm` (see the *Popularity* section in the paper). It can be invaluable when debugging, say, learning rates of a machine learning framework. At 5M downloads per month (see below) I don't know how it compares to other JOSS software.
