### 2019-02-20 #1258 @cMadan
Thanks for your response, @casperdcl! I looked through some other JOSS submissions and there are other instances of software that is more intended towards developers than a specific application/group of users, so I agree that it is sufficiently within-scope.

Do you have any suggestions for reviewers?
