### 2019-02-20 #1258 @casperdcl
Well since it's so widely applicable as you've noted, I don't have any reviewer preference. Just someone who uses python regularly!
