### 2019-02-21 6b463f5
prettify and unify contributors/maintainers/authors
