### 2019-02-23 c919eab
Drop support for EOL Python 2.6 and 3.3
