+++ tox.ini
@@ -4,8 +4,8 @@
 # and then run "tox" from this directory.
 
 [tox]
-# deprecation warning: py{26,32,33,34}
-envlist = py{26,27,33,34,35,36,37,py,py3}, flake8, setup.py, perf
+# deprecation warning: py34
+envlist = py{27,34,35,36,37,py,py3}, flake8, setup.py, perf
 
 [coverage]
 deps =
@@ -35,30 +35,13 @@ deps =
     pandas
 commands = {[extra]commands}
 
-# no cython/numpy/pandas for py{py,py3,26,33,34}
-
-[testenv:py26]
-# no codecov and timer for py26
-deps =
-    nose
-    coverage
-    coveralls==1.2.0
-    codecov
-    pycparser==2.18
-    idna==2.7
-commands =
-    {[coverage]commands}
-    codecov
+# no cython/numpy/pandas for py{py,py3,34}
 
 [testenv:pypy]
 deps = {[extra]deps}
 commands = {[extra]commands}
 
 [testenv:pypy3]
-deps = {[extra]deps}
-commands = {[extra]commands}
-
-[testenv:py33]
 deps = {[extra]deps}
 commands = {[extra]commands}
 
