### 2019-02-23 d351a56
Upgrade Python syntax with pyupgrade
