### 2019-02-24 #1277 @casperdcl
@cMadan @GregaVrbancic, @Benjamin-Lee please use the [`tqdm:devel`](https://github.com/tqdm/tqdm/tree/devel) for review, I can commit any suggested changes there and make a new release before (hopeful) acceptance
