### 2019-03-01 35f2f77
add binder, tidy notebooks.ai links
closes #679
