### 2019-03-01 2660a0a
minor README italicisation
