+++ .readme.rst
@@ -10,8 +10,8 @@ tqdm
 |DOI-URI| |LICENCE| |OpenHub-Status| |binder-demo| |notebook-demo|
 
 
-``tqdm`` means "progress" in Arabic (taqadum, تقدّم)
-and is an abbreviation for "I love you so much" in Spanish (te quiero demasiado).
+``tqdm`` means "progress" in Arabic (*taqadum*, تقدّم)
+and is an abbreviation for "I love you so much" in Spanish (*te quiero demasiado*).
 
 Instantly make your loops show a smart progress meter - just wrap any
 iterable with ``tqdm(iterable)``, and you're done!
