+++ .travis.yml
@@ -13,8 +13,24 @@ matrix:
     env: TOXENV=py36
   - python: 3.7
     dist: xenial
-    sudo: true
+    sudo: true  # required for py37
     env: TOXENV=py37
+    before_deploy:
+      - pip install py-make
+      - pymake build  # releases
+      #- pip install .[dev]
+      #- pymake submodules
+      #- cd wiki && pymake && cd ..
+    deploy:
+      - provider: releases
+        api_key: $GITHUB_TOKEN
+        file_glob: true
+        file: dist/tqdm-*.whl*
+        skip_cleanup: true
+        draft: true
+        name: "tqdm v$TRAVIS_TAG stable"
+        on:
+          tags: true
   - python: pypy2.7-5.10.0
     env: TOXENV=pypy
   - python: pypy3.5-5.10.0
