### 2019-03-01 f0129b5
update deployment dev docs
