### 2019-03-02 f1c722d
add snap:edge channel
