+++ .travis.yml
@@ -17,9 +17,9 @@ matrix:
     env: TOXENV=py37
     before_deploy:
       - pip install .[dev]
-      - pymake build
-      #- pymake submodules
-      #- cd wiki && pymake && cd ..
+      - make build
+      #- make submodules
+      #- cd wiki && make && cd ..
       - openssl aes-256-cbc -K $encrypted_a6d6301302b7_key
         -iv $encrypted_a6d6301302b7_iv -in .tqdm.gpg.enc -out .tqdm.gpg -d
       - gpg --import .tqdm.gpg
