### 2019-03-02 eec3188
fix travis deploy snapcraft
