### 2019-03-02 4c22a38
strip whitespace
