### 2019-03-02 d98871e
add docker documentation
