### 2019-03-03 c52d7af
add travis deploy docker devel
