### 2019-03-03 d2265fe
TODO: revert before merge: travis snap build branch
