+++ .travis.yml
@@ -158,7 +158,7 @@ jobs:
       channel: edge
       skip_cleanup: true
       on:
-        branch: devel
+        branch: snap
 before_install:
 # fix a crash with multiprocessing on Travis
 # - sudo rm -rf /dev/shm
