### 2019-03-04 #1277 @GregaVrbancic
So, here is my review of this submission. In general, the package seems really nice, but in order to publish this paper I have some minor suggestions or remarks:

1. Stephen Larroque appears to be a significant contributor. Have you considered adding him in the authors' list?
2. I have read the discussion about the scope of the paper and the target audience in #1258. My opinion is that the target audience could be more explicitly defined, but if @cMadan is fine with the current state, I will not insist on making any changes.
3. In the references list, the reference with id tqdm-ar has release year and publisher defined differently as [here](https://eprints.soas.ac.uk/29058/).
4. The iterable-based usage example featuring trange is not working out of the box. It would be nice if the example would also have the import statement for trange defined.
5. Some guidelines on reporting issues or seeking support could be added.

@casperdcl please take a look at the listed remarks.

@cMadan I have a question regarding the `whedon check references` command. Is it possible to run it against the devel branch? Maybe in the same manner as `whedon generate pdf from branch devel`?
