### 2019-03-04 #1277 @casperdcl
thanks @GregaVrbancic; will respond in more detail later.

1. Stephen (@lrq3000), I know you haven't contributed in years but would be great to have you back on the team
2. Any suggestions?
3. Will check
4. Will update
5. Will add links to relevant documentation
