### 2019-03-04 #1277 @Benjamin-Lee
Initial thoughts on the paper:

1. If @arfon and @cMadan are happy with `tqdm` being in JOSS's scope, then I have no issues. To my mind, the more peer reviewed published software, the better.
1. The structure of the paper requires work. Specifically, I would move the popularity section to the introduction and convert it to prose from bullets and eliminate the "see also" and "references in blogs" sections. JOSS papers generally use a single "summary" section. I would suggest attempting to do the same if at all possible. If you think you can't make it work in one section, [this paper](https://doi.org/10.21105/joss.01141) is a very good example of a well structured but complex JOSS paper. In general, I would suggest collapsing the bullets to paragraphs (with the exception of the first set in the summary, which I think works well)
1. I don't think there's any need for the logo in the paper.
1. The third sentence, explaining the meaning of the `tqdm`, is unnecessary.
1. The bibliography seems to have a lot of redundant information. To start, I would suggest collapsing the GitHub references down to one.
1. In the license section, I would remove the wording about citing tqdm on Zenodo since it'll be cited in JOSS!

Overall, `tqdm` is a great tool with wide applicability to both research and general software development. The repo is quite well organized and the code well documented. If I were to choose an area for improvement in the repo (though there isn't really much), I would suggest more developer documentation describing the structure of the `tqdm/tqdm` directory, since some of the files such as `_main.py` `__main__.py` are similarly named.
