### 2019-03-05 59efd5b
paper: respond to reviews
https://github.com/openjournals/joss-reviews/issues/1277
