+++ examples/paper.bib
@@ -1,9 +1,8 @@
-@book{tqdm-ar,
+@phdthesis{tqdm-ar,
  author="Maḥmūd Alī Ġūl",
  title="Early Southern Arabian Languages and Classical Arabic Sources: A Critical Examination of Literary and Lexicographical Sources by Comparison with the Inscriptions",
- publisher="Yarmouk University",
- address="Irbid",
- year="1993"
+ school="{SOAS} University of London",
+ year="1963"
 }
 @misc{tqdm-es,
  year="2009",
@@ -26,7 +25,13 @@
 }
 @misc{snapcraft,
  year="2019",
- author="Snapcraft",
+ author="{Docker Inc.}",
+ title="{tqdm}/{tqdm} - Docker Hub",
+ url="https://hub.docker.com/r/tqdm/tqdm"
+}
+@misc{docker,
+ year="2019",
+ author="Docker",
  title="Installing {tqdm} for Linux using the Snap Store",
  url="https://snapcraft.io/tqdm"
 }
@@ -36,6 +41,18 @@
  title="{tqdm} stable",
  publisher="Zenodo",
  doi="10.5281/zenodo.595120"
+}
+@misc{notebooks,
+ year="2019",
+ author="{Notebooks {AI}}",
+ title="{tqdm}",
+ url="https://notebooks.ai/demo/gh/tqdm/tqdm"
+}
+@misc{binder,
+ year="2019",
+ author="{Binder (beta)}",
+ title="{tqdm}",
+ url="https://mybinder.org/v2/gh/tqdm/tqdm/master?filepath=DEMO.ipynb"
 }
 @misc{stdout,
  year="2019",
