+++ examples/paper.md
@@ -29,9 +29,8 @@ affiliations:
 date: 16 February 2019
 bibliography: paper.bib
 ---
-![](../logo.png)
 
-# Summary
+# Introduction
 
 **`tqdm`** is a progress bar library designed to be fast and extensible. It is
 written in Python, though ports in other languages are available. `tqdm` means
@@ -62,10 +61,6 @@ in the most demanding of scenarios.
 
 Exhaustive documentation may be found on the project's [home
 page](https://github.com/tqdm/tqdm/#documentation).
-
-The package supports both Python versions 2 and 3, and is available for download
-via `conda` [@conda], `pip` [@pypi], `snap` [@snapcraft], and *Zenodo*
-[@zenodo].
 
 The two basic use cases are within Python code and within a Command-line
 interface:
@@ -116,62 +111,50 @@ $ cat *.txt | python3 -m tqdm --unit loc --unit_scale --total 1075075 
 1075075
 ```
 
-# Popularity
+# Availability
 
-As of January 2019, `tqdm` has received:
+The package supports both Python versions 2 and 3, and is available for download
+via `conda` [@conda], `pip` [@pypi], `snap` [@snapcraft], `docker` [@docker],
+and *Zenodo* [@zenodo].
+Web-based Jupyter interactive demonstrations are also available
+[@notebooks;@binder]
 
-- over 20 million downloads [@pypi-downloads];
-- over 315 thousand code inclusions [@tqdm-results];
-  * over 23 thousand dependent repositories [@tqdm-dependents];
-  * 7 thousand dependent libraries [@lib-io];
-- a SourceRank of 22 [@sourcerank], in the world's top 20 Python packages as of
-  early 2019 [@sourcerank-descending];
-- 9 thousand stars on GitHub [@stars;@stars-hist], and the top trending
-  repository during a period in December 2015 [@trend-hist];
-- 500 thousand documentation hits [@hits];
-- usage in several textbooks [@miller;@boxel;@nandy] and peer-reviewed
-  scientific publications
-  [@stein;@cook;@madhikar;@palmer;@knight;@moriwaki;@jackson].
+Unit tests are run at least weekly on cloud-based continuous integration
+[@travis], with code style and security issues checked on
+[Codacy](https://app.codacy.com/project/tqdm/tqdm/dashboard) [@code-review].
+Coverage is reported on [Coveralls](https://coveralls.io/github/tqdm/tqdm) and
+[Codecov](https://codecov.io/gh/tqdm/tqdm), and performance is monitored against
+regression [@asv].
 
-## References in Blogs and Public Media
+# Impact
 
-- A `tqdm` release becomes [Zenodo's 1 millionth
-  record](https://twitter.com/WikimediaItalia/status/914448810117545985)
-- [A Hymn to Progress](http://www.metafilter.com/161265/An-alternative-to-an-eternally-spinning-wheel#6644017),
-  a poem or song with suggested tune of *For those in Peril on the C*,
-  where "C" is a pun on *Sea* and the *C programming language*
-- [My top 5 'new' Python modules of
-  2015](http://blog.rtwilson.com/my-top-5-new-python-modules-of-2015/)
-- [`tqdm`, `imageio` and `Seaborn`: Three essential python modules
-  (Nov 2018)](https://dominikschmidt.xyz/python-modules-tqdm-imageio-seaborn/)
-- <https://pythontips.com/2018/06/03/top-14-most-famous-python-libraries-frameworks>
-- <https://github.com/IliaGavrilov/Machine_Learning_libraries_and_Python_3.6_tricks>
-- <https://github.com/robclewley/TqdmAudioRicker>
+As of January 2019, `tqdm` has accumulated over 20 million downloads
[…]
