### 2019-03-06 #1277 @GregaVrbancic
@casperdcl please consider the following proposed improvements.

> 2. > I have read the discussion about the scope of the paper and the target audience in #1258. My opinion is that the target audience could be more explicitly defined, but if @cMadan is fine with the current state, I will not insist on making any changes.
>
> * Any suggestions as to how the audience could be more explicitly defined? On [![PyPI](https://camo.githubusercontent.com/91df3b8889b8102db255bfbeaefc87c713c4e181/68747470733a2f2f696d672e736869656c64732e696f2f62616467652f507950492d7471646d2d626c75652e7376673f6c6f676f3d707974686f6e)](https://pypi.org/project/tqdm/), the intended audience is listed (based on https://pypi.org/pypi?%3Aaction=list_classifiers) - I can mirror this in the paper if you'd like.

In my opinion, a brief description of  the target audience (based on https://pypi.org/pypi?%3Aaction=list_classifiers) would be useful and would improve the overall quality of the paper. Keep in mind that not all of the JOSS audience will go through the classifiers listed on PyPI package site.

> 5. > Some guidelines on reporting issues or seeking support could be added.
>
> * Does this comment refer to the repository or the paper itself?
>   If it's the repository, there's https://github.com/tqdm/tqdm/#contributions or https://tqdm.github.io/contributing/.

Yes, I was refering to the repository. Look at the review checklist, secton Documentation, subsection Community guidelines. You have adressed the the 1) nicely in the documentation, but I was aiming more toward the points 2) and 3), or have I missed something?

My other remarks you've addressed properly in my opinion.
