### 2019-03-06 73962a4
tqdm: fix boolean casting to work for sizeless self.iterable
