### 2019-03-06 fa57ecc
paper: misc fixes
- response to https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-470007598
