+++ examples/paper.bib
@@ -23,15 +23,15 @@
  title="{tqdm} :: Anaconda Cloud",
  url="https://anaconda.org/conda-forge/tqdm"
 }
-@misc{snapcraft,
+@misc{docker,
  year="2019",
  author="{Docker Inc.}",
  title="{tqdm}/{tqdm} - Docker Hub",
  url="https://hub.docker.com/r/tqdm/tqdm"
 }
-@misc{docker,
+@misc{snapcraft,
  year="2019",
- author="Docker",
+ author="Snapcraft",
  title="Installing {tqdm} for Linux using the Snap Store",
  url="https://snapcraft.io/tqdm"
 }
@@ -50,7 +50,7 @@
 }
 @misc{binder,
  year="2019",
- author="{Binder (beta)}",
+ author="Binder",
  title="{tqdm}",
  url="https://mybinder.org/v2/gh/tqdm/tqdm/master?filepath=DEMO.ipynb"
 }
