### 2019-03-06 #1277 @casperdcl
Thanks @GregaVrbancic. I've update my response above (but also explicit the changes below):

2. > In my opinion, a brief description of the target audience (based on https://pypi.org/pypi?%3Aaction=list_classifiers) would be useful and would improve the overall quality of the paper. Keep in mind that not all of the JOSS audience will go through the classifiers listed on PyPI package site.
- Thanks,  I have appended a description to the Introduction.
5. > Yes, I was refe[r]ring to the repository. Look at the review checklist, sect[i]on Documentation, subsection Community guidelines. You have ad[d]ressed the the 1) nicely in the documentation, but I was aiming more toward the points 2) and 3), or have I missed something?
I have now updated the latter to point to all issues (including closed ones and PRs), and added both links to the website (https://tqdm.github.io/) now so they're harder to miss.
N.B. anyone attempting to file a new issue/PR is presented with a message:
    > - [ ] I have visited the [source website], and in particular
    >   read the [known issues]
    > - [ ] I have searched through the [issue tracker] for duplicates
    > - [ ] I have mentioned version numbers, operating system and
    >   environment, where applicable:
    >   ```python
    >   import tqdm, sys
    >   print(tqdm.__version__, sys.version, sys.platform)
    >   ```
    > - [ ] If applicable, I have mentioned the relevant/related issue(s)

[source website]: https://github.com/tqdm/tqdm/
[known issues]: https://github.com/tqdm/tqdm/#faq-and-known-issues
[issue tracker]: https://github.com/tqdm/tqdm/issues?q=
