### 2019-03-10 #1277 @cMadan
@casperdcl @GregaVrbancic @Benjamin-Lee, just to follow-up re: scope of work and JOSS. I appreciate both reviewers commenting on this--I do agree that the project is on the edge of JOSS' scope (and had brought this up in the pre-review). However, we have published tools that are similarly tools for researchers/development (rather than researcher end-users) and tdqm has clearly demonstrated it's utility as a development tool, so we are considering it within-scope here. I think the statement of need provided in the paper is sufficient, but suggestions to improve it are, of course, welcome.
