### 2019-03-16 d230a55
minor potential fix snapcraft.yml with numeric commit ID
also add `[python setup.py] make run`
