### 2019-03-17 d1187ac
Workaround lack of multiprocessing on Jython
