### 2019-03-21 c2a8b1b
Fix 'Set changed size during iteration' bug
