+++ tqdm/_tqdm.py
@@ -481,8 +481,9 @@ class tqdm(Comparable):
     @classmethod
     def _get_free_pos(cls, instance=None):
         """Skips specified instance"""
-        positions = set(abs(inst.pos) for inst in cls._instances
-                        if inst is not instance and hasattr(inst, "pos"))
+        with cls._lock:
+            positions = set(abs(inst.pos) for inst in cls._instances
+                            if inst is not instance and hasattr(inst, "pos"))
         return min(set(range(len(positions) + 1)).difference(positions))
 
     @classmethod
@@ -932,11 +933,10 @@ class tqdm(Comparable):
 
         # if nested, at initial sp() call we replace '\r' by '\n' to
         # not overwrite the outer progress bar
-        with self._lock:
-            if position is None:
-                self.pos = self._get_free_pos(self)
-            else:  # mark fixed positions as negative
-                self.pos = -position
+        if position is None:
+            self.pos = self._get_free_pos(self)
+        else:  # mark fixed positions as negative
+            self.pos = -position
 
         if not gui:
             # Initialize the screen printer
