### 2019-04-17 58f9f1c
Add missing isatty() method to DummyTqdmFile
