+++ examples/redirect_print.py
@@ -32,6 +32,9 @@ class DummyTqdmFile(object):
     def flush(self):
         return getattr(self.file, "flush", lambda: None)()
 
+    def isatty(self):
+        return getattr(self.file, "isatty", lambda: False)()
+
 
 @contextlib.contextmanager
 def std_out_err_redirect_tqdm():
