### 2019-04-17 e52835d
Fix usage docstrings
- replace `tqdm_notebook` with `tqdm.tnrange()` (or `._tqdm_notebook`)
- replace `tqdm_gui` with `tqdm.tgrange` (or `._tqdm_gui`)
- closes #714
- closes #715
