### 2019-04-17 bc96fd5
Implement fixed width bar [bar_width] and chop overflow of ncols
- added bar_width to initializer and format_meter that sets a fixed
  bar width that never changes (unless:)
- added better ncols handling for overflows on l_bar, bar, r_bar
  specifically, output is restricted to one line (of length ncols)
  and will be chopped to ensure that is always true
- added tests

This fixes #623
This fixes #690
This helps prevent the spillover seen in #630
