### 2019-04-19 #1277 @Benjamin-Lee
One last tiny change and then I think it's good to go:

> within a Command-line interface

Should this be `CLI` (since you defined the acronym in the previous section) or `within a command-line interface` with a lowercase `c`>

Either way, congrats @casperdcl!
