### 2019-04-22 c1eec57
minor paper update
- fixes https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-484753574
