### 2019-04-29 108eb6c
Fixes: #723, #722, #721
