### 2019-05-07 a545c98
Unwrap fps before comparing them
This fixes issue #727 by unwrapping file objects before comparing them
for equality. On Python 2.7, this makes text output look reasonable
using the example scripts.
