### 2019-05-07 8ee73bc
Use ANSI escape to clear to end of line
This switches from trying to remember the length of the previous output,
and instead uses an ANSI sequence to clear to the end of the line.

This fixes issue #728
