+++ tqdm/_tqdm.py
@@ -243,12 +243,8 @@ class tqdm(Comparable):
             fp.write(_unicode(s))
             fp_flush()
 
-        last_len = [0]
-
         def print_status(s):
-            len_s = len(s)
-            fp_write('\r' + s + (' ' * max(last_len[0] - len_s, 0)))
-            last_len[0] = len_s
+            fp_write("\r" + s + "\x1b[K")
 
         return print_status
 
