### 2019-05-09 4ed04a7
Provide itertools wrappers that know about their length.
