### 2019-05-10 0f0c1a6
add and update documentation
- example of dynamic usage (#735, #545, #547, #432, #374)
- note writing issues #737
- update badges
