### 2019-05-10 140de47
support negative update
- fixes #545
- fixes #432
