### 2019-05-10 c7015f4
add a reset() function
- fixes #545 (https://github.com/tqdm/tqdm/issues/545#issuecomment-471090550)
