### 2019-05-13 #1277 @cMadan
@casperdcl, everything looks good and I'm almost ready to accept. Let me know after you merge into master and have the new GitHub release and new Zenodo DOI minted, then I'll finalise things here.
