### 2019-05-13 0e1961c
documentation, tidy, and function reordering
