+++ tqdm/_tqdm.py
@@ -480,7 +480,7 @@ class tqdm(Comparable):
 
     @classmethod
     def _get_free_pos(cls, instance=None):
-        """Skips specified instance"""
+        """Skips specified instance."""
         positions = set(abs(inst.pos) for inst in cls._instances
                         if inst is not instance and hasattr(inst, "pos"))
         return min(set(range(len(positions) + 1)).difference(positions))
@@ -517,9 +517,7 @@ class tqdm(Comparable):
 
     @classmethod
     def write(cls, s, file=None, end="\n", nolock=False):
-        """
-        Print a message via tqdm (without overlap with bars)
-        """
+        """Print a message via tqdm (without overlap with bars)."""
         fp = file if file is not None else sys.stdout
         with cls.external_write_mode(file=file, nolock=nolock):
             # Write the message
@@ -965,21 +963,6 @@ class tqdm(Comparable):
     def __del__(self):
         self.close()
 
-    @property
-    def format_dict(self):
-        """Public API for read-only member access"""
-        return dict(
-            n=self.n, total=self.total,
-            elapsed=self._time() - self.start_t
-            if hasattr(self, 'start_t') else 0,
-            ncols=self.dynamic_ncols(self.fp)
-            if self.dynamic_ncols else self.ncols,
-            prefix=self.desc, ascii=self.ascii, unit=self.unit,
-            unit_scale=self.unit_scale,
-            rate=1 / self.avg_time if self.avg_time else None,
-            bar_format=self.bar_format, postfix=self.postfix,
-            unit_divisor=self.unit_divisor)
-
     def __repr__(self):
         return self.format_meter(**self.format_dict)
 
@@ -1152,9 +1135,7 @@ class tqdm(Comparable):
                 self.last_print_t = cur_t
 
     def close(self):
-        """
-        Cleanup and (if leave=False) close the progressbar.
-        """
+        """Cleanup and (if leave=False) close the progressbar."""
         if self.disable:
             return
 
@@ -1195,13 +1176,44 @@ class tqdm(Comparable):
                 if not pos:
                     fp_write('\r')
 
+    def clear(self, nolock=False):
+        """Clear current bar display."""
+        if self.disable:
+            return
+
+        if not nolock:
+            self._lock.acquire()
+        self.moveto(abs(self.pos))
+        self.sp('')
+        self.fp.write('\r')  # place cursor back at the beginning of line
+        self.moveto(-abs(self.pos))
+        if not nolock:
+            self._lock.release()
+
+    def refresh(self, nolock=False):
+        """Force refresh the display of this bar."""
+        if self.disable:
+            return
+
+        if not nolock:
+            self._lock.acquire()
+        self.display()
+        if not nolock:
+            self._lock.release()
+
     def unpause(self):
-        """
-        Restart tqdm timer from last print time.
-        """
+        """Restart tqdm timer from last print time."""
         cur_t = self._time()
         self.start_t += cur_t - self.last_print_t
         self.last_print_t = cur_t
+
+    def reset(self, total=None):
+        """Resets a bar to n=0 for repeated use."""
+        self.last_print_n = self.n = 0
+        self.last_print_t = self.start_t = self._time()
+        if total is not None:
+            self.total = total
+        self.refresh()
 
     def set_description(self, desc=None, refresh=True):
         """
@@ -1218,9 +1230,7 @@ class tqdm(Comparable):
             self.refresh()
 
     def set_description_str(self, desc=None, refresh=True):
-        """
-        Set/modify description without ': ' appended.
-        """
+        """Set/modify description without ': ' appended."""
         self.desc = desc or ''
         if refresh:
             self.refresh()
[…]
