### 2019-05-13 e2bd4eb
better reset() docs
