+++ mkdocs.py
@@ -20,14 +20,20 @@ name  : type, optional
 """
 
 
-def doc2rst(doc, arglist=True):
+def doc2rst(doc, arglist=True, raw=False):
     """
     arglist  : bool, whether to create argument lists
+    raw  : bool, ignores arglist and indents by 2 spaces
     """
-    doc = dedent(doc).replace('`', '``')
-    if arglist:
-        doc = '\n'.join([i if not i or i[0] == ' ' else '* ' + i + '  '
-                         for i in doc.split('\n')])
+    doc = doc.replace('`', '``')
+    if raw:
+        doc = doc.replace('\n ', '\n   ')
+        #doc = '\n'.join(i.rstrip() for i in doc.split('\n'))
+    else:
+        doc = dedent(doc)
+        if arglist:
+            doc = '\n'.join([i if not i or i[0] == ' ' else '* ' + i + '  '
+                             for i in doc.split('\n')])
     return doc
 
 
@@ -42,6 +48,11 @@ DOC_tqdm_init_args = DOC_tqdm_init.partition(doc2rst(HEAD_ARGS))[-1]\
 DOC_tqdm_init_args, _, DOC_tqdm_init_rets = DOC_tqdm_init_args\
     .partition(doc2rst(HEAD_RETS))
 DOC_cli = doc2rst(tqdm._main.CLI_EXTRA_DOC).partition(doc2rst(HEAD_CLI))[-1]
+DOC_tqdm_tqdm = {}
+for i in dir(tqdm.tqdm):
+    doc = getattr(tqdm.tqdm, i).__doc__
+    if doc:
+        DOC_tqdm_tqdm[i] = doc2rst(doc, raw=True)
 
 # special cases
 DOC_tqdm_init_args = DOC_tqdm_init_args.replace(' *,', ' ``*``,')
@@ -51,6 +62,8 @@ README_rst = README_rst.replace('{DOC_tqdm}', DOC_tqdm)\
     .replace('{DOC_tqdm.tqdm.__init__.Parameters}', DOC_tqdm_init_args)\
     .replace('{DOC_tqdm._main.CLI_EXTRA_DOC}', DOC_cli)\
     .replace('{DOC_tqdm.tqdm.__init__.Returns}', DOC_tqdm_init_rets)
+for k, v in DOC_tqdm_tqdm.items():
+    README_rst = README_rst.replace('{DOC_tqdm.tqdm.%s}' % k, v)
 
 if __name__ == "__main__":
     fndoc = path.join(src_dir, 'README.rst')
