+++ .readme.rst
@@ -308,92 +308,39 @@ Returns
 .. code:: python
 
       def update(self, n=1):
-          """
-          Manually update the progress bar, useful for streams
-          such as reading files.
-          E.g.:
-          >>> t = tqdm(total=filesize) # Initialise
-          >>> for current_buffer in stream:
-          ...    ...
-          ...    t.update(len(current_buffer))
-          >>> t.close()
-          The last line is highly recommended, but possibly not necessary if
-          ``t.update()`` will be called in such a way that ``filesize`` will be
-          exactly reached and printed.
-
-          Parameters
-          ----------
-          n  : int, optional
-              Increment to add to the internal counter of iterations
-              [default: 1].
-          """
+          """{DOC_tqdm.tqdm.update}"""
 
       def close(self):
-          """Cleanup and (if leave=False) close the progressbar."""
+          """{DOC_tqdm.tqdm.close}"""
 
       def clear(self, nomove=False):
-          """Clear current bar display."""
+          """{DOC_tqdm.tqdm.clear}"""
 
       def refresh(self):
-          """Force refresh the display of this bar."""
+          """{DOC_tqdm.tqdm.refresh}"""
 
       def unpause(self):
-          """Restart tqdm timer from last print time."""
+          """{DOC_tqdm.tqdm.unpause}"""
 
       def reset(self, total=None):
-          """
-          Resets to 0 iterations for repeated use.
-
-          Consider combining with ``leave=True``.
-
-          Parameters
-          ----------
-          total  : int, optional. Total to use for the new bar.
-          """
+          """{DOC_tqdm.tqdm.reset}"""
 
       def set_description(self, desc=None, refresh=True):
-          """
-          Set/modify description of the progress bar.
-
-          Parameters
-          ----------
-          desc  : str, optional
-          refresh  : bool, optional
-              Forces refresh [default: True].
-          """
+          """{DOC_tqdm.tqdm.set_description}"""
 
       def set_postfix(self, ordered_dict=None, refresh=True, **kwargs):
-          """
-          Set/modify postfix (additional stats)
-          with automatic formatting based on datatype.
-
-          Parameters
-          ----------
-          refresh  : bool, optional
-              Forces refresh [default: True].
-          """
+          """{DOC_tqdm.tqdm.set_postfix}"""
 
       @classmethod
       def write(cls, s, file=sys.stdout, end="\n"):
-          """Print a message via tqdm (without overlap with bars)."""
+          """{DOC_tqdm.tqdm.write}"""
 
       @property
       def format_dict(self):
-          """Public API for read-only member access."""
+          """{DOC_tqdm.tqdm.format_dict}"""
 
       def display(self, msg=None, pos=None):
-          """
-          Use ``self.sp`` to display ``msg`` in the specified ``pos``.
-
-          Consider overloading this function when inheriting to use e.g.:
-          ``self.some_frontend(**self.format_dict)`` instead of ``self.sp``.
-
-          Parameters
-          ----------
-          msg  : str, optional. What to display (default: ``repr(self)``).
-          pos  : int, optional. Position to ``moveto``
-            (default: ``abs(self.pos)``).
-          """
+          """{DOC_tqdm.tqdm.display}"""
 
     def trange(*args, **kwargs):
         """
