### 2019-05-13 #1277 @casperdcl
@cMadan all done: merged, tagged (v4.32.0) and released on all channels (GitHub/Zenodo/PyPI/DockerHub/Snapcraft). Just waiting for Conda.
