### 2019-05-13 c901886
minor auto-deploy 'v' prefix strip
