### 2019-05-13 #1277 @casperdcl
note that this is more recent than the version mentioned in the original (https://github.com/openjournals/joss-reviews/issues/1277#issue-413844605)
