### 2019-05-13 #1277 @casperdcl
correction just released v4.32.1 with a minor bugfix
