### 2019-05-14 #1277 @cMadan
@openjournals/joss-eics, I think we're all set to accept here!
