### 2019-05-14 #1277 @labarba
👋  @casperdcl — Could you edit the metadata of the Zenodo deposit, so the title and author list matches the paper? You may also want to add your ORCID.
