### 2019-05-16 71fe001
tidy bool() and tests
