### 2019-05-17 #1277 @arfon
@casperdcl - your paper is now accepted into JOSS and the DOI is https://doi.org/10.21105/joss.01277 :zap::rocket::boom:
