### 2019-05-18 #1277 @casperdcl
thanks @cMadan @arfon - the metadata associated witht he DOI doesn't actually match the submission well, could you update it? Specifically the bibtex fields:

```
author={{da Costa-Luis}, Casper O.},
title={{tqdm}: A Fast, Extensible Progress Meter for Python and {CLI}},
```
