### 2019-05-18 #1277 @arfon
@casperdcl - I've updated the author metadata in the Crossref submission, those changes should be corrected in the next few hours.
