### 2019-05-19 f3ad68e
add JOSS DOI, minor documentation code fix
