### 2019-06-05 86dfc61
Fix race for tqdm._instances between __init__ and other methods
