### 2019-06-09 caa7065
Clear _closed_taken_positions when closing
