### 2019-06-09 091cc3a
Only refresh bars that changed position
