### 2019-06-11 5a980e1
Implements platform-specific progress reporters.
* a common interface for them
* Taskbar Extensions for Windows
* Progresses in trays using FreeDesktop (both v2 and v3 are implemented) and similar DBus interfaces
* progresses in taskbar buttons using Unity DBus interface

Closes: #552, #442
