### 2019-06-11 a27cd4e
Implements platform-specific progress reporters.
* a common interface for them
* Taskbar Extensions for Windows
* Progresses in trays using FreeDesktop (both v2 and v3 are implemented) and similar DBus interfaces
* progresses in taskbar buttons using Unity DBus interface

Closes: #552, #442
