### 2019-06-11 7bd9482
Added GUI progress reporting for Windows (using Taskbar Extensions) and Linux ( KDE and Unity using DBus)
Closes: #552, #442
