### 2019-06-18 9011f33
100% _tqdm.py coverage
