+++ tqdm/tests/tests_tqdm.py
@@ -1688,10 +1688,11 @@ def test_external_write():
 def test_unit_scale():
     """Test numeric `unit_scale`"""
     with closing(StringIO()) as our_file:
-        for _ in tqdm(_range(100), unit_scale=9, file=our_file):
+        for _ in tqdm(_range(9), unit_scale=9, file=our_file,
+                      miniters=1,  mininterval=0):
             pass
         out = our_file.getvalue()
-        assert '900/900' in out
+        assert '81/81' in out
 
 
 @with_setup(pretest, posttest)
