### 2019-06-18 36b3b7d
update notable contributors list
