### 2019-06-18 89d1494
flake8 and minor dev docs
