+++ tqdm/_tqdm.py
@@ -985,6 +985,17 @@ class tqdm(Comparable):
     def __hash__(self):
         return id(self)
 
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        try:
+            self.update()
+            return await self.iterable.__anext__()
+        except StopIteration:
+            self.close()
+            raise StopAsyncIteration
+
     def __iter__(self):
         """Backward-compatibility to use: for x in tqdm(iterable)"""
 
