### 2019-07-11 5b42cd1
more typos
- closes #773
