### 2019-07-14 7d79047
remove the = between options and values in tqdm --help
