### 2019-07-19 397d88d
fix broken url link to awesome-python
