### 2019-07-19 e7eae14
fix broken url link to awesome-python
