### 2019-07-23 f2dd9c9
Fix typo in awesome-python link
