### 2019-08-02 4c2186c
Added Additional Base Functionality
Added functionality:
All prints now show the filename, function, line number, and description
