### 2019-08-06 ef8409f
Added current time
TQDM now outputs the print time as well
