### 2019-08-06 07b5a6e
Printing also conforms to bar format
Bar format has been ported over to printing
