### 2019-08-08 7277b70
fix pandas==0.25 API change
- fixes #780
