### 2019-08-08 aecd4e0
potential fix for py26 travis CI
