+++ .travis.yml
@@ -3,6 +3,7 @@ matrix:
   include:
   - python: 2.6
     env: TOXENV=py26
+    dist: trusty
   - python: 2.7
     env: TOXENV=py27
   - python: 3.4
