### 2019-08-08 bb3936e
actually fix
- fixes #778
- fixes #782
