### 2019-08-08 d2c4d75
actually support both ' ' and '=' for CLI
- closes #774
- fixes #761
