+++ tqdm/_main.py
@@ -89,7 +89,7 @@ def posix_pipe(fin, fout, delim='\n', buf_size=256,
 # ((opt, type), ... )
 RE_OPTS = re.compile(r'\n {8}(\S+)\s{2,}:\s*([^,]+)')
 # better split method assuming no positional args
-RE_SHLEX = re.compile(r'\s*(?<!\S)--?([^\s=]+)(?:\s*|=|$)')
+RE_SHLEX = re.compile(r'\s*(?<!\S)--?([^\s=]+)(\s+|=|$)')
 
 # TODO: add custom support for some of the following?
 UNSUPPORTED_OPTS = ('iterable', 'gui', 'out', 'file')
@@ -128,7 +128,12 @@ def main(fp=sys.stderr, argv=None):
     try:
         log = argv.index('--log')
     except ValueError:
-        logLevel = 'INFO'
+        for i in argv:
+            if i.startswith('--log='):
+                logLevel = i[len('--log='):]
+                break
+        else:
+            logLevel = 'INFO'
     else:
         # argv.pop(log)
         # logLevel = argv.pop(log)
@@ -151,7 +156,7 @@ def main(fp=sys.stderr, argv=None):
     # d = RE_OPTS.sub(r'  --\1=<\1>  : \2', d)
     split = RE_OPTS.split(d)
     opt_types_desc = zip(split[1::3], split[2::3], split[3::3])
-    d = ''.join('\n  --{0} <{0}>  : {1}{2}'.format(*otd)
+    d = ''.join('\n  --{0}=<{0}>  : {1}{2}'.format(*otd)
                 for otd in opt_types_desc if otd[0] not in UNSUPPORTED_OPTS)
 
     d = """Usage:
@@ -172,7 +177,7 @@ Options:
         sys.exit(0)
 
     argv = RE_SHLEX.split(' '.join(["tqdm"] + argv))
-    opts = dict(zip(argv[1::2], argv[2::2]))
+    opts = dict(zip(argv[1::3], argv[3::3]))
 
     log.debug(opts)
     opts.pop('log', True)
