### 2019-08-10 129311e
Fix tqdm_notebook red block on explicit file=None
