### 2019-08-17 d0c85fc
minor doc update
