### 2019-08-17 aafa889
clear bars before shifting up
TODO: fix unit tests

- fixes #795
