### 2019-08-18 08261f1
add leave=None, make close() display at pos=0 if leave
- fixes #230
  + https://github.com/tqdm/tqdm/issues/230#issuecomment-277364386
  + note that other bars are not refreshed
