### 2019-08-18 45a475c
minor closing when leave=True stats fix
