### 2019-08-18 7c3d2e1
update/fix unit tests
