### 2019-08-18 81998b5
more test fixes
