+++ tqdm/tests/tests_tqdm.py
@@ -1179,21 +1179,28 @@ def test_position():
              '\n\n\rpos2 bar:   0%',
              '\n\n\rpos2 bar:  50%',
              '\n\n\rpos2 bar: 100%',
-             '\n\rpos1 bar:  50%',
+             '\rpos2 bar: 100%',
+             '\n\n\rpos1 bar:  50%',
              '\n\n\rpos2 bar:   0%',
              '\n\n\rpos2 bar:  50%',
              '\n\n\rpos2 bar: 100%',
-             '\n\rpos1 bar: 100%',
-             '\rpos0 bar:  50%',
+             '\rpos2 bar: 100%',
+             '\n\n\rpos1 bar: 100%',
+             '\rpos1 bar: 100%',
+             '\n\rpos0 bar:  50%',
              '\n\rpos1 bar:   0%',
              '\n\n\rpos2 bar:   0%',
              '\n\n\rpos2 bar:  50%',
              '\n\n\rpos2 bar: 100%',
-             '\n\rpos1 bar:  50%',
+             '\rpos2 bar: 100%',
+             '\n\n\rpos1 bar:  50%',
              '\n\n\rpos2 bar:   0%',
              '\n\n\rpos2 bar:  50%',
              '\n\n\rpos2 bar: 100%',
-             '\n\rpos1 bar: 100%',
+             '\rpos2 bar: 100%',
+             '\n\n\rpos1 bar: 100%',
+             '\rpos1 bar: 100%',
+             '\n\rpos0 bar: 100%',
              '\rpos0 bar: 100%',
              '\n']
     pos_line_diff(res, exres)
