### 2019-08-18 8727386
tidy pos_line_diff test fail output
