### 2019-08-18 d854b5b
final test fixes
