### 2019-08-18 b1f2124
linting: flake8 fixes& upgrade
