### 2019-08-18 4cf9cbd
temp fix GitHub docker package deployment
