### 2019-08-18 5b1c130
update package docs about `leave=None`
