+++ tqdm/tests/tests_perf.py
@@ -298,7 +298,7 @@ def test_iter_overhead_simplebar_hard():
 
     # Compute relative overhead of tqdm against native range()
     try:
-        assert time_tqdm() < 3 * time_bench()
+        assert time_tqdm() < 3.5 * time_bench()
     except AssertionError:
         raise AssertionError('trange(%g): %f, simple_progress(%g): %f' %
                              (total, time_tqdm(), total, time_bench()))
@@ -321,7 +321,8 @@ def test_manual_overhead_simplebar_hard():
                 t.update(10)
 
         simplebar_update = simple_progress(
-            total=total, file=our_file, leave=True, miniters=1, mininterval=0)
+            total=total * 10, file=our_file, leave=True, miniters=1,
+            mininterval=0)
         a = 0
         with relative_timer() as time_bench:
             for i in _range(total):
@@ -330,7 +331,7 @@ def test_manual_overhead_simplebar_hard():
 
     # Compute relative overhead of tqdm against native range()
     try:
-        assert time_tqdm() < 3 * time_bench()
+        assert time_tqdm() < 3.5 * time_bench()
     except AssertionError:
         raise AssertionError('tqdm(%g): %f, simple_progress(%g): %f' %
                              (total, time_tqdm(), total, time_bench()))
