### 2019-08-18 1481617
update multiprocessing and multithreading examples
- part of #796
