### 2019-08-18 e877ea1
update multiprocessing and multithreading examples
- part of #796
