### 2019-08-23 c410a5a
initial notebook upgrade
- fixes #790
- fixes #740
