+++ tqdm/_tqdm_notebook.py
@@ -135,55 +135,63 @@ class tqdm_notebook(tqdm):
             container.layout.flex_flow = 'row wrap'
         display(container)
 
-        def print_status(s='', close=False, bar_style=None, desc=None):
-            # Note: contrary to native tqdm, s='' does NOT clear bar
-            # goal is to keep all infos if error happens so user knows
-            # at which iteration the loop failed.
+        return container
 
-            # Clear previous output (really necessary?)
-            # clear_output(wait=1)
+    def display(self, msg=None, pos=None,
+                # additional signals
+                close=False, bar_style=None):
+        # Note: contrary to native tqdm, msg='' does NOT clear bar
+        # goal is to keep all infos if error happens so user knows
+        # at which iteration the loop failed.
 
-            # Get current iteration value from format_meter string
-            if total:
-                # n = None
-                if s:
-                    npos = s.find(r'/|/')  # cause we use bar_format=r'{n}|...'
-                    # Check that n can be found in s (else n > total)
-                    if npos >= 0:
-                        n = float(s[:npos])  # get n from string
-                        s = s[npos + 3:]  # remove from string
+        # Clear previous output (really necessary?)
+        # clear_output(wait=1)
 
-                        # Update bar with current n value
-                        if n is not None:
-                            pbar.value = n
+        # Update description
+        if self.desc:
+            pbar.description = self.desc
+            self.desc = None  # trick to place description before the bar
+            if IPYW >= 7:
+                pbar.style.description_width = 'initial'
 
-            # Print stats
-            if s:  # never clear the bar (signal: s='')
-                s = s.replace('||', '')  # remove inesthetical pipes
-                s = escape(s)  # html escape special characters (like '?')
-                ptext.value = s
+        if not msg and not close:
+            msg = self.__repr__()
 
-            # Change bar style
-            if bar_style:
-                # Hack-ish way to avoid the danger bar_style being overridden by
-                # success because the bar gets closed after the error...
-                if not (pbar.bar_style == 'danger' and bar_style == 'success'):
-                    pbar.bar_style = bar_style
+        pbar, ptext = self.container.children
 
-            # Special signal to close the bar
-            if close and pbar.bar_style != 'danger':  # hide only if no error
-                try:
-                    container.close()
-                except AttributeError:
-                    container.visible = False
+        # Get current iteration value from format_meter string
+        if self.total:
+            # n = None
+            if msg:
+                npos = msg.find(r'/|/')  # cause we use bar_format=r'{n}|...'
+                # Check that n can be found in msg (else n > total)
+                if npos >= 0:
+                    n = float(msg[:npos])  # get n from string
+                    msg = msg[npos + 3:]  # remove from string
 
-            # Update description
-            if desc:
-                pbar.description = desc
-                if IPYW >= 7:
-                    pbar.style.description_width = 'initial'
+                    # Update bar with current n value
+                    if n is not None:
+                        pbar.value = n
 
-        return print_status
+        # Print stats
+        if msg:  # never clear the bar (signal: msg='')
+            msg = msg.replace('||', '')  # remove inesthetical pipes
+            msg = escape(msg)  # html escape special characters (like '?')
+            ptext.value = msg
+
+        # Change bar style
+        if bar_style:
+            # Hack-ish way to avoid the danger bar_style being overridden by
+            # success because the bar gets closed after the error...
+            if not (pbar.bar_style == 'danger' and bar_style == 'success'):
+                pbar.bar_style = bar_style
+
+        # Special signal to close the bar
[…]
