### 2019-08-23 bda07d1
update notebook to fully support custom `bar_format`
- fixes #594
- fixes #649
