### 2019-08-24 ecd2ed6
add Bar.__format__ helper
- supporting `{bar:[width][type]}` format string
- fixes #623
