### 2019-08-24 5624c31
add Bar.__format__ tests
