### 2019-08-24 1aa3745
move and document empty format helper
