### 2019-08-24 273e0c7
fix py26 tests
