+++ tqdm/tests/tests_tqdm.py
@@ -332,9 +332,9 @@ def test_si_format():
 
 def test_bar_formatspec():
     """Test Bar.__format__ spec"""
-    assert "{:5a}".format(Bar(0.3)) == "#5   "
-    assert "{:2}".format(Bar(0.5, charset=" .oO0")) == "0 "
-    assert "{:2a}".format(Bar(0.5, charset=" .oO0")) == "# "
+    assert "{0:5a}".format(Bar(0.3)) == "#5   "
+    assert "{0:2}".format(Bar(0.5, charset=" .oO0")) == "0 "
+    assert "{0:2a}".format(Bar(0.5, charset=" .oO0")) == "# "
 
 
 @with_setup(pretest, posttest)
