### 2019-08-24 3c17710
negative bar len for auto-trimming
