### 2019-08-24 532d433
blank bar format spec type
- fixes #623
- e.g. for fixed bar width of 10 and right-justified text:
  bar_format='{l_bar}{bar:10}|{bar:-10b}some text'
