### 2019-08-24 8888db4
tidy perf tests
