### 2019-08-24 86db96c
update perf test thresholds
