### 2019-08-24 ee68bac
add bar formatspec documentation
