### 2019-08-24 6d1d38f
linting (flake8, codacy)
