### 2019-08-24 359864c
parallel `(py)make test`
