### 2019-08-26 5d373de
support custom bar_format with unknown total
