### 2019-08-26 1340672
minor gui update
