### 2019-08-26 b8a28f2
update tqdm_gui logic from tqdm
