### 2019-08-26 13911cf
display-ify tqdm_gui
