+++ tqdm/_tqdm_gui.py
@@ -106,10 +106,6 @@ class tqdm_gui(tqdm):  # pragma: no cover
         maxinterval = self.maxinterval
         miniters = self.miniters
         dynamic_miniters = self.dynamic_miniters
-        unit = self.unit
-        unit_scale = self.unit_scale
-        ascii = self.ascii
-        start_t = self.start_t
         last_print_t = self.last_print_t
         last_print_n = self.last_print_n
         n = self.n
@@ -117,15 +113,6 @@ class tqdm_gui(tqdm):  # pragma: no cover
         smoothing = self.smoothing
         avg_time = self.avg_time
         time = self._time
-        bar_format = self.bar_format
-
-        plt = self.plt
-        ax = self.ax
-        xdata = self.xdata
-        ydata = self.ydata
-        zdata = self.zdata
-        line1 = self.line1
-        line2 = self.line2
 
         for obj in iterable:
             yield obj
@@ -139,7 +126,6 @@ class tqdm_gui(tqdm):  # pragma: no cover
                 if delta_t >= mininterval:
                     cur_t = time()
                     delta_it = n - last_print_n
-                    elapsed = cur_t - start_t
                     # EMA (not just overall average)
                     if smoothing and delta_t and delta_it:
                         rate = delta_t / delta_it
@@ -147,58 +133,7 @@ class tqdm_gui(tqdm):  # pragma: no cover
                         self.avg_time = avg_time
 
                     self.n = n
-                    # Inline due to multiple calls
-                    total = self.total
-                    # instantaneous rate
-                    y = delta_it / delta_t
-                    # overall rate
-                    z = n / elapsed
-                    # update line data
-                    xdata.append(n * 100.0 / total if total else cur_t)
-                    ydata.append(y)
-                    zdata.append(z)
-
-                    # Discard old values
-                    # xmin, xmax = ax.get_xlim()
-                    # if (not total) and elapsed > xmin * 1.1:
-                    if (not total) and elapsed > 66:
-                        xdata.popleft()
-                        ydata.popleft()
-                        zdata.popleft()
-
-                    ymin, ymax = ax.get_ylim()
-                    if y > ymax or z > ymax:
-                        ymax = 1.1 * y
-                        ax.set_ylim(ymin, ymax)
-                        ax.figure.canvas.draw()
-
-                    if total:
-                        line1.set_data(xdata, ydata)
-                        line2.set_data(xdata, zdata)
-                        try:
-                            poly_lims = self.hspan.get_xy()
-                        except AttributeError:
-                            self.hspan = plt.axhspan(0, 0.001, xmin=0,
-                                                     xmax=0, color='g')
-                            poly_lims = self.hspan.get_xy()
-                        poly_lims[0, 1] = ymin
-                        poly_lims[1, 1] = ymax
-                        poly_lims[2] = [n / total, ymax]
-                        poly_lims[3] = [poly_lims[2, 0], ymin]
-                        if len(poly_lims) > 4:
-                            poly_lims[4, 1] = ymin
-                        self.hspan.set_xy(poly_lims)
-                    else:
-                        t_ago = [cur_t - i for i in xdata]
-                        line1.set_data(t_ago, ydata)
-                        line2.set_data(t_ago, zdata)
-
-                    ax.set_title(self.format_meter(
-                        n, total, elapsed, 0,
-                        self.desc, ascii, unit, unit_scale,
-                        1 / avg_time if avg_time else None, bar_format),
-                        fontname="DejaVu Sans Mono", fontsize=11)
-                    plt.pause(1e-9)
[…]
