### 2019-08-26 f7cf8c9
fix minor cached var optimisation
