### 2019-08-26 2670a17
migrate to new submodule architecture
- fixes #198
