+++ tqdm/auto.py
@@ -1,6 +1,6 @@
 import warnings
-from .._tqdm import TqdmExperimentalWarning
+from .std import TqdmExperimentalWarning
 with warnings.catch_warnings():
     warnings.simplefilter("ignore", category=TqdmExperimentalWarning)
-    from ..autonotebook import tqdm, trange
+    from .autonotebook import tqdm, trange
 __all__ = ["tqdm", "trange"]
