+++ MANIFEST.in
@@ -11,10 +11,6 @@ include tox.ini
 recursive-include tqdm/tests *.py
 include requirements-dev.txt
 
-# Sub-packages
-recursive-include tqdm/autonotebook *.py
-recursive-include tqdm/auto *.py
-
 # Examples/Documentation
 recursive-include examples *.py
 include README.rst
