+++ tqdm/std.py
@@ -11,7 +11,7 @@ from __future__ import absolute_import
 # integer division / : float, // : int
 from __future__ import division
 # compatibility functions and utilities
-from ._utils import _supports_unicode, _environ_cols_wrapper, _range, _unich, \
+from .utils import _supports_unicode, _environ_cols_wrapper, _range, _unich, \
     _term_move_up, _unicode, WeakSet, _basestring, _OrderedDict, \
     Comparable, RE_ANSI, _is_ascii, SimpleTextIOWrapper, FormatReplace
 from ._monitor import TMonitor
