+++ tqdm/tests/tests_tqdm.py
@@ -1318,8 +1318,8 @@ def test_deprecated_gui():
         try:
             t.update(1)
         except TqdmDeprecationWarning as e:
-            if 'Please use `tqdm_gui(...)` instead of `tqdm(..., gui=True)`' \
-                    not in our_file.getvalue():
+            if ('Please use `tqdm.gui.tqdm(...)` instead of'
+                ' `tqdm(..., gui=True)`') not in our_file.getvalue():
                 raise e
         else:
             raise DeprecationError('Should not allow manual gui=True without'
@@ -1334,8 +1334,8 @@ def test_deprecated_gui():
             for _ in t:
                 pass
         except TqdmDeprecationWarning as e:
-            if 'Please use `tqdm_gui(...)` instead of `tqdm(..., gui=True)`' \
-                    not in our_file.getvalue():
+            if ('Please use `tqdm.gui.tqdm(...)` instead of'
+                ' `tqdm(..., gui=True)`') not in our_file.getvalue():
                 raise e
         else:
             raise DeprecationError('Should not allow manual gui=True without'
