### 2019-08-26 2d1a7f5
update pandas, notebook, gui docs
- fixes #245
- part of #800
