### 2019-08-26 c4220fd
update submodule contribution guide
- fixes #252
