### 2019-08-26 be1275b
Hardening the last commit
This is what the last commit should have been.
Robustified computing the numebr columns across *NIX and PC (by using a default for PC)
Removed commented out debug code
Added print functionality
Some basic refactoring for future maintenance
