### 2019-08-27 a347706
potential thread safety fix
- related: #548
