### 2019-09-07 5be8ff9
Add tests for wide characters in description
Making tests work with Python 2 was annoying, so I just skip them there.
It'll be EOLd in a few months anyway.
