### 2019-09-17 1bd486c
expose more custom format params
