### 2019-09-17 46135ce
minor doc updates
- sync inline code documentation
- add PyData London links
