### 2019-10-02 1b1ec3e
remove broken funding link
