+++ .github/FUNDING.yml
@@ -1,2 +1,2 @@
-github: casperdcl
+# github: casperdcl
 custom: https://caspersci.uk.to/donate
