### 2019-10-03 fa2e30f
MNT: set stack level on warnings
This makes it much easier to find where in client code needs to be
fixed.
