+++ tqdm/_tqdm_gui.py
@@ -4,4 +4,4 @@ from .std import TqdmDeprecationWarning
 from warnings import warn
 warn("This function will be removed in tqdm==5.0.0\n"
      "Please use `tqdm.gui.*` instead of `tqdm._tqdm_gui.*`",
-     TqdmDeprecationWarning)
+     TqdmDeprecationWarning, stacklevel=2)
