### 2019-10-10 e601dd9
attempt to fix perf
- fixes #803
- closes #410
- closes #805
