+++ tqdm/utils.py
@@ -285,15 +285,11 @@ def _term_move_up():  # pragma: no cover
     return '' if (os.name == 'nt') and (colorama is None) else '\x1b[A'
 
 
-def _text_width(s):  # pragma: no cover
+try:
     # TODO consider using wcswidth third-party package for 0-width characters
-    try:
-        from unicodedata import east_asian_width
-    except ImportError:
-        return len(s)
-    else:
-        try:
-            return len(s) + sum(east_asian_width(ch) in 'FW' for ch in s)
-        except TypeError:  # Py2
-            s = s.decode('utf-8')
-            return len(s) + sum(east_asian_width(ch) in 'FW' for ch in s)
+    from unicodedata import east_asian_width
+except ImportError:
+    _text_width = len
+else:
+    def _text_width(s):
+        return len(s) + sum(east_asian_width(ch) in 'FW' for ch in _unicode(s))
