### 2019-10-11 b3d7739
unicode width speed improvement, clear fix
