+++ tqdm/std.py
@@ -300,7 +300,7 @@ class tqdm(Comparable):
         last_len = [0]
 
         def print_status(s):
-            len_s = _text_width(s)
+            len_s = len(s)
             fp_write('\r' + s + (' ' * max(last_len[0] - len_s, 0)))
             last_len[0] = len_s
 
