### 2019-10-12 6e50fa2
add stacklevel=2 to more warnings
