+++ tqdm/_tqdm.py
@@ -4,4 +4,4 @@ from .std import TqdmDeprecationWarning
 from warnings import warn
 warn("This function will be removed in tqdm==5.0.0\n"
      "Please use `tqdm.std.*` instead of `tqdm._tqdm.*`",
-     TqdmDeprecationWarning)
+     TqdmDeprecationWarning, stacklevel=2)
