### 2019-10-22 9baa46e
fix example/tqdm_wget unknown total
- fixes #826
