### 2019-10-25 5f0eb91
Revert "remove broken funding link"
This reverts commit 1b1ec3e3abc798b6d4a01426657acc89d01137f6.
