### 2019-10-29 84102ab
update issue/pr templates
would fix things like #830
