+++ .github/ISSUE_TEMPLATE.md
@@ -1,3 +1,8 @@
+- [ ] I have marked all applicable categories:
+    + [ ] exception-raising bug
+    + [ ] visual output bug
+    + [ ] documentation request (i.e. "X is missing from the documentation." If instead I want to ask "how to use X?" I understand [StackOverflow#tqdm] is more appropriate)
+    + [ ] new feature request
 - [ ] I have visited the [source website], and in particular
   read the [known issues]
 - [ ] I have searched through the [issue tracker] for duplicates
@@ -11,3 +16,4 @@
   [source website]: https://github.com/tqdm/tqdm/
   [known issues]: https://github.com/tqdm/tqdm/#faq-and-known-issues
   [issue tracker]: https://github.com/tqdm/tqdm/issues?q=
+  [StackOverflow#tqdm]: https://stackoverflow.com/questions/tagged/tqdm
