### 2019-10-29 59f457a
potential future pandas fix
- related to #824
