### 2019-10-31 b17b594
update parallel example
- fixes #407
