### 2019-10-31 c9413df
attempt to move to Travis stages
