### 2019-10-31 a46caa4
linting/fix flake8
