+++ examples/parallel_bars.py
@@ -15,7 +15,8 @@ def progresser(n, auto_position=True, write_safe=False):
     interval = random() * 0.002 / (NUM_SUBITERS - n + 2)
     total = 5000
     text = "#{}, est. {:<04.2}s".format(n, interval * total)
-    for _ in tqdm(range(total), desc=text, position=None if auto_position else n):
+    for _ in trange(total, desc=text,
+                    position=None if auto_position else n):
         sleep(interval)
     # NB: may not clear instances with higher `position` upon completion
     # since this worker may not know about other bars #796
