+++ setup.cfg
@@ -2,6 +2,6 @@
 universal = 1
 
 [flake8]
-ignore = W503,W504
+ignore = W503,W504,E722
 max_line_length = 80
 exclude = .asv,.tox,.ipynb_checkpoints,build,dist,.git,__pycache__
