+++ tox.ini
@@ -77,7 +77,7 @@ commands =
 [testenv:flake8]
 deps = flake8
 commands =
-    flake8 -j 8 --count --statistics --exit-zero .
+    flake8 -j 8 --count --statistics .
 
 [testenv:setup.py]
 deps =
