### 2019-11-01 4c7cb44
initial keras.callback
