+++ tqdm/keras.py
@@ -14,10 +14,17 @@ class TqdmCallback(Callback):
                     [logs.pop(i) for i in pop]
                 bar.set_postfix(logs, refresh=False)
             bar.update()
+
         return callback
 
-    def __init__(self, epochs, data_size=None, batch_size=None, verbose=1,
-                 tqdm_class=tqdm_auto):
+    def __init__(
+        self,
+        epochs,
+        data_size=None,
+        batch_size=None,
+        verbose=1,
+        tqdm_class=tqdm_auto,
+    ):
         """
         Parameters
         ----------
@@ -34,31 +41,33 @@ class TqdmCallback(Callback):
             `tqdm` class to use for bars [default: `tqdm.auto.tqdm`].
         """
         self.tqdm_class = tqdm_class
-        self.te = tqdm_class(total=epochs, unit='epoch')
-        self.on_epoch_end = self.bar2callback(self.te)
+        self.epoch_bar = tqdm_class(total=epochs, unit="epoch")
+        self.on_epoch_end = self.bar2callback(self.epoch_bar)
         if data_size and batch_size:
-            self.batches = batches = (
-                data_size + batch_size - 1) // batch_size
+            self.batches = batches = (data_size + batch_size - 1) // batch_size
         else:
             if verbose:
                 print("W:missing batch and data size")
             verbose = 0
         self.verbose = verbose
         if verbose == 1:
-            self.tb = tqdm_class(total=batches, unit='batch', leave=False)
-            self.on_batch_end = self.bar2callback(self.tb, ['batch'])
+            self.batch_bar = tqdm_class(
+                total=batches, unit="batch", leave=False
+            )
+            self.on_batch_end = self.bar2callback(self.batch_bar, pop=["batch"])
 
     def on_epoch_begin(self, *_, **__):
         if self.verbose == 2:
-            if hasattr(self, 'tb'):
-                self.tb.close()
-            self.tb = self.tqdm_class(
-                total=self.batches, unit='batch', leave=True)
-            self.on_batch_end = self.bar2callback(self.tb, ['batch'])
+            if hasattr(self, "batch_bar"):
+                self.batch_bar.close()
+            self.batch_bar = self.tqdm_class(
+                total=self.batches, unit="batch", leave=True
+            )
+            self.on_batch_end = self.bar2callback(self.batch_bar, pop=["batch"])
         elif self.verbose == 1:
-            self.tb.reset()
+            self.batch_bar.reset()
 
     def on_train_end(self, *_, **__):
         if self.verbose:
-            self.tb.close()
-        self.te.close()
+            self.batch_bar.close()
+        self.epoch_bar.close()
