### 2019-11-07 90774d3
Allow hiding progressbar until some time interval has elapsed.
This avoids printing the progressbar for very short loops.
