+++ tests/tests_tqdm.py
@@ -622,6 +622,21 @@ def test_max_interval():
         t2.close()
 
 
+def test_waituntil():
+    """Test waituntil"""
+    timer = DiscreteTimer()
+    with closing(StringIO()) as our_file:
+        t = tqdm(total=2, file=our_file, leave=True, waituntil=3)
+        cpu_timify(t, timer)
+        timer.sleep(2)
+        t.update(1)
+        assert not our_file.getvalue()
+        timer.sleep(2)
+        t.update(1)
+        assert our_file.getvalue()
+        t.close()
+
+
 def test_min_iters():
     """Test miniters"""
     with closing(StringIO()) as our_file:
