### 2019-11-07 83f2752
Replace dedent calls by dedented strings.
short literals are faster and more readable than using dedent()

Co-authored by: Casper da Costa-Luis <tqdm@caspersci.uk.to>
