### 2019-11-07 b989bc4
support args for `TqdmDefaultWriteLock.acquire()`
- related to #838
