### 2019-11-07 23e7fb1
use `self.refresh()` in more places
- working towards #838
