### 2019-11-07 dba8a28
default `lock_args=(False,)` for intermediate progress
- fixes #838
