### 2019-11-07 66037ef
default lock_args=None and expose
- avoids errors with custom locks not supporting acquire(False)
