### 2019-11-08 37038f0
fix acquisition and perf debugging
