### 2019-11-08 52b985e
fix lock_args on creation, add tests
