### 2019-11-08 7c38d0b
fix flake8 and add test logic
