### 2019-11-08 d301a12
debugging thread races
