### 2019-11-08 213f302
fix silly lock bug, tighten perf tests
