### 2019-11-08 464165c
fix py3.4 tests
