+++ examples/parallel_bars.py
@@ -4,6 +4,7 @@ from tqdm import tqdm, trange
 from random import random
 from multiprocessing import Pool, freeze_support
 from concurrent.futures import ThreadPoolExecutor
+from threading import RLock
 from functools import partial
 import sys
 
@@ -11,11 +12,12 @@ NUM_SUBITERS = 9
 PY2 = sys.version_info[:1] <= (2,)
 
 
-def progresser(n, auto_position=True, write_safe=False):
+def progresser(n, auto_position=True, write_safe=False, blocking=True):
     interval = random() * 0.002 / (NUM_SUBITERS - n + 2)
     total = 5000
     text = "#{}, est. {:<04.2}s".format(n, interval * total)
     for _ in trange(total, desc=text,
+                    lock_args=None if blocking else (False,),
                     position=None if auto_position else n):
         sleep(interval)
     # NB: may not clear instances with higher `position` upon completion
@@ -45,6 +47,9 @@ if __name__ == '__main__':
         ncols = t.ncols or 80
     print(("{msg:<{ncols}}").format(msg="Multi-threading", ncols=ncols))
 
+    # explicitly set just threading lock for nonblocking progress
+    tqdm.set_lock(RLock())
     with ThreadPoolExecutor() as p:
-        progresser_thread = partial(progresser, write_safe=not PY2)
+        progresser_thread = partial(
+            progresser, write_safe=not PY2, blocking=False)
         p.map(progresser_thread, L)
