+++ .travis.yml
@@ -20,7 +20,8 @@ notifications:
 stages:
 - check
 - test
-- deploy
+- name: deploy
+  if: repo = tqdm/tqdm
 jobs:
   include:
   - name: py2.6
@@ -74,6 +75,7 @@ jobs:
       -iv $encrypted_a6d6301302b7_iv -in .meta/.tqdm.gpg.enc -out .tqdm.gpg -d
     - gpg --import .tqdm.gpg
     - rm .tqdm.gpg
+    - git log --pretty='format:- %s%n%b---' $(git tag --sort=creatordate | tail -n2 | head -n1)..HEAD > CHANGES.md
     deploy:
     - provider: script
       script: twine upload -s -i tqdm@caspersci.uk.to dist/tqdm-*
@@ -87,6 +89,8 @@ jobs:
       skip_cleanup: true
       draft: true
       name: tqdm $TRAVIS_TAG stable
+      edge: true
+      release_notes_file: CHANGES.md
       on:
         tags: true
   - name: docker
