### 2019-11-09 76e9c04
add basic keras test
