### 2019-11-10 a60552b
clean CI stages
