+++ .travis.yml
@@ -32,6 +32,11 @@ jobs:
   - name: py2.7
     python: 2.7
     env: TOXENV=py27
+  - name: py2.7-win
+    os: windows
+    language: shell
+    python: 2.7
+    env: TOXENV=py27
   - name: py3.4
     python: 3.4
     env: TOXENV=py34
@@ -42,6 +47,11 @@ jobs:
     python: 3.6
     env: TOXENV=py36
   - name: py3.7
+    python: 3.7
+    env: TOXENV=py37
+  - name: py3.7-win
+    os: windows
+    language: shell
     python: 3.7
     env: TOXENV=py37
   - name: pypy2.7
