### 2019-11-10 7812648
fix py2.6, allow win fail
