### 2019-11-13 80ade93
Add check for pandas builtin for progress apply
Applied func will be checked against pandas internal builtins in order
to apply proper transformation, i.e max->amax

Closes:#697
