### 2019-11-15 49c5e15
remove unneeded bar logic
