### 2019-11-19 505669b
add tqdm.tqdm.wrapattr
- fixes #84
