### 2019-11-19 9a949ec
move to classmethod wrapattr, rename low-level to self_wrapattr
