### 2019-11-20 76abd5d
document default argument changing
- fixes #370
