### 2019-11-20 c032868
add `CallbackIOWrapper`, add tests
