### 2019-11-20 2fd7eb2
nuke self_wrapattr
