### 2019-11-21 06930c9
add and test wrapattr(bytes=True)
