### 2019-11-21 4c63100
fix and tidy snap
