+++ tqdm/taskbar/meta_reporter.py
@@ -1,8 +1,8 @@
-__all__=("MetaReporter",)
 import sys
 import importlib
 from functools import wraps
 from .PlatformSpecificProgressReporter import PlatformSpecificProgressReporter
+__all__ = ["MetaReporter"]
 
 systemToReporterMapping={
     "win32": ("windows_taskbar_extensions", ),
@@ -25,24 +25,24 @@ class MetaReporter(PlatformSpecificProgressReporter):
         self.reporters=[]
         for cls in reporterClassesToUse:
             self.reportersCandidates.append(cls(*args, **kwargs))
-    
+
     @wraps(PlatformSpecificProgressReporter.progress)
     def progress(self, *args, **kwargs):
         for rep in self.reporters:
             rep.progress(*args, **kwargs)
-    
+
     def fail(self, reason:str=None):
         for rep in self.reporters:
             rep.fail(reason)
-    
+
     def success(self):
         for rep in self.reporters:
             rep.success()
-    
+
     def prefix(self, prefix:str):
         for rep in self.reporters:
             rep.prefix(prefix)
-    
+
     def message(self, message:str):
         for rep in self.reporters:
             rep.message(message)
@@ -50,7 +50,7 @@ class MetaReporter(PlatformSpecificProgressReporter):
     def postfix(self, postfix:str):
         for rep in self.reporters:
             rep.postfix(postfix)
-    
+
     def __enter__(self):
         for rep in self.reportersCandidates:
             #try:
@@ -58,11 +58,11 @@ class MetaReporter(PlatformSpecificProgressReporter):
             #except:
             #    pass
         return self
-    
+
     def __exit__(self, exception_type, exception_value, traceback):
         for rep in self.reporters:
             rep.__exit__(exception_type, exception_value, traceback)
-    
+
     def clear(self):
         for rep in self.reporters:
             rep.clear()
