+++ tqdm/taskbar/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import division, absolute_import
+from .taskbar import tqdm, trange
+from .meta_reporter import MetaReporter
+
+__author__ = {"github.com/": ["KOLANICH", "casperdcl"]}
+__all__ = ["MetaReporter", "tqdm", "trange"]
