+++ tqdm/std.py
@@ -19,7 +19,6 @@ from warnings import warn
 from weakref import WeakSet
 
 from ._monitor import TMonitor
-from .platform_specific import MetaReporter
 from .utils import (
     CallbackIOWrapper, Comparable, DisableOnWriteError, FormatReplace,
     SimpleTextIOWrapper, _basestring, _is_ascii, _range, _screen_shape_wrapper,
@@ -957,7 +956,6 @@ class tqdm(Comparable):
         -------
         out  : decorated iterator.
         """
-        self._entered=0
         if write_bytes is None:
             write_bytes = file is None and sys.version_info < (3,)
 
@@ -1076,7 +1074,6 @@ class tqdm(Comparable):
         self.postfix = None
         self.colour = colour
         self._time = time
-        self.platform_specific_reporter = None
         if postfix:
             try:
                 self.set_postfix(refresh=False, **postfix)
@@ -1122,33 +1119,18 @@ class tqdm(Comparable):
              else getattr(self, "total", None))
 
     def __enter__(self):
-        if not self._entered:
-            self.platform_specific_reporter=MetaReporter(self.total, self.unit).__enter__()
-            self._entered+=1
         return self
 
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        if self._entered:
-            if self.platform_specific_reporter:
-                if exc_traceback:
-                    import traceback
-                    self.platform_specific_reporter.fail("".join(traceback.format_exception(exc_type, exc_value, exc_traceback, limit=1)))
-                elif self.total and self.n < self.total:
-                    self.platform_specific_reporter.fail("The iteration not reached its finishing amount, though no exception has been thrown.")
-                self.platform_specific_reporter.__exit__(exc_type, exc_value, exc_traceback)
-                self.platform_specific_reporter = None
-            try:
-                self.close()
-            except AttributeError:
-                # maybe eager thread cleanup upon external error
-                if (exc_type, exc_value, traceback) == (None, None, None):
-                    raise
-                warn("AttributeError ignored", TqdmWarning, stacklevel=2)
-            self._entered-=1
-            return False
+    def __exit__(self, exc_type, exc_value, traceback):
+        try:
+            self.close()
+        except AttributeError:
+            # maybe eager thread cleanup upon external error
+            if (exc_type, exc_value, traceback) == (None, None, None):
+                raise
+            warn("AttributeError ignored", TqdmWarning, stacklevel=2)
 
     def __del__(self):
-        self.__exit__(None, None, None)
         self.close()
 
     def __str__(self):
@@ -1163,8 +1145,6 @@ class tqdm(Comparable):
 
     def __iter__(self):
         """Backward-compatibility to use: for x in tqdm(iterable)"""
-        if not self._entered and self.iterable:
-            self.__enter__()
 
         # Inlining instance variables as locals (speed optimisation)
         iterable = self.iterable
@@ -1197,10 +1177,6 @@ class tqdm(Comparable):
                         last_print_t = self.last_print_t
         finally:
             self.n = n
-            if self.platform_specific_reporter:
-                self.platform_specific_reporter.success()
-                self.platform_specific_reporter.__exit__(None, None, None)
-                self.platform_specific_reporter = None
             self.close()
 
     def update(self, n=1):
@@ -1277,9 +1253,6 @@ class tqdm(Comparable):
         # Prevent multiple closures
         self.disable = True
 
-        if self.platform_specific_reporter:
-            self.platform_specific_reporter.clear()
-
         # decrement instance pos and remove from internal set
         pos = abs(self.pos)
         self._decr_instances(self)
@@ -1306,8 +1279,6 @@ class tqdm(Comparable):
[…]
