### 2019-11-22 e3ba1c5
minor contributing note abotu snap automation
