### 2019-11-23 f7a4b06
document snap channels
