### 2019-11-23 d93cf30
officially document float n/total guidelines
- fixes #802
