### 2019-11-23 127af5c
fix most tests and logic
