+++ tqdm/utils.py
@@ -360,8 +360,10 @@ def disp_trim(data, length):
     if len(data) == disp_len(data):
         return data[:length]
 
+    ansi_present = bool(RE_ANSI.search(data))
     while disp_len(data) > length:  # carefully delete one char at a time
         data = data[:-1]
-    if RE_ANSI.search(data):  # assume ANSI reset is required
-        return data + "\033[0m"
+    if ansi_present and bool(RE_ANSI.search(data)):
+        # assume ANSI reset is required
+        return data if data.endswith("\033[0m") else data + "\033[0m"
     return data
