### 2019-11-26 4c9638d
fix unicode format args
- fixes #851
