### 2019-11-26 ac2d5f9
allow float imprecision
- fixes #849
