+++ tqdm/std.py
@@ -370,7 +370,7 @@ class tqdm(Comparable):
         """
 
         # sanity check: total
-        if total and n > total:
+        if total and n >= (total + 0.5):  # allow float imprecision (#849)
             total = None
 
         # apply custom scale if necessary
