### 2019-11-28 d091635
added a missing import
added from tqdm import tqdm in the first example of Description and additional stats. Required import for working of second part of code.
