### 2019-11-30 4c48242
refactor common into utils.ObjectWrapper
- expose utils.DummyTqdmFile
- expose utils.CallbackIOWrapper
