### 2019-11-30 1b40056
document `wrapattr`, `utils.CallbackIOWrapper`
