### 2019-11-30 ef653d8
expand example of `wrapattr`
