### 2019-11-30 866180c
self-contained `wrapattr` example
