+++ README.rst
@@ -1003,9 +1003,6 @@ A reusable canonical example is given below:
     from tqdm.contrib import DummyTqdmFile
 
 
-        def isatty(self):
-            return getattr(self.file, "isatty", lambda: False)()
-
     @contextlib.contextmanager
     def std_out_err_redirect_tqdm():
         orig_out_err = sys.stdout, sys.stderr
