### 2019-12-04 03f8300
fix auto gh releases
resurface of https://github.com/travis-ci/dpl/issues/914
