### 2019-12-05 f422502
fix floating point imprecision raising errors
- fixes #859
