+++ tqdm/std.py
@@ -143,7 +143,9 @@ class Bar(object):
     BLANK = "  "
 
     def __init__(self, frac, default_len=10, charset=UTF):
-        assert 0 <= frac <= 1
+        if not (0 <= frac <= 1):
+            warn("clamping frac to range [0, 1]", TqdmWarning, stacklevel=2)
+            frac = max(0, min(1, frac))
         assert default_len > 0
         self.frac = frac
         self.default_len = default_len
