### 2019-12-05 ffb706e
add tests for float total < n
