### 2019-12-09 d8152a4
fix pandas unhashable func
- fixes #862
