+++ tqdm/std.py
@@ -732,7 +732,10 @@ class tqdm(Comparable):
                         " Use keyword arguments instead.",
                         fp_write=getattr(t.fp, 'write', sys.stderr.write))
 
-                func = df._is_builtin_func(func)
+                try:
+                    func = df._is_builtin_func(func)
+                except TypeError:
+                    pass
 
                 # Define bar updating wrapper
                 def wrapper(*args, **kwargs):
