### 2019-12-09 a7c565e
add pandas unhashable test
