### 2019-12-19 2bd851b
document notebook `ncols`
- fixes #276
