### 2019-12-19 e5863ff
auto-detect epochs and batches
