### 2019-12-19 8431d2c
patch py2 threading tests
