+++ tqdm/tests/tests_synchronisation.py
@@ -3,7 +3,9 @@ from tqdm import tqdm, trange, TMonitor
 from tests_tqdm import with_setup, pretest, posttest, SkipTest, \
     StringIO, closing
 from tests_tqdm import DiscreteTimer, cpu_timify
+from tests_perf import retry_on_except
 
+import sys
 from time import sleep
 from threading import Event
 
@@ -188,6 +190,8 @@ def test_imap():
     assert res[-1] == 100
 
 
+# py2: locks won't propagate to incr_bar so may cause `AttributeError`
+@retry_on_except(n=3 if sys.version_info < (3,) else 1)
 @with_setup(pretest, posttest)
 def test_threadpool():
     """Test concurrent.futures.ThreadPoolExecutor"""
