### 2019-12-19 aa7e631
keras notebook auto-epochs
