### 2019-12-19 b45f248
document keras `fit(verbose=0)`
