+++ .meta/.readme.rst
@@ -665,7 +665,7 @@ A ``keras`` callback is also available:
 
     ...
 
-    model.fit(..., callbacks=[TqdmCallback()])
+    model.fit(..., verbose=0, callbacks=[TqdmCallback()])
 
 IPython/Jupyter Integration
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
