### 2019-12-20 3c42b02
better keras tests
