### 2019-12-20 b7dcfbf
fix batch scaling
