### 2019-12-20 2902de2
keras missing pop('size')
