### 2019-12-20 588eab9
minor keras tidy
