### 2020-01-15 07c3bce
Update notebook.py
