### 2020-01-15 3d2d3c0
Update notebook.py
