### 2020-01-17 589eafa
notebook:disable=None => False
- closes #880
