### 2020-01-17 966572d
initial convenience map functions
- related to #225
