### 2020-01-17 5f4a4c1
neater backward-compat
