### 2020-01-17 9332c16
add `contrib.wraps.tenumerate`
- closes #840
