### 2020-01-17 39f3a1c
add `contrib.wraps` tests
