### 2020-01-17 2636098
fix `ndenumerate`
- see https://github.com/tqdm/tqdm/issues/402#issuecomment-309747737
- related #480
