### 2020-01-17 9c2cd9c
tidy `contrib.wraps.tenumerate`
