+++ tqdm/contrib/wraps.py
@@ -14,16 +14,15 @@ def tenumerate(iterable, start=0, total=None, **tqdm_kwargs):
     """
     Equivalent of `numpy.ndenumerate` or builtin `enumerate`.
     """
-    _enumerate = enumerate
     try:
         import numpy as np
     except ImportError:
         pass
     else:
         if isinstance(iterable, np.ndarray):
-            _enumerate = np.ndenumerate
-    return tqdm(
-        _enumerate(iterable), total=total or len(iterable), **tqdm_kwargs)
+            return tqdm(np.ndenumerate(iterable),
+                        total=total or len(iterable), **tqdm_kwargs)
+    return enumerate(tqdm(iterable, **tqdm_kwargs))
 
 
 def _tzip(iter1, *iter2plus, **tqdm_kwargs):
