### 2020-01-17 73e9d1e
expose `tqdm_class` in `contrib.wraps`
