+++ tqdm/contrib/wraps.py
@@ -1,7 +1,7 @@
 """
 Thin wrappers around common functions.
 """
-from tqdm.auto import tqdm
+from tqdm.auto import tqdm as tqdm_auto
 from copy import deepcopy
 import functools
 import sys
@@ -10,9 +10,14 @@ __author__ = {"github.com/": ["casperdcl"]}
 __all__ = ['tenumerate', 'tzip', 'tmap', 'thread_map', 'process_map']
 
 
-def tenumerate(iterable, start=0, total=None, **tqdm_kwargs):
+def tenumerate(iterable, start=0, total=None, tqdm_class=tqdm_auto,
+               **tqdm_kwargs):
     """
     Equivalent of `numpy.ndenumerate` or builtin `enumerate`.
+
+    Parameters
+    ----------
+    tqdm_class  : [default: tqdm.auto.tqdm].
     """
     try:
         import numpy as np
@@ -20,22 +25,32 @@ def tenumerate(iterable, start=0, total=None, **tqdm_kwargs):
         pass
     else:
         if isinstance(iterable, np.ndarray):
-            return tqdm(np.ndenumerate(iterable),
-                        total=total or len(iterable), **tqdm_kwargs)
-    return enumerate(tqdm(iterable, **tqdm_kwargs))
+            return tqdm_class(np.ndenumerate(iterable),
+                              total=total or len(iterable), **tqdm_kwargs)
+    return enumerate(tqdm_class(iterable, **tqdm_kwargs))
 
 
 def _tzip(iter1, *iter2plus, **tqdm_kwargs):
     """
     Equivalent of builtin `zip`.
+
+    Parameters
+    ----------
+    tqdm_class  : [default: tqdm.auto.tqdm].
     """
-    for i in zip(tqdm(iter1, **tqdm_kwargs), *iter2plus):
+    kwargs = deepcopy(tqdm_kwargs)
+    tqdm_class = kwargs.pop("tqdm_class", tqdm_auto)
+    for i in zip(tqdm_class(iter1, **tqdm_kwargs), *iter2plus):
         yield i
 
 
 def _tmap(function, *sequences, **tqdm_kwargs):
     """
     Equivalent of builtin `map`.
+
+    Parameters
+    ----------
+    tqdm_class  : [default: tqdm.auto.tqdm].
     """
     for i in _tzip(*sequences, **tqdm_kwargs):
         yield function(*i)
@@ -44,17 +59,26 @@ def _tmap(function, *sequences, **tqdm_kwargs):
 def _executor_map(PoolExecutor, fn, *iterables, **tqdm_kwargs):
     """
     Implementation of `thread_map` and `process_map`.
+
+    Parameters
+    ----------
+    tqdm_class  : [default: tqdm.auto.tqdm].
     """
     kwargs = deepcopy(tqdm_kwargs)
     kwargs.setdefault("total", len(iterables[0]))
+    tqdm_class = kwargs.pop("tqdm_class", tqdm_auto)
     with PoolExecutor(max_workers=kwargs.pop("max_workers", None)) as ex:
-        return list(tqdm(ex.map(fn, *iterables), **kwargs))
+        return list(tqdm_class(ex.map(fn, *iterables), **kwargs))
 
 
 def thread_map(fn, *iterables, **tqdm_kwargs):
     """
     Equivalent of `list(map(fn, *iterables))`
     driven by `concurrent.futures.ThreadPoolExecutor`.
+
+    Parameters
+    ----------
+    tqdm_class  : [default: tqdm.auto.tqdm].
     """
     from concurrent.futures import ThreadPoolExecutor
     return _executor_map(ThreadPoolExecutor, fn, *iterables, **tqdm_kwargs)
[…]
