### 2020-01-17 cb95769
fix codacy issues
