### 2020-01-17 41a6f8b
fix py34
- see https://docs.python.org/3/library/concurrent.futures.html#threadpoolexecutor
