+++ tqdm/contrib/wraps.py
@@ -4,6 +4,7 @@ Thin wrappers around common functions.
 from tqdm.auto import tqdm as tqdm_auto
 from copy import deepcopy
 import functools
+import os
 import sys
 
 __author__ = {"github.com/": ["casperdcl"]}
@@ -67,7 +68,8 @@ def _executor_map(PoolExecutor, fn, *iterables, **tqdm_kwargs):
     kwargs = deepcopy(tqdm_kwargs)
     kwargs.setdefault("total", len(iterables[0]))
     tqdm_class = kwargs.pop("tqdm_class", tqdm_auto)
-    with PoolExecutor(max_workers=kwargs.pop("max_workers", None)) as ex:
+    max_workers = kwargs.pop("max_workers", min(32, os.cpu_count() + 4))
+    with PoolExecutor(max_workers=max_workers) as ex:
         return list(tqdm_class(ex.map(fn, *iterables), **kwargs))
 
 
