### 2020-01-17 078980c
fix `contrib.wraps` auto `cpu_count`
