### 2020-01-17 cf74393
add missing contrib module to Makefile
