### 2020-01-19 e06a544
Don’t strip postfix to enable padding it
Currently, doing `obar.set_postfix(header=f'{header:<30}')` does nothing.
This restores the ability to have things not jump all over the place.
