### 2020-01-19 eb9377f
add itertools stub
- addresses #225
