### 2020-01-19 59a48ba
minor linting
