### 2020-01-19 fdf6a2e
itertools test stub
