+++ tqdm/tests/tests_concurrent.py
@@ -1,5 +1,5 @@
 """
-Tests for `tqdm.contrib`
+Tests for `tqdm.contrib.concurrent`.
 """
 from tqdm.contrib.concurrent import thread_map, process_map
 from tests_tqdm import with_setup, pretest, posttest, SkipTest, StringIO, \
