### 2020-01-19 8cc777f
complete itertools tests
