+++ tqdm/tests/tests_itertools.py
@@ -6,11 +6,22 @@ from tests_tqdm import with_setup, pretest, posttest, StringIO, closing
 import itertools
 
 
+class NoLenIter(object):
+    def __init__(self, iterable):
+        self._it = iterable
+
+    def __iter__(self):
+        for i in self._it:
+            yield i
+
+
 @with_setup(pretest, posttest)
 def test_product():
     """Test contrib.itertools.product"""
     with closing(StringIO()) as our_file:
         a = range(9)
-        b = a[::-1]
-        assert list(product(a, b, file=our_file)) == \
-            list(itertools.product(a, b))
+        assert list(product(a, a[::-1], file=our_file)) == \
+            list(itertools.product(a, a[::-1]))
+
+        assert list(product(a, NoLenIter(a), file=our_file)) == \
+            list(itertools.product(a, NoLenIter(a)))
