+++ tqdm/contrib/itertools.py
@@ -5,6 +5,8 @@ from __future__ import absolute_import
 from tqdm.auto import tqdm as tqdm_auto
 from copy import deepcopy
 import itertools
+__author__ = {"github.com/": ["casperdcl"]}
+__all__ = ['product']
 
 
 def product(*iterables, **tqdm_kwargs):
