### 2020-01-21 #1277 @lrq3000
Hello,

I come late to the party, but I strongly disagree with what @casperdcl did. And for how this paper submission was handled. I will keep it short, as these points should be self-explanatory:

1. Although he did write the paper on its own, and thus deserves credit, he did NOT contact any of the main co-developers of the software, which I am only one (I think major, but nevertheless only one) of the many contributors who could make this happen. They also deserved to be invited to contribute to the paper, and to get due credit. At the very least, they should be mentioned in an Acknowledgements section.
2. Casper is not even the original author. Although we had a hard time reaching the original author, for publishing, we ought to try harder before publishing a paper. I can not even find a mention of his name: Noam Yorav-Raphael. That is unethical.
3. A lot (most?) of the foundational code leading to the current form of tqdm was made by me (in particular the object-oriented approach which was initially dismissed by Casper and was necessary for most use cases tqdm can cover nowadays). Most features that were implemented after I departed (because of a change of to a more restrictive license Casper forcefully did on his own), such as `tqdm_auto`, and which now profit from a quite popular recognition were initially proposed by me, and were pushed/slowed down by Casper (another reason I left development).

Side-note: don't just look at the commits, but at the PR from where they come from, Casper had the habit of always merging himself without properly attributing to the real author due to a misconfiguration of the git client (at least that's what I thought, but I can see from the current history that there is no credit to other devs).

And I can probably raise up some other points, but this is I think enough to raise some big red flags, that should have been caught.

Nevertheless, I recognize Casper's dedication to the project, he's done a great work, and whatever his motives, I am glad the project is still maintained.

But the way Casper is trying to make this a one-man project, stripping due recognition to fellow co-developers and taking full control as he did (including the license), is highly questionable. The paper as it is can only be considered unethical and should be amended or retracted. And it's a shame, because it could have been a great opportunity to bring back on board fellow old-time developers, but Casper once again chose to do it solo.

Pinging old time/foundational developers: @noamraph, @obiwanus, @kmike, @hadim.
