### 2020-01-21 #1277 @kmike
I haven't contributed much to tqdm, and don't need any recognition for tqdm work.

However, I've been involved in a similar case in past, where a technical report was published for a piece of software (theano). See https://arxiv.org/abs/1605.02688 - it was handled in a different way - all contributions who made non-trivial fixes or improvements (so, no typo fixes) were asked for comments on the paper itself, and added as authors (if they're ok with that). It is a relatively small effort to list all people who helped, and a super-nice gesture, which still makes me smile. No way I'd be Yoshua Bengio's, Ian Goodfellow's, Xavier Glorot's, etc. co-author otherwise :)

Recognizing people who made important contributions (not me :) is a must though, regardless of when these contributions happened, especially if that's the first technical report. As I recall, for Theano there were several papers, separated by a few years - in this case I think that's fine to give recognition only to people who contributed since the last report, and cite previous work.
