### 2020-01-21 #1277 @casperdcl
Hello, just like to say I'm happy to hear from @lrq3000 but must clear up a few points.

> I come late to the party [...] Although he did write the paper on its own [...] he did NOT contact beforehand any of the main co-developers [...] **That is highly unethical**

As far as I know, in order of first contribution, @noamraph, @hadim, @kmike, @casperdcl, and @lrq3000 were instrumental in resuscitating the current `tqdm` library in 2015 (https://github.com/tqdm/tqdm/pull/35). However as @lrq3000 mentions, the original author (@noamraph, original repo https://github.com/noamraph/tqdm with 35 commits, of which 11 commits (27 lines of code) survive in the current repo, making up under 1%, vis https://github.com/tqdm/tqdm#contributions) abandoned the project and has yet to accept any invites to continue maintenance. Meanwhile @kmike, @hadim, and then @lrq3000 stopped contributing, leaving me (@casperdcl) with the sole burden of maintenance. As of paper writing, it had been **years** since their last contribution. In particular, I had `@mention`ed @lrq3000 in several issues/PRs and sent several emails to him over the last couple of years which have received no response, leading me to the obvious conclusion that he'd abandoned the project too. Still wanting to acknowledge contributions and provide public incentivisation to contribute again, I explicitly `@mention`ed him in this review (vis https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-469222023) to no avail.

> I am only one [...] I think major [...] of the many contributors who could make this happen

"Make this happen" is a bit ambiguous. There are many contributors of OS projects, including `tqdm`. I ensure that the documentation (https://github.com/tqdm/tqdm#contributions) explicitly refers to as many as possible using dynamic updating links, avatars, tables and references. However a loose definition of "contributing" does not satisfy the JOSS requirements for authorship. Again, I reiterate that during review I also hoped that @lrq3000, being the only other arguably major contributor to current library, would rejoin.

> they should be mentioned in an Acknowledgements section, regardless of whether they would have contributed to the paper or not.

I am unclear about whether an Acknowledgements section can be included, and what exactly should be in it. Perhaps simply a link to https://github.com/tqdm/tqdm#contributions? The issue is that there are many ongoing contributions from new people. For this purpose, in this review we settled on simply adding "`tqdm` developers" as an author of the paper; a similar tactic used in other JOSS papers.

> A lot (most?) of the foundational code leading to the current form of tqdm was made by [@lrq3000] [...] which was initially dismissed by [@casperdcl] [...] features that were implemented after [@lrq3000] departed [...] were initially proposed by [@lrq3000] [...] were pushed/slowed down by [@casperdcl] (another reason [@lrq3000] left development)

I think this is a highly contestable point which I disagree with, but probably both off topic and also not worth debating given my response above.
I'd like to make it clear that I would've always been willing to discuss project directions or reservations of contributors. The statements above are completely surprising to me. Note that an incredibly widely used project (millions of downloads per month; often in the world's top 10 Python packages) needs a very pedantic level of care and attention to detail to avoid breaking changes that could affect millions. This makes review slow and frustrating (both for maintainers and contributors) but **I don't see how it could concern JOSS editors**.

> change of to a more restrictive license [@casperdcl] forcefully did on his own

This is not true. This being said, again **I fail to see how this could concern JOSS editors**.

There's a long history of debate and discussion over licensing, initially kicked off by a very concerning commit message slipped through by a minor contributor. I had changed the licence to negate any negative consequences. Subsequently, @hadim, @kmike, @lrq3000 and I actively participated in discussion (both on public GitHub issues https://github.com/tqdm/tqdm/pull/124, https://github.com/tqdm/tqdm/issues/121 and via private emails) and eventually *settled upon the current licence together*. This included explicit public confirmation from @lrq3000. The so-called resulting "more restrictive licence" is FLOSS so cannot be called restrictive.

> [@casperdcl] had the habit of always merging himself without properly attributing to the real author due to a misconfiguration of the git client [...] but [...] can see from the current history that there are still no attributions to other devs nowadays

This is absolutely not true in any way.

I'd suggest the editors ask for evidence to backup this claim or politely request for it to be formally retracted.

I can think of only one instance where a PR by @lrq3000 was mysteriously closed without the option to re-open (https://github.com/tqdm/tqdm/pull/321), and in that case the replacement (https://github.com/tqdm/tqdm/pull/543) explicitly mentioned this fact.

In addition, I frequently agree with a contributor's ideas but (having in-depth knowledge as the sole maintainer) disagree with implementation. In these cases I have always made it a point to add my own correctional commits to PRs rather than outright remove original author's contributions. In cases where forked PRs do not "allow commits from maintainers", I go out of my way to rebase/cherry-pick commits onto a local branch, once again so that people can show up in https://github.com/tqdm/tqdm/graphs/contributors.

You can also see from https://github.com/tqdm/tqdm/commits/master?before=cc53d86f86cef63ba8dbbfb8e9cb27d4f247fd20+70 that original PR authors do indeed show up because of my maintenance efforts.

> And I can probably raise up some other points

I strongly encourage you to raise any and all reservations you believe to be valid in the interest of transparency and fruitful discussion.

> [@casperdcl]'s wish of showing this as a one-man project, stripping due recognition to fellow co-developers and taking full control

Again, absolutely defamatory and I must insist on a retraction. Even if I had no claim to authorship whatsoever, any claims to know what my "wishes" are are completely speculative and unfounded.

> a great opportunity to bring back on board fellow old-time developers and streighten the community's bonds around this wonderful project.

For avoidance of doubt I would like to state that this was always precisely my intention. Given that there has been some responses in this review (however delayed and negative) I think this has actually been somewhat successful.

> PS: for those reading the emails notifications, please read the post on Github, I have made some changes to clarify.

Please do mention if any points are missing responses.

> [@kmike] [has] been involved in a similar case in past [...] all contributions who made non-trivial fixes or improvements [...] were asked for comments on the paper itself, and added as authors (if they're ok with that).

I'm not sure about the definition of "non-trivial" but this sounds like something this journal may potentially consider adding to their guidelines in future.

I would also be happy to retrospectively change this paper too upon request.

I leave the decision (if any) to the editors and sympathise that the question of authorship and acknowledgements (paper writer, code contributor, major contributor, dev, co-dev, etc) is a complex issue.

----

Finally, in case the current `tqdm` master branch (https://github.com/tqdm/tqdm/commit/cc53d86f86cef63ba8dbbfb8e9cb27d4f247fd20):

```bash
tqdm (master)$ git fame -wMC
Blame: 100%|██████████| 77/77 [00:02<00:00, 27.81file/s]
Total commits: 1259
Total ctimes: 1335
Total files: 203
Total loc: 2869
```

| Author                     |   loc |   coms |   fils |  distribution   |
|:---------------------------|------:|-------:|-------:|:----------------|
| Casper da Costa-Luis       |  2234 |    851 |     66 | 77.9/67.6/32.5  |
| Stephen Larroque           |   369 |    202 |     23 | 12.9/16.0/11.3  |
| Matthew Stevens            |    31 |      3 |      2 | 1.1/ 0.2/ 1.0   |
| Noam Yorav-Raphael         |    27 |     11 |      8 | 0.9/ 0.9/ 3.9   |
| Guangshuo Chen             |    23 |     18 |      6 | 0.8/ 1.4/ 3.0   |
| Hadrien Mary               |    17 |     31 |     10 | 0.6/ 2.5/ 4.9   |
| Mikhail Korobov            |    16 |     11 |      7 | 0.6/ 0.9/ 3.4   |
| Daniel Panteleit           |    14 |      2 |      2 | 0.5/ 0.2/ 1.0   |
| Kyle Altendorf             |    13 |     31 |      3 | 0.5/ 2.5/ 1.5   |
| Ivan Ivanov                |    10 |     11 |      3 | 0.3/ 0.9/ 1.5   |
| Martin Zugnoni             |    10 |      3 |      3 | 0.3/ 0.2/ 1.5   |
| Matthew D. Pagel           |     8 |      3 |      3 | 0.3/ 0.2/ 1.5   |
| Marcel Bargull             |     7 |      6 |      2 | 0.2/ 0.5/ 1.0   |
| Hugo                       |     5 |      2 |      5 | 0.2/ 0.2/ 2.5   |
| Greg Gandenberger          |     5 |      1 |      2 | 0.2/ 0.1/ 1.0   |
| Min ho Kim                 |     5 |      1 |      5 | 0.2/ 0.1/ 2.5   |
| Veith Röthlingshöfer       |     5 |      1 |      2 | 0.2/ 0.1/ 1.0   |
| Thomas A Caswell           |     4 |      1 |      4 | 0.1/ 0.1/ 2.0   |
| immerrr                    |     4 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| James E. King III          |     4 |      1 |      3 | 0.1/ 0.1/ 1.5   |
| Johannes Hansen            |     3 |      1 |      2 | 0.1/ 0.1/ 1.0   |
| Max Nordlund               |     3 |      2 |      3 | 0.1/ 0.2/ 1.5   |
| Adnan Umer                 |     3 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| Antony Lee                 |     3 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| Orivej Desh                |     3 |      2 |      1 | 0.1/ 0.2/ 0.5   |
| Charles Newey              |     3 |      3 |      2 | 0.1/ 0.2/ 1.0   |
| Mike Kutzma                |     2 |      1 |      2 | 0.1/ 0.1/ 1.0   |
| zz                         |     2 |      1 |      2 | 0.1/ 0.1/ 1.0   |
| Kuang-che Wu               |     2 |      1 |      2 | 0.1/ 0.1/ 1.0   |
| David W.H. Swenson         |     2 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| Jesus Cea                  |     2 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| shongololo                 |     2 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| Alexander                  |     2 |      2 |      2 | 0.1/ 0.2/ 1.0   |
| Julien Chaumont            |     2 |      1 |      2 | 0.1/ 0.1/ 1.0   |
| Yaroslav Halchenko         |     2 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| Tomas Ostasevicius         |     2 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| RedBug312                  |     2 |      2 |      1 | 0.1/ 0.2/ 0.5   |
| ReadmeCritic               |     2 |      1 |      2 | 0.1/ 0.1/ 1.0   |
| William Turner             |     2 |      1 |      1 | 0.1/ 0.1/ 0.5   |
| James Lu                   |     2 |      3 |      2 | 0.1/ 0.2/ 1.0   |
| Igor Ljubuncic             |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| stonebig                   |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| FichteFoll                 |     1 |      2 |      1 | 0.0/ 0.2/ 0.5   |
| Sergei Izmailov            |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| Cheng Chen                 |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| Staffan Malmgren           |     1 |      2 |      1 | 0.0/ 0.2/ 0.5   |
| Darin                      |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| Todd                       |     1 |      2 |      1 | 0.0/ 0.2/ 0.5   |
| Peter VandeHaar            |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| Anurag Pandey              |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| Robert Krzyzanowski        |     1 |      1 |      1 | 0.0/ 0.1/ 0.5   |
| Alex Rothberg              |     1 |      2 |      1 | 0.0/ 0.2/ 0.5   |
| Rafael Lukas Maers         |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Jan Schlüter               |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Lev Velykoivanenko         |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Pablo Zivic                |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Albert Kottke              |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Edward Betts               |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Jack McCracken             |     0 |      2 |      0 | 0.0/ 0.2/ 0.0   |
| JoshKarpel                 |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Dyno Fu                    |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Socialery                  |     0 |      2 |      0 | 0.0/ 0.2/ 0.0   |
| Jose Tiago Macara Coutinho |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Arun Persaud               |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Andrey Portnoy             |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Riccardo Coccioli          |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| David Bau                  |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Ford Hurley                |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| zed                        |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Dénes Türei                |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Shirish Pokharel           |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
| Fabian Dill                |     0 |      2 |      0 | 0.0/ 0.2/ 0.0   |
| Javi Merino                |     0 |      1 |      0 | 0.0/ 0.1/ 0.0   |
