### 2020-01-21 #1277 @arfon
@lrq3000 @kmike - thank you for reporting this issue. We're going to conduct our own investigation and report back our findings. Given JOSS is run by a volunteer editorial team this may take a few weeks.

Arfon Smith
Editor-in-Chief, JOSS
