### 2020-01-21 #1277 @arfon
Please note that I'm also locking this issue for now while we investigate the claims here.
