+++ tqdm/notebook.py
@@ -192,8 +192,10 @@ class tqdm_notebook(std_tqdm):
         kwargs['gui'] = True
         kwargs.setdefault('bar_format', '{l_bar}{bar}{r_bar}')
         kwargs['bar_format'] = kwargs['bar_format'].replace('{bar}', '<bar/>')
+        display = kwargs.get('display')
+        kwargs['display'] = display if display is not None else False
         super(tqdm_notebook, self).__init__(*args, **kwargs)
-        if self.disable is not False or not kwargs['gui']:
+        if self.disable or not kwargs['gui']:
             return
 
         # Get bar width
@@ -207,7 +209,7 @@ class tqdm_notebook(std_tqdm):
         self.sp = self.display
 
         # Print initial bar state
-        if self.disable is not False:
+        if self.disable:
             self.display()
 
     def __iter__(self, *args, **kwargs):
