### 2020-01-22 17db575
Update notebook.py
