### 2020-01-24 6afc159
allow tensorflow.keras without keras
- fixes #885
