### 2020-01-24 319aeba
add tensorflow without keras CI test
