+++ tox.ini
@@ -5,7 +5,7 @@
 
 [tox]
 # deprecation warning: py{26,32,33,34}
-envlist = py{26,27,33,34,35,36,37,py,py3}, flake8, setup.py, perf
+envlist = py{26,27,33,34,35,36,37,py,py3}, tf-no-keras, perf, flake8, setup.py
 
 [coverage]
 deps =
@@ -68,6 +68,13 @@ commands = {[extra]commands}
 # py34-compatible pandas
 deps = {[extra]deps}
 commands = {[extra]commands}
+
+[testenv:tf-no-keras]
+deps =
+    {[extra]deps}
+    tensorflow
+commands =
+    nosetests --with-timer tqdm/tests/tests_keras.py -d -v
 
 [testenv:perf]
 deps =
