+++ tqdm/keras.py
@@ -2,18 +2,17 @@ from __future__ import absolute_import, division
 from .auto import tqdm as tqdm_auto
 from copy import deepcopy
 try:
-    from keras.callbacks import Callback
+    import keras
 except ImportError as e:
     try:
-        from tensorflow.keras.callbacks import Callback
+        from tensorflow import keras
     except ImportError:
         raise e
-
 __author__ = {"github.com/": ["casperdcl"]}
 __all__ = ['TqdmCallback']
 
 
-class TqdmCallback(Callback):
+class TqdmCallback(keras.callbacks.Callback):
     """`keras` callback for epoch and batch progress"""
     @staticmethod
     def bar2callback(bar, pop=None, delta=(lambda logs: 1)):
