### 2020-01-24 2c3b64c
actually add tf-no-keras to travis
