### 2020-01-24 a518817
contrib:fix numpy.ndenumerate auto-total
