+++ tqdm/contrib/__init__.py
@@ -37,7 +37,7 @@ def tenumerate(iterable, start=0, total=None, tqdm_class=tqdm_auto,
     else:
         if isinstance(iterable, np.ndarray):
             return tqdm_class(np.ndenumerate(iterable),
-                              total=total or len(iterable), **tqdm_kwargs)
+                              total=total or iterable.size, **tqdm_kwargs)
     return enumerate(tqdm_class(iterable, **tqdm_kwargs), start)
 
 
