### 2020-01-24 30c2d8f
examples/parallel_bars with contrib.concurrent
