### 2020-01-24 ad87480
examples with `tqdm.auto`
