+++ examples/include_no_requirements.py
@@ -1,6 +1,6 @@
-# How to import tqdm without enforcing it as a dependency
+# How to import tqdm in any frontend without enforcing it as a dependency
 try:
-    from tqdm import tqdm
+    from tqdm.auto import tqdm
 except ImportError:
 
     def tqdm(*args, **kwargs):
