### 2020-01-25 #1277 @arfon
Dear @casperdcl, @lrq3000, @noamraph and other tqdm contributors. The JOSS editorial board has considered the concerns raised by @lrq3000 and agree the authorship of this paper is problematic. JOSS’s [policy on authorship states](https://joss.theoj.org/about#ethics):

> Author lists must be correct and complete. All listed authors must have made a contribution to the work, and all significant contributors should be included in the author list.

While we recognize that @casperdcl made an [attempt to include](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-469222023) @lrq3000 as an author after @GregaVrbancic [identified](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-469208417) @lrq3000’s contributions during their review, we’ve concluded that this effort was insufficient and therefore a violation of JOSS’s ethical standards.

### How we plan to handle this submission

We ask that the tqdm contributors resolve this authorship question between yourselves and report back here on the agreed resolution. Past submissions have successfully used a public thread on their repository (e.g., an issue posted on https://github.com/tqdm/tqdm) inviting past authors to be authors.

**If no acceptable resolution can be found then we will retract this paper in one calendar month from today (25 February 2020).**

Should the tqdm team arrive at an acceptable solution then we would handle the paper update as follows:

- Retract the original paper.
- Publish a new tqdm JOSS paper with the updated author list/acknowledgements. This paper will be fast-tracked through the review process (likely a quick review by one or more of our editors).
- Ensure that the Crossref metadata associated with the new paper includes a [`isVersionOf`](https://support.crossref.org/hc/en-us/articles/214357426-Relationships-between-DOIs-and-other-objects) identifier making it clear that the new paper is a version of the earlier record (https://doi.org/10.21105/joss.01277).

### Which of the tqdm contributors should be included on this paper?

Deciding authorship can be challenging and ultimately authors need to decide this between themselves. That said, in this situation, the JOSS editorial team expects any resolution here to include the explicit consent of @casperdcl, @lrq3000, & @noamraph.

We trust this is an acceptable solution for all parties involved. JOSS is a civilized platform, and we value civility in resolving disputes. If in this one-month window the tqdm contributors are able to resolve the co-authorship issue, peacefully and amicably, then an updated version of this submission can be published in place of the current version, which will be retracted. If not, then we will be forced to retract this paper with no replacement.

Arfon Smith on behalf of the JOSS editorial team.
