### 2020-01-25 #1277 @hadim
I guess I am kind of late to the party. I am the person who forked the original repo to `tqdm/tqdm` and maintained it for a few weeks before I let the other devs taking over.

I am not sure I should be on that paper but at least my name should be cited in it wherever it's in an acknowledgment paragraph or a an history paragraph about the birth and re-birth of the project (this section would be cool to have actually).

That being said, maybe we should continue that discussion at https://github.com/tqdm/tqdm.
