### 2020-01-25 #1277 @arfon
> That being said, maybe we should continue that discussion at https://github.com/tqdm/tqdm.

Yes, please discuss this topic in https://github.com/tqdm/tqdm and report back here when you have a resolution.
