### 2020-01-25 #1277 @casperdcl
@arfon thanks for your time and review.

> the JOSS editorial team expects any resolution here to include the explicit consent of @casperdcl, @lrq3000, & @noamraph.

Clearly we shouldn't have any issues with @lrq3000 (and @hadim/@kmike) going forward but I'm concerned about @noamraph - I've never managed to contact him and as far as I know nobody else has either, despite very serious efforts (year ago, we were about to rename the project - https://github.com/tqdm/tqdm/issues/2, https://github.com/tqdm/tqdm/issues/30, https://github.com/tqdm/tqdm/issues/33, and eventually PyPI admins granted access to the old project). If you also look at https://github.com/tqdm/tqdm/issues?q=involves%3Anoamraph you'll see there's no participation either.

Given the current ~1% code contribution, nonexistent website/wiki/paper/testing/framework contribution, and last contribution dating back to Jan 2014, would it be an issue if we get no response?
