### 2020-01-25 #1277 @lrq3000
Yes we can surely work it out together.

About Noam, the total code contribution is irrelevant IMO in his case as he invented the very foundation and concept of this project, so the code quantity is irrelevant in his case.

But indeed he may be hard to reach, although I'll try my best to. In the worst case, if he cannot be reached despite our best attempts, I propose Noam to be mentioned in context of the introduction or a history section. Would this be a ok solution for the JOSS board?
