### 2020-01-25 #1277 @arfon
> Clearly we shouldn't have any issues with @lrq3000 (and @hadim/@kmike) going forward but I'm concerned about @noamraph

I’ve actually heard directly from @noamraph over email twice in the last couple of days. I will ask him to participate in this discussion.
