### 2020-01-25 #1277 @lrq3000
Dear @noamraph, I am very pleased and honored to talk with you, that's awesome! I will open an issue on tqdm repo before tomorrow with some propositions so we can further discuss there. Thank you for coming in and no worries about the time! :D
