### 2020-01-26 tqdm/tqdm#887 @casperdcl
update paper
Discuss updating [paper](https://github.com/tqdm/tqdm/blob/master/examples/paper.md) (archived at https://doi.org/10.21105/joss.01277) as per https://github.com/openjournals/joss-reviews/issues/1277.

- in particular, authors and acknowledgements
- upon request of Arfon Smith (@arfon) on behalf of the JOSS editorial team
- in addition to assignees; cc @noamraph, @Benjamin-Lee, @GregaVrbancic
