### 2020-01-26 tqdm/tqdm#887 @casperdcl
In an attempt to avoid any debate (so we can focus on tackling other open issues), hopefully we can get an immediate decision:

Reading the Editor-in-Chief's comments very carefully, in particular where and when others are `@mention`ed, the *impression* I get is he may be implying authors: @casperdcl, @lrq3000; acknowledgements: @noamraph. I might be wrong and it would be great if @arfon could confirm either way. Further I would suggest (as per @hadim's request) to also add @hadim in the acknowledgements. I would also suggest adding @kmike if he is amenable. I am unsure about adding others but will not refuse any suggestions.

To further prompt a decision:

1. I reiterate my willingness to accept the Editorial team's decisions.
1. I think one of the main responsibilities of a journal is to (with the aid of/via the reviewers and Editorial team) define and make decisions on authorship.

I understand that is a big request with a lot of responsibility and research associated, especially considering the voluntary nature of the JOSS team's work. However I would very much appreciate if at least a *suggestion* if not a decision is provided.
