### 2020-01-26 tqdm/tqdm#887 @noamraph
Hi! Here I am!

Arfon wrote: "all significant contributors should be included in the author list." I consider myself to be a significant contributor, by writing the initial working version. I acknowledge that a lot of significant work has been done since, probably with much more time spent than my initial effort. However, I still consider myself to be a significant contributor.

Cheers,
Noam
