### 2020-01-26 tqdm/tqdm#887 @casperdcl
Hey @noamraph,

Happy to hear from you.

Unfortunately though in this instance I disagree for a long list of reasons. However I really hope to avoid or at least shorten any debate if we could get a response from @arfon here...
