### 2020-01-26 tqdm/tqdm#887 @arfon
> @casperdcl, @lrq3000, tqdm developers; acknowledgements: @noamraph. I might be wrong and it would be great if @arfon could confirm either way.

Apologies if this is how my comment was interpreted, this is not what I was trying to imply because...

> I think one of the main responsibilities of a journal is to (with the aid of/via the reviewers and editorial team) define and make decisions on authorship.

I disagree. The journal's responsibility is to ensure that there has been ethical behaviour on the part of authors, reviewers, and editors. It it *not* the responsibility of the journal to decide who should be an author on a given paper. As I said in [this comment](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-578402029) authorship is a challenging subject and something that the authors/contributors of the package are best-placed to decide between themselves.

To be absolutely clear here I think if @noamraph wants to be an author, he should be.

Hopefully this is useful additional context. I am hoping this additional input from myself will be sufficient to guide you all in your decision. As I said over in https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-578402029, this is a conversation that needs to happen between you all, and really isn't something that I can/or should be steering.
