### 2020-01-26 tqdm/tqdm#887 @arfon
Hi all, I'd like to introduce @gkthiruvathukal who is a member of the JOSS editorial team. @gkthiruvathukal has kindly volunteered to act as a neutral mediator/third party in these discussions. As past Editor in Chief of [CiSE](https://www.computer.org/csdl/magazines/cs) he has a wealth of experience with discussions such as these.
