### 2020-01-26 tqdm/tqdm#887 @gkthiruvathukal
@arfon Thanks for the introduction. I've taken the time to look at tqdm. It's great to see all of the work that has gone into this effort. As the neutral mediator (and observer in general) I'd like to encourage us to start from the `git fame` output shown on the landing page.

The `git fame` output shows that there are multiple contributors who probably should be recognized as co-authors, per JOSS guielines. I've done some preliminary analysis below.

Strong case for co-authorship:
@casperdcl
@lrq3000
@noamraph

Strong case for acknowledgments and possible co-authorship:
@altendky (top 3 on `git fame` list, possible co-author?)
@chengs
@mjstevens777
@hadim
@hmike

Again, I remind everyone that this is coming *straight from the tqdm landing page*.

In Software Engineering, I teach that not all contributions can be measured by LOC or number of commits. There are multiple ways to make an impact, and sometimes an enormous amount of impact can be made without the singular focus on LOC (or commits). For example, there can be developers who write good tests, configuration scripts, and documentation. These all make for great projects, and tqdm seems to have most of the good things in SE going for itself. This project is lucky to have such a great group of contributors, and it is my hope that you all can work to understand the importance of your *software* being well-cited as opposed to *individuals*.

@arfon and I have discussed this case, and we don't want to make decisions for you. We are both in agreement that this submission needs to recognize all co-authors in order for it to remain a JOSS submission, so if you are mentioned above, perhaps you can respond (concisely) with whether you agree with the above initial analysis. If you think you should be on a different list (co-authors or acknowledgments) than my initial triage, please make this clear in your follow up. I promise to hear everyone's input and help you to reach a conclusion that is good for the project and those who have made the most significant contributions to its success. This is a nice project and is the sort of contribution we really like to see submitted to JOSS.

I'd like to see if we can bring the authorship issues to a resolution by Friday, 31 January, so JOSS can decide on the next steps, which will involve revising the submission one way or another.
