### 2020-01-27 tqdm/tqdm#887 @altendky
@gkthiruvathukal, I have no concerns with your initial groupings for attribution.

@lrq3000's comments about lack of attribution of changes had me a little worried as I remembered some funny feelings with some of my contributions.  I think the primary missing part on my contributions was that two PRs showed as closed, not merged (#673, #622).  The commits they were closed by do show me as author but for some reason have @casperdcl as committer.  #674 (adding four words on one line...) was closed and a different solution committed without referencing me (no complaints here).  #598 (my only significant contribution) was merged per normal GitHub methods.  I closed #590 because I took a different approach in #598 that replaced it.  I won't speak to @lrq300's issues but the only comment I would have related to mine is that it would probably be good practice to use the GitHub merge button so it shows up as merged in GitHub.  It supports merge, squash-merge, and rebase so hopefully there's sufficient flexibility to satisfy personal git preferences.

Based on the infamously questionable metric of lines of code I would give @lrq3000 'author' level attribution.  https://github.com/tqdm/tqdm/graphs/contributors (5220 lines compared to @casperdcl's 23,267 lines).  It sounds like if I actually bothered to research it I would give 'author' level attribution for other reasons as well.

Based on being the original author (I think there isn't contention on this 'fact'?) I would give @noamraph 'author' level attribution.

Should I be an 'author'?  Aside from the whole conflict of interest etc in me weighing in on credit due to me...  I think it depends.  Is an author someone that authored a block of code for tqdm?  Or is an author someone that has contributed primary architecture or a large featureset?  I certainly contributed one non-trivial bug-fix.  I certainly did not design or author a significant portion of tqdm.  I suspect that the sensible answer is to put me in the 'acknowledgements' group, not co-author.  If I get co-author for my one bug-fix then we should look at everyone else that made non-trivial bug fixes and feature additions and include them as well.

Oh, to be complete, I'd put @casperdcl in the 'author' group as well.  `:]`  And no, I'm not going to bother to justify that.
