### 2020-01-27 tqdm/tqdm#887 @hadim
I would give @noamraph, @lrq3000 and @casperdcl 'author' level.

As for me, I don't think I should choose for myself rather I will let others choose for me. That being said I am fine either way. The only thing I ask is to at least be cited in a 'history' paragraph.

My contribution to the code is much more modest than the devs I've cited above but I would like to highlight that I am the one who originally forked `noamraph/tqdm` repo to create `tqdm/tqdm`. Now to be honest I don't remember exactly everything so others please feel free to tell if I am wrong.

Quickly after the creation of `tqdm/tqdm` a bunch of devs showed up (including @lrq3000, @casperdcl and @kmike ) and showed a very strong interest in the project. Everyone started to open issues and create PRs and very quickly the project became very popular. I remember giving up quickly on the development after the fork and letting the other devs maintaining tqdm.

I don't know if the project would be different today without me forking it. I guess probably not, since someone would have done it anyway at some point.

That was my modest contribution to the debate. I hope everyone will get their deserved acknowledgment/authorship without any more drama :-)
