### 2020-01-27 tqdm/tqdm#887 @gkthiruvathukal
@hadim and @altendky, thanks for your quick follow-up. I am going to continue observing, but my initial reaction is that you probably deserve more than an acknowledgment for your contributions as significant contributors. I always take an *inclusive* approach to contributions, no matter how large or small.  Hopefully, the lead contributors will chime in with whether we could live with the top contributors as minor co-authors. We can easily arrange the co-authors in any order. I don't think there is much doubt about @casperdcl, @lrq3000, and @noamraph  but am at this point hoping the three of them could agree to be co-authors and invite any *additional* authors with a strong case for inclusion.
