### 2020-01-27 tqdm/tqdm#887 @altendky
@gkthiruvathukal, agreed as far as tending towards more inclusivity.  Is there something that restricts to co-authors and acknowledgments as opposed to adding a contributors group in between?  I wouldn't feel slighted being 'merely' acknowledged but if there's any contention on the topic for others then perhaps more categories would help (I mean...  three anyways, not 17).
