### 2020-01-27 tqdm/tqdm#887 @lrq3000
Hello everyone! Thank you all (including Casper) so much for your feedbacks! Sorry for my delay, I tend to always spend more (too much?) on research before doing anything :-) Thank you very much @gkthiruvathukal and @arfon for accompanying us and your very valuable insights on this issue!

About the main issue at hand, here are my suggestions:
* I agree we should strive to be **as inclusive as possible** with the end goal of **being fair**. If we choose not everyone to be in the paper's authors list (although there is a [precedent](https://arxiv.org/abs/1605.02688) as @kmike pointed out), there is ample space in the acknowledgements. Not only for the old-time devs, but for also the new contributors since then. But we need to define a metric indeed. Personally, I would be very glad to see our fellow co-developers in the author list!
* @casperdcl has without any doubts contributed the most to this project, maintained it all this time making it a very successful package used in lots of other applications, and made this paper happen. I think we can all agree that he deserves to be the first author.
* @noamraph, being the original developer who conceived the whole foundation of this project, should be an author as well (see [scikit-learn for a precedent](http://www.jmlr.org/papers/v12/pedregosa11a)). I would suggest him to be 2nd co-author. In addition, I suggest Noam could provide valuable insights, by writing about why and how he conceived tqdm (ie, his vision), and this IMO would be a great addition to the paper.
* as @hadim suggested in the other thread, I agree a history section would be very pertinent, and it would be an opportunity to acknowledge several developers. If not a history section, an addition to the introduction would also be fine, it depends on how we work it out. Please see the etherpad below for a rough first draft :-)

Questions that need to be resolved:
* Arfon, you said the paper will be reviewed through a "fast-track", so @gkthiruvathukal do you know if this means we should stick with minimal modifications (ie, we shouldn't rewrite significant portions of the paper), right? Would a new History section (along with Acknowledgements) be ok?
* How do we organize the attributions? There are three possible places:
    * In paper's authors list.
    * In paper's body in context.
    * In an Acknowledgements section.
* In relation to the previous point, we also need to assess the contributions to the code. There are several tools for that (next I'll experiment with them). @gkthiruvathukal is there a tool (apart from git fame) you would advise to explore in more fine grained details the git history? We should also review issues since some were not properly merged in.

For efficiency, I have created an etherpad (collaborative text editor) with a rough outline of my idea of the changes we could make, so that everyone here can contribute (if there is any vandalism, I can make it private and invite manually), and it's still of course meant to change depending what we agree on:

https://mypads.framapad.org/mypads/?/mypads/group/tqdm-i211b7ca/pad/view/tqdm-paper-x912b7p6
