### 2020-01-27 tqdm/tqdm#887 @lrq3000
(For those who received the email notification, I have change the URL for the pad to be in English by default, please prefer to use this URL: https://board.net/p/tqdm-paper )
