### 2020-01-27 tqdm/tqdm#887 @gkthiruvathukal
Responding to @lrq3000's questions:

Q: Arfon, you said the paper will be reviewed through a "fast-track", so @gkthiruvathukal do you know if this means we should stick with minimal modifications (ie, we shouldn't rewrite significant portions of the paper), right? Would a new History section (along with Acknowledgements) be ok?

A: Yes, the modifications would be minimal, and a complete rewrite should not be necessary.

Q:  How do we organize the attributions? There are three possible places: In paper's authors list, In paper's body in context, In an Acknowledgements section.

A: Speaking as an observer, I'd like to see *as many authors* as possible. I know tqdm is important, but let's remember that the paper with the record number of authors on LHC (Large Hadron Collider, a super important project, too) is more than 5,000 authors, https://www.nature.com/news/physics-paper-sets-record-with-more-than-5-000-authors-1.17567.  Did everyone contribute equally to this paper? Did everyone even write it? Or do the actual research? We'll never know, but they're all co-authors. We have a much smaller number of potential authors here? I'd like to see as many included as possible, provided their contributions can be justified in quantitative or qualitative terms.

Q: In relation to the previous point, we also need to assess the contributions to the code. There are several tools for that (next I'll experiment with them). @gkthiruvathukal is there a tool (apart from git fame) you would advise to explore in more fine grained details the git history? We should also review issues since some were not properly merged in.

A: I continue to research on this topic, but there are no really great tools, which is why I'm doing research on software metrics (hoping to change the state of affairs). One possible way is to look at other data that might show activity. You seem to have a fairly active issue tracker. Perhaps you can look at how many members of the team helped to raise and/or resolve issues. Did any contributors help, specifically, with user-facing aspects of the software (testing, docs, etc.?)

In the end, however, this can't be left entirely to data. It comes down to team dynamics and people advocating for others who may have made significant contributions. It seems like we already have convergence on some key contributors. The question is, are there any other minor contributors who are worthy of co-authorship? I would like those in a strong position to speak for those who may not be willing to advocate for themselves. Once I see all of the responses, I will synthesize a list of co-authors.  Again, I must stress that Arfon and I do not want to make the decisions for you. And similar to @lrq3000, I prefer to focus on research. Oddly, this project is helping to motivate the need for metrics, which is the subject of some of my current research. (Lest I digress.)

I won't speak to the "end game" because @arfon and the board are still discussing how to handle papers where the DOI and its metadata may need to be changed. My goal is to help mediate the (ethics) issues that were raised and help everyone to bring the matter to a graceful (and civilized) conclusion. There seems to be a lot of goodwill here, and I am pleased with the responses I have seen thus far. I'm optimistic we can come together to get the co-author list right!
