### 2020-01-27 tqdm/tqdm#887 @lrq3000
@gkthiruvathukal Thank you so much for your again very enlightening reply! And I hope you'll get all the funding needed to pursue your goal, it would really help to have better metrics indeed ;-)

I get what you said, although I have I think a fairly good memory of what happened and all contributors (I tend to have a good memory of those things, as I am always grateful to anyone who freely spend time to help others such as in opensource projects), I don't know what happened after, so I'm now going to try to explore the history with tools and manually to get a more comprehensive view and then I'll offer suggestions :-) (this may take some time, any help is welcome, if anyone has some contributors in mind please feel free to cite them!)
