### 2020-01-27 tqdm/tqdm#887 @gkthiruvathukal
Thanks for your support, @lrq3000. I had some funding for work on this topic, but it was only enough to prove some initial concepts. If interested to hear more, we can correspond sometime after getting tqdm's issues sorted out!

It would be great if we can get a HISTORY.md document produced. This would go a long way to clarifying how the project came to be--and how it evolved. Although I suggested Friday as a deadline for gathering initial input, that doesn't mean we can't wait, especially if it means we will include everyone who deserves to be included.
