### 2020-01-27 tqdm/tqdm#887 @lrq3000
@gkthiruvathukal Oh yes, I would be delighted to hear more!

Yes great idea for HISTORY.md, I'll direct my effort towards drafting that then :-)
