### 2020-01-27 tqdm/tqdm#887 @lrq3000
Actually, @casperdcl did a great job documenting changes and PRs in Releases, so maybe we can use that to build a HISTORY.md?
