### 2020-01-27 tqdm/tqdm#887 @lrq3000
Here's a first attempt [HISTORY.md](https://gist.github.com/lrq3000/c3cff5839d9c4a3831e0652f534ed430) generated by the wonderful [github-changelog-generator](https://github.com/github-changelog-generator/github-changelog-generator). It neatly shows the contributor for PR that were merged in. I think this should be usable as a basis to refine :-)
