### 2020-01-27 tqdm/tqdm#887 @noamraph
Hi everyone, apart from my wish to be listed as an author, I would accept any decision that you take. I have no problem with the list of authors being inclusive. I'm grateful for all the hard work that you spent making my little project into something so awesome!

Regarding history, I first wrote tqdm where I worked, at about 2006. Iterators where pretty new, and I really liked the possibility of showing a progress meter with only adding a few characters. I chose the name "tqdm" because I looked for a short and unique name, and I really like Arabic. (I'm a Jewish Israeli). "tqdm", which is Arabic "تَقَدُّم" in Latin letters, pronounced "taqaddum", simply means "progress".
