### 2020-01-27 tqdm/tqdm#887 @kmike
Hey! I think that any arrangement which includes @casperdcl, @noamraph and @lrq3000 as authors and mentions @hadim in some way is fine; their contributions were critical for having tqdm as it is now. It seems most proposals here suggest something similar anyways 👍
