### 2020-01-28 tqdm/tqdm#887 @gkthiruvathukal
Ok, I am synthesizing here.  As you read this, please keep in mind that I do not know any of you personally and am focusing on the input you've given and what git tells us.

I think there is strong consensus for @casperdcl, @lrq3000, @noamraph (probably in that order).  Based on my earlier report, I think there is a strong case for @altendky as the 4th author, based on my earlier analysis and his follow up (where he took great care not to say whether he should be a co-author). In his case, however, he seems willing not to be a co-author but has made non-trivial contributions to the success of the project. For this reason alone, I take the view that he should be a co-author. I'm aware that number of commits and lines of code are not the only measures, but in looking at the actual commits, I am convinced his contributions are significant.

As I mentioned, I am hoping to have any and all input by Friday of this week, after which I think @arfon can take the next editorial steps. Based on the information I have in git and on this issue thread, I feel comfortable moving toward these four as co-authors in the order mentioned but would certainly appreciate if all who have commented would react (an emoji will do) to this proposal.
