### 2020-01-28 tqdm/tqdm#887 @casperdcl
There are two very distinct issues: ongoing maintenance of the library
(paramount importance), and secondly the publication [joss.01277].

# 1) `tqdm` maintenance practices

- I absolutely consider **all** contributors to `tqdm` (any commits/lines of code/issues/PRs, whether surviving or not) to be "joint authors" of the *library*. More specifically; joint FOSS copyright holders. This is non-negotiable.

> [@altendky] I remembered some funny feelings with some of my contributions. I think the primary missing part [...]

This implies there are other more minor things you haven't mentioned. Please mention them! Also let me know if I haven't responded to any of your other points.

> [@altendky] two PRs showed as closed, not merged (#673, #622). The commits they were closed by do show me as author but for some reason have @casperdcl as committer.

Did you check "allow commits from maintainers" when opening your PR? If not, or if I felt there was a lot more work I had to do to tidy up, I would've rebased your commits. This is *extremely common good `git` practice*. You will show up as the author (both on the commits and in any metrics such as those generated by GitHub, and those generated by `git fame`). I want you to show up as the author! You obviously won't be the committer on such rebased/cherry-picked commits unless I do something really clandestine contrary to the design of `git`. Author takes precedence over committer. Please just take it to mean "the committer's machine was incidentally used at some point."

> [@altendky]  #674 (adding four words on one line...) was closed and a different solution committed without referencing me (no complaints here).

Yes, it looks like I accidentally independently implemented something which made your PR redundant. Based on the discussion it was also unclear whether it should have been required at all. In any case, at the time, I acknowledged all of this explicitly and also thanked you and added you to the list in the README.

> [@altendky] #598 (my only significant contribution) was merged per normal GitHub methods [...] the only comment I would have related to mine is that it would probably be good practice to use the GitHub merge button so it shows up as merged in GitHub

Sort of, I think because you kindly helped with rebasing, debugging and tidying everything so I didn't have to do much!
It wasn't "normal" in the sense that I never use the "merge" button as merging locally and pushing is much more explicit and thus reliable.
I try my best to ensure that whatever I do locally shows up correctly on GitHub but in general don't trust the merged-vs-closed/fixed-vs-closed status of PRs/issues, which is why I use a whole bunch of other metrics when acknowledging contributions.

> [@altendky] Should I be an 'author' [...] suspect that the sensible answer is to put me in the 'acknowledgements' group

You *are already* an author (of `tqdm`). See my first comment. Regarding [joss.01277], though, see the next section (basically happy for you to be acknowledged, and hopefully if JOSS updates their guidelines have you as author).

> [@hadim] I hope everyone will get their deserved acknowledgment/authorship without any more drama.

Referring solely to the library (incl. docs/website but not the paper, which I'll address separately below) please let me know if you feel you need more prominent acknowledgement. Let's face it, the library's dons have been viewed over 818 thousand times; several orders of magnitude more than any paper could hope for.

> [@gkthiruvathukal] You seem to have a fairly active issue tracker. Perhaps you can look at how many members of the team helped to raise and/or resolve issues.

As I mentioned above, I don't trust the "resolved" status. "Participated" might be a better thing to look for.

> [@gkthiruvathukal] sometimes an enormous amount of impact can be made without the singular focus on LOC (or commits). For example, there can be developers who write good tests, configuration scripts, and documentation [...] Did any contributors help, specifically, with user-facing aspects of the software (testing, docs, etc.?)

- "Sometimes" should be "most of the time."
- There is no singular focus on LoC, as evidenced by the various other metrics in https://github.com/tqdm/tqdm/#contributions. Let me know if you can think of any others.
- Tests, configuration scripts, and documentation are all included in my definition of LoC.
- The website (https://tqdm.github.io/) and associated helper scripts in their entirety were written by me but not included in the LoC though. I have no plans to count this. Let me know if you disagree.

> [@gkthiruvathukal] This project is lucky to have such a great group of contributors

- I find this a highly offensive offhand statement which sounds like a personal attack on all contributors. Luck had nothing to do with it - it took effort.

> [@gkthiruvathukal] understand the importance of your *software* being well-cited

It's not important for FOSS software to be well-cited compared to the important for it to be well-used.

> [@gkthiruvathukal] we don't want to make decisions for you

Yes, I understand you don't want to (and that's within your rights), but the consequences are MUCH worse.

> [@noamraph] Regarding history, I first wrote tqdm where I worked, at about 2006 [...] turned out to be very popular in my work place, so I decided to open source it.

I hope you had authorisation from your company https://www.theregister.co.uk/2019/12/12/nginx_moscow_office_raided/.

# 2) JOSS publication

Oh dear. This is exactly the sort of libellous argument *ad-hominem* which could have been avoided if JOSS could step in. I understand JOSS is volunteer-run, does great work, and is under no obligation to help, though it would be most useful if JOSS policies are revised to be more helpful.

I have mixed views about various posts in this issue. I am tempted to respond point-by-point, however I believe my views (and indeed anyone else's) are irrelevant to the underlying issue. Attempting to change people's views (in order to avoid the underlying issue) is not something I am comfortable with. Everyone is entitled to their beliefs. Also note that who meets the current and/or future JOSS definition of paper author will unlikely be the same as what anyone *feels* (i.e. "who do I wish were paper authors" versus "who actually are the paper authors").

To summarise the issue:

- there is a disagreement about the author list of [joss.01277];
- the contributing parties participating here all seem willing to accept JOSS decision(s)/overruling;
- JOSS will not overrule/decide;
- therefore JOSS will withdraw the paper.

There are some considerations which hopefully will make JOSS better:

1. Introduce explicit guidelines:
    + The [IEEE definition of authorship][ieee-auth] and [citation guidelines][ieee-cite] are a good place to start.
    + Modifications would of course be needed tailored to a FOSS-specific journal.
    + It is unreasonable to expect non-legally trained software devs to discuss this issue with no concrete guidance. It is also unethical to force volunteer FOSS devs to spend time debating this instead of just *telling them what to do* so they can move on to contributing more code and ideas.
    + Any numeric thresholds (e.g. min 5% LoC contribution), while bound to be wrong, are less wrong than not having thresholds at all (e.g. asking a teacher to mark and accepting marking errors will occur rather than asking students to rank themselves).
2. Resolve discrepancies between current published guidelines and unpublished internal guidelines/beliefs:
    + Published guidelines https://joss.readthedocs.io/en/latest/review_criteria.html#authorship:
      * > submitting author has made a 'substantial contribution' [...] determined by the commit history"
      * > the full list of paper authors seems appropriate and complete
      * does not state "*all* 'substantial' project contributors should be paper authors," which I think would be a harmful requirement but nevertheless something JOSS is entitled to add
    + Unofficial JOSS comment https://github.com/tqdm/tqdm/issues/887#issuecomment-578495632
      * > [@arfon] It it not the responsibility of the journal to decide who should be an author on a given paper
      * Conflicts with the review criteria. Criteria for deciding are indeed given (but evidently need to be improved upon).
      * > [@arfon] authors/contributors of the package are best-placed to decide between themselves
      * Once again seems to be confounding the issue by being ambiguous about whether we are talking about *author* (which I define to be of a paper) and *contributor* (which I define to be any activity on GitHub). Also stating that authors should decide who are authors is a tautological logical fallacy.
    + Unofficial JOSS comment https://github.com/tqdm/tqdm/issues/887#issuecomment-578523686
      * > [@gkthiruvathukal] not all contributions can be measured by LOC or number of commits
      * implies JOSS guidelines explicitly mentioning "commit history" need updating to include additional sources
    + Unofficial JOSS comment https://github.com/tqdm/tqdm/issues/887#issuecomment-578571433 and https://github.com/tqdm/tqdm/issues/887#issuecomment-578583505 and contributor comments https://github.com/tqdm/tqdm/issues/887#issuecomment-578558307 and https://github.com/tqdm/tqdm/issues/887#issuecomment-578579211
      * > [@gkthiruvathukal] probably deserve more than an acknowledgment for your contributions as significant contributors. I always take an *inclusive* approach to contributions, no matter how large or small. Hopefully, the lead contributors will chime in with whether we could live with the top contributors as minor co-authors
      * > [@gkthiruvathukal] I'd like to see *as many authors* as possible
      * > [@altendky] If I get co-author for my one bug-fix then we should look at everyone else that made non-trivial bug fixes and feature additions and include them as well.
      * > [@lrq3000] But we need to define a metric indeed. Personally, I would be very glad to see our fellow co-developers in the author list!
      * I fully agree - if paper authors are defined to be code authors then I insist on including the original Zenodo author list (https://doi.org/10.5281/zenodo.2792998). **I do not consider anything non-trivial**. Permission need not be sought from all code authors if we go down this route as we would be defining paper authors to be code authors, making the paper an accessory to the library rather than a stand-alone journal publication-with-associated-code. Please note that as-is, simply "liking" or "wanting" this to happen is incompatible with the current JOSS guidelines.
    + Contributor comment https://github.com/tqdm/tqdm/issues/887#issuecomment-578574650
      * > [@altendky] Is there something that restricts to co-authors and acknowledgments as opposed to adding a contributors group in between
      * Absolutely agree this is something JOSS should think about
    + Contributor comment https://github.com/tqdm/tqdm/issues/887#issuecomment-578579211
      * > [@lrq3000] contributed the most to this project, maintained it all this time making it a very successful package used in lots of other applications, and made this paper happen. I think we can all agree that he deserves to be the first author [of the paper]
      * I don't think any of these (apart from the "made this paper happen" and "contributed" -- it's irrelevant that it's "the most") are part of the current JOSS requirements.
3. Address the difference between a) project author(s), b) project maintainer(s), and c) paper author(s):
    + Definitions of all three are required (e.g. "(c) is expected to be the same as (b)," and "(b) is anyone who has ...,").
    + Zenodo author records need not match JOSS author records *unless JOSS updates its authorship definition to include all code contributors*. It is wrong to expect the other way around (Zenodo to update its authorship definition to include only a subset of contributors).
    + Consider this hypothetical extreme example
      * A FOSS project contributor responsible for writing a small yet "significant" (assuming a definition is provided and satisfied) amount of all code, documentation, unit tests, etc wishes to write a paper by themselves.
      * Due to the FOSS licence, this person is not required to seek any permission of anyone, nor are they forced to invite co-authorship (otherwise Python maintainers would be invited co-authors of every Python-based library!).
      * By the [IEEE definition of authorship][ieee-auth], this person will be the sole author of the paper.
      * By the [IEEE citation guidelines][ieee-cite] and by the FOSS licence terms, they will need to cite the project correctly.

[ieee-auth]: https://journals.ieeeauthorcenter.ieee.org/become-an-ieee-journal-author/publishing-ethics/definition-of-authorship/
[ieee-cite]: https://journals.ieeeauthorcenter.ieee.org/become-an-ieee-journal-author/publishing-ethics/cite-sources-appropriately/
[joss.01277]: https://doi.org/10.21105/joss.01277

## 2.1) Conclusions

- The JOSS **requirements** could be improved.
- The JOSS **suggestions** are trying to redefine and prune a project's full contributor/author/maintainer list to match the purposes of a paper rather than the other way around.

In the absence of any concrete guidance, I can only revert to my own personal code of ethics.

### 2.1.1) Suggestions

- Would @arfon be OK with opening an issue at JOSS discussing "**(2) JOSS publication**?" This is non-specific to https://github.com/openjournals/joss-reviews/issues/1277, and something JOSS could really benefit from.
- Based on my personal definitions of a) project authorship, b) project maintainer, and c) paper authorship, I suggest the either:
    + The paper author list is either left as-is. Based on https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain it seems an Acknowledgements section is indeed allowed, and I suggest one be added to the paper with an *inclusive* approach, even giving a history and praising noteworthy efforts if acceptable to editors and reviewers.
    + Or the paper author list is updated to match https://doi.org/10.5281/zenodo.2792998. Acknowledgements giving a history and praising noteworthy efforts if acceptable to editors and reviewers may be included.
- I suggest everyone take a look at https://github.com/tqdm/tqdm#contributions and the links therein. If any person or anything is missing please let me know ASAP.
- https://joss.theoj.org/about#ethics should be updated to replace "required to have [...] financial support" with "required to cite [...] financial support if any was received."

### 2.1.2) Actions

In the interim period while discussion/decision(s) are ongoing, I will do the following:

1. Remove https://github.com/tqdm/tqdm/blob/2e58f011c425291d1176ccdf1b2f1ae860ee2dc7/.zenodo.json#L11-L12
    + This was requested in https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-492158041 and made me extremely uncomfortable.
    + The Zenodo DOI is a software citation (not a paper citation) and as such should include their automatically-generated complete list of **all** contributors.
1. Update the table under https://github.com/tqdm/tqdm#contributions.
    + Only minor updates required.
    + This needs to be manually updated (last done in https://github.com/tqdm/tqdm/commit/36b3b7d34f5e50b9a13b73e34daefaf1b8318391 ~7 months ago).
    + Happy to take suggestions about how to automate the update process, improve, and/or remove the table.
1. Continue to maintain `tqdm` amongst other FOSS efforts.
