### 2020-01-29 tqdm/tqdm#887 @kmike
A good point about authors of paper vs authors of the software. I can see how the whole situation could have arised from different expectations here.

From https://joss.readthedocs.io/en/latest/submitting.html:

> Your paper should include:
>    A list of the authors **of the software** and their affiliations, using the correct format (see the example below).

In the example below a "correct format" is used to define authors with affiliations, which is then shows as authors of the paper. From this I conclude that in JOSS it is expected to have **authors of the software as authors of the paper**.

https://joss.readthedocs.io/en/latest/review_criteria.html#authorship asks for the list of authors to be reasonably complete. From this it follows that a list of paper authors should be a **reasonably complete** list of software authors.

Then they ask community to decide what would be a reasonable list of authors. Wording may not be the most precise, but I think then intention is clear. So let's do it :)

How to decide who's an author is subjective, and several approaches are possible. I don't think any legal stuff about authorship matters here. Rules are different in different countries, and we're not in the court. The whole thing is just about fairly acknowledging people for their contributions.

Picking some top contributors, considering their code, discussion and historical contributions looks like a fair way to define an author list. Considering everyone who contributed an author also sounds fine to me. There is also an Acknowledgement section, which can be used to highlight contributions of people who're not in the author list. Alternatively, I think it can be used to highlight top contributors, if "all contributors are authors" approach is taken.

It doesn't look that complicated, after all. Two concrete proposal on how to update the paper, which seems to be in line with the guidelines:

Proposal A.
1. Put @casperdcl, @lrq3000 and @noamraph as authors.
2. In the Acknowledgments section put a link to https://github.com/tqdm/tqdm/graphs/contributors and maybe mention that @hadim forked the repo in the dark times.

Proposal B.
1. Put all contributors to the author list. Getting affiliations can be tricky though.
2. In the Acknowledgments section say that @noamraph was the original author, @casperdcl has been the main maintainer and contributor in the recent years, @lrq3000 was another prolific contributor, and @hadim forked the repo in the dark times.
