### 2020-01-29 tqdm/tqdm#887 @lrq3000
@casperdcl , you are shooting yourself in the foot. **This situation is your fault, not JOSS's.**

This situation arose because you wrote the paper on your own without reaching to any other co-developers, with the exception of me when a JOSS editor requested it. This was evidently an unreasonable procedure.

As you have worked in research, you know that the authors of a paper must include all those who contributed to the experiment. It would be foolish for any of the team member to quickly write a paper secretively, submit it as the sole author, and then claim they are right to do so because noone else contributed to the paper, that's an obvious circular reasoning.

If you had reached to other developers to invite in the writing, and noone would have answered, then it would make for a far better case for you to be the sole paper author. But pre-emptively avoiding contacts barred any other contribution.

So no need for legalese, just common ethical sense. I suggest we stop on this slippery path and come back to fixing the issue at hand. Although @gkthiruvathukal makes suggestions, which I am thankful for, we ultimately are the ones to decide, which is only logical (they only arbitrate).

I agree the [Zenodo](https://zenodo.org/record/2792998) list of authors can be a good basis, but maybe complemented? Also the order seems subjective, I would like to ensure some degree of objective fairness (if that is at all possible)...

Side-note, about the legal status, if it comes to the worst, the whole project can probably be easily rebased on my minimalist implementation which I made from scratch. I don't know if it still lives here (in the PRs?) but we can certainly find it somewhere.
