### 2020-01-29 tqdm/tqdm#887 @altendky
> This implies there are other more minor things you haven't mentioned. Please mention them! Also let me know if I haven't responded to any of your other points.

Even if I did have anything in mind, I'll pass.  I stated that you handled my contributions appropriately and what did you do?  Rebutted that statement point by point.  *boggle*  I also commented that while I understand personal preference in workflow that your preference has the presumably unintended side effect of marking PRs as closed rather than merged which can be a bit confusing.  Instead of understanding that point you...  again, rebutted it.

Please consider not responding to this at all and instead just thinking about the slight feedback on workflow.

> > [@gkthiruvathukal] This project is lucky to have such a great group of contributors
>
> I find this a highly offensive offhand statement which sounds like a personal attack on all contributors. Luck had nothing to do with it - it took effort.

Clearly that is not an insulting statement.  At the same time, there was certainly chance involved in venvs (was it mkenv at the time?) deciding to use tqdm before py2 went eol and capturing the output in the tests thus showing the bug and me being involved enough at that time that I decided to try to fix the issue rather than just dropping use of tqdm, etc.

So, @casperdcl, is the summary of all that you wrote here that you are not clear as to what 'authors' are being discussed and whether they are supposed to be 'the people that wrote the paper' vs. 'the people that wrote the library'?  That seems like something that @gkthiruvathukal could probably clear up if it's still in question.  Though I haven't been involved in research so I may really be missing something.
