### 2020-01-29 tqdm/tqdm#887 @lrq3000
> [@gkthiruvathukal] This project is lucky to have such a great group of contributors

I am currently reviewing carefully the commits, and I am in awe with what various developers contributed in their free time. They solved lots of small and hard problems I had no idea how to fix at the time. Yes, the success of tqdm is due to a lot of work, particularly by Casper, but it's also lucky that so many found this library useful and interesting enough to dedicate some of their free time to help out.

> So, @casperdcl, is the summary of all that you wrote here that you are not clear as to what 'authors' are being discussed and whether they are supposed to be 'the people that wrote the paper' vs. 'the people that wrote the library'?

The thing is that the correct process would have been to contact other developers (software authors) beforehand to invite them in authoring/writing the paper. Since that was not done and the paper was published, and now on hold, we are in a special case. So now, there is no way to go back and define who should be the paper's author on the criterion on who wrote the paper, and we have to select another criterion. The criterion is ours to choose, JOSS can't choose for us since it's a special case. Or there is always the other possibility of simply retracting the paper if we can't agree, but I think this would be a waste and a loss-loss situation.

We already have several points of agreements. Let's not overcomplicate this, I am convinced we can all reach agreement on a final solution that would content everyone.
