### 2020-01-29 tqdm/tqdm#887 @gkthiruvathukal
Alright, thanks to everyone for your participation. Looking back, I probably shouldn't have chosen the word *lucky* to describe how your team has such a *great* group of developers, despite nothing but the best of intentions.  The term *luck* was not by any means intended to slight your efforts but rather the opposite, which many of you seem to have understood.  I was doing what many world cultures do by recognizing the importance of being *lucky* as part of *success*. We have a phrase in USA, attributable to a famous baseball player (look it up) that says, "I'd rather be lucky than good." (He was onto something.)

Of course, many people mistakenly think that all of the efforts are attributable to their own individual talents, which seems to be at work here. You could also have been *unlucky* and not had such a great group of people to help evolve the software, but you weren't. It's unlikely that tqdm would be what it is today, if left to the efforts of one person, given that it started with someone who is not the largest committer. Many have contributed to its success, which is important when it comes to the paper/publishing aspect of all this.

For every successful project like yours, there are hundreds that are not. Most projects--in the end--fail for human (not technical) reasons. I hope yours continues to be successful. Resolving the human issues will probably be crucial to ongoing interest in the project.

**I'm going to conclude my participation and turn things back over to Arfon.** We are sure *lucky* to have him as our EIC. In the end, I came here to help mediate the situation, and I think I was at least successful in opening the lines of communication and reaching some degree of consensus how to proceed. In psychology, one thing we learn is that getting people to talk is usually the hardest part.

Here are some resources that I hope will be helpful to your project's ongoing success:

On the role of luck, which most underestimate and also fail to acknowledge:
https://www.nytimes.com/2017/01/09/your-money/stop-and-acknowledge-how-much-luck-has-to-do-with-your-success.html

A guide for working well with others on open source projects, written by folks who have a lot of experience working on open source projects (developers of Subversion):
http://shop.oreilly.com/product/0636920018025.do

Wishing all of you the best!
