### 2020-01-29 tqdm/tqdm#887 @lrq3000
Thank you again @gkthiruvathukal for your insights and resources, I will surely read them (particularly the book once I buy it!).

I have finished my review of the commits. Before writing the results, here is the methodology I have used:

1. Overview of authors ordered by contribution count: `git-fame` (great package by @casperdcl) and `git shortlog -ns` to get a first glance of top authors by loc and by commits count. "Contribution" here is to be understood vaguely, can be loc or commits count.
2. Overview of commits grouped by author: `git shortlog -n` to review the description of the commits by each author
3. Manual deep review: reviewing manually each commits by each author via the github interface. I also identified via github the top authors by number of commits and by number of additions (results slightly differ).
4. Finding missing authors: in the [log](https://gist.github.com/lrq3000/c3cff5839d9c4a3831e0652f534ed430#file-log) of the process of generating the CHANGELOG.MD by github-changelog-generator, there are warnings such as:
`Warning: PR 671 merge commit was not found in the release branch or tagged git history and no rebased SHA comment was found`

I have checked each PR from these warnings and found a few orphaned authors.
5. Quick checkup of the most commented issues to see if there were "community managers" or another kind of user who significantly contributed apart from code.

Here are the results: apart from the 3 authors mentioned previously (I let you guys decide the order), here are the other software authors that I think should be considered for paper authorship, along with a succinct description of their contribution (feel free to correct me if I missed anything), and in order of significance and effort (so an aggregate subjectively evaluated measure of the commits count without reverts and the content of these commits and how much it made the project evolve) according to me:

```
* hadim made the initial tqdm logo and initially coordinated the effort towards the new tqdm and merged (casper later refined the logo and animated it) see https://github.com/noamraph/tqdm/issues/18 and https://github.com/tqdm/tqdm/issues/2 and https://github.com/tqdm/tqdm/issues/30 and https://github.com/tqdm/tqdm/issues/33
* Gangshuo Chen major fixes for notebook and pandas + TMonitor minor
* Kyle Altendorf significant changes to output management - note there were lots of reverts, so the commits count does not directly reflect the changes done, but this suggests it was a quite hard feature to implement
* kmike mid, tox setup and various packaging fixes and added the disable keyword, very * important
* obiwanus initial testing and tox setup
* mbargull significantly fixed TMonitor
* mjstevens777 idem
* orivej idem
* martinzugnoni provided an interactive demo ipython notebook
* kcwu significant fix for precision of remaining time calculation (ema)
* charlienewey minor changes to core tqdm, on deduplicating ema calculation
* MPagel significant few changes to tqdm core printing and code organization
* jcea initial arg to restart meter
* yarikoptic lazy loading of modules
* RunOrVeith inf total iterables support
* JackMc support for any file-like output
* tacaswell stack level change of warnings
* CrazyPython community management and copyediting
```

All the authors above made usually more than 5 commits, some have made only 2 or 3 commits but were significant enough to suggest authorship IMO (example: martinzugnoni, 3 commits  1,656 ++  476 --, because he added a full ipython notebook demo he made from scratch).

I am not sure whether all these software authors should be in the paper's author list, but I would suggest to include at least down to martinzugnoni (included).

All other software authors that are not listed here, who did minor - but very useful! - changes (which is the primary criterion I have used to differenciate, and who happen to have a number of commits of 3 at most), should I think be thanked in an Acknowledgement sections.

Orphaned authors who should be in the acknowledgements too:
cfperez PR#220
jwilk PR#82
enriquefynn PR#76

Also, I think it's important that Noam is credited as the original conceptor of tqdm in the paper.

Finally, here is a Gource visualization of tqdm development history, a nice overview of all the hard work everyone put in:

https://youtu.be/uiOFEbdC1uo

Side-note: please @casperdcl [fix this](https://github.com/tqdm/tqdm/wiki/History).
