### 2020-01-29 tqdm/tqdm#887 @lrq3000
Also please let me commend @casperdcl work in improving recognition of contributions, both in the [readme](https://github.com/tqdm/tqdm#contributions), by making the [git-fame package](https://github.com/casperdcl/git-fame), and also by adding [badges linking to the contributors](https://github.com/casperdcl/git-fame#authors).

To further improve recognition of contributions, in line with gkthiruvathukal and with minimal overhead in development, I would suggest to systematize the use of [git-changelog-generator](https://github.com/github-changelog-generator/github-changelog-generator) to make and update a CHANGELOG.md file. This will serves both as a change log and an attributon list, since it also includes the authors of each merge.
