### 2020-01-29 tqdm/tqdm#887 @HsuanTingLu
>
> Finally, here is a Gource visualization of tqdm development history, a nice overview of all the hard work everyone put in:
>
> https://youtu.be/uiOFEbdC1uo
>

@lrq3000 I think you accidentally made the video private
