### 2020-01-29 tqdm/tqdm#887 @lrq3000
@HsuanTingLu Sorry in fact it was still uploading, now it's done, you can watch it :-)
