### 2020-02-02 tqdm/tqdm#887 @lrq3000
Let's try to wrap this up. As Arfon suggested, and to simplify choice as well as to ensure some form of balance, I suggest me, @casperdcl and @noamraph try to agree on a list of co-authors for the paper.

Here are the 3 options I suggest, note that contributors who are not selected as co-authors will anyway be mentioned in Acknowledgements:
1. Only @casperdcl , me and @noamraph, this is the minimal common denominator most (all?) participants here agreed on.
2. In addition to authors in 1, add the following who made significant contributions (significant being defined as > 5 commits or major changes/bugfixes to tqdm, such as adding a full ipython notebook demo or fixing the TMonitor), this would include the following contributors:

* hadim
* Gangshuo Chen
* Kyle Altendorf
* kmike
* obiwanus
* mbargull
* mjstevens777
* martinzugnoni

Note I have removed orivej, who made minor fixes to TMonitor after I had a 2nd look.

3. All contributors to the tqdm software will be co-authors.

----

I have a preference for 2, but I am also ok with 1 and 3 without any preference between the two.

@casperdcl, @noamraph: what do you guys think? :-)

PS: @casperdcl the order and values in [this commit](https://github.com/tqdm/tqdm/commit/7580b90e718128222d881deecf88c62daf8de145) look a bit weird to me given what I saw during my code review of contributions, so I'm wondering if maybe LoC in this table is in fact the SLoC (surviving LoC)?
