### 2020-02-02 tqdm/tqdm#887 @noamraph
Hi, I'm fine with all 3 options.
