### 2020-02-05 tqdm/tqdm#887 @lrq3000
Thank you @noamraph !

@casperdcl what do you think?
