+++ README.rst
@@ -688,7 +688,7 @@ Nested progress bars
 
     for i in trange(4, desc='1st loop'):
         for j in trange(5, desc='2nd loop'):
-            for k in trange(50, desc='3nd loop', leave=False):
+            for k in trange(50, desc='3rd loop', leave=False):
                 sleep(0.01)
 
 On Windows `colorama <https://github.com/tartley/colorama>`__ will be used if
