### 2020-02-10 tqdm/tqdm#887 @lrq3000
Hello @casperdcl , did you have some time to take a look at my propositions? Please just let us know your preferences and I'll take care of the practical stuff.

We are about halfway through the deadline Arfon set for us (1 month), I would regret if we couldn't reach an agreement on time.
