### 2020-02-13 #1277 @casperdcl
Hello everyone. Thank you for your patience!

After much discussion I think we're all confident that we can arrive at a conclusion.

However before I can contribute I need to understand JOSS principles/ethos so that I can then follow them. I need an answer to this hypothetical question:

> A FOSS project *Proj* exists. *SC-A* is a significant contributor to *Proj*. *SC-A* invites all other significant contributors of *Proj* to collaborate on a JOSS paper *Pap*. Everything goes smoothly apart from *SC-B*. Everyone agrees *SC-B* is a significant contributor to *Proj*. However *SC-B* states that they have no time to write or review *Pap*, but still would like to be listed as an author of *Pap*. All significant contributors of *Proj* non-negotiably refuse to volunteer any opinions on the matter, and insist that JOSS decide.

**Question**: should *SC-B* be listed as an author?

1. JOSS official answer: :tada: **Yes**
2. JOSS official answer: :rocket: **No**
3. JOSS official answer: :confused: `None`/undecided
    - In this case JOSS must reject the paper as everyone else has refused to decide.

Please note I will personally fully support JOSS without argument/debate no matter what answer is picked - it is fully within its rights to pick any. It's also a hypothetical question so any answer is not taken to be a decision on this issue.

Feel free to vote using emoji
