### 2020-02-13 tqdm/tqdm#887 @casperdcl
# 1) `tqdm` maintenance practices

As I have always stated, the project maintenance is my primary concern, including actively (aggressively? :)) fostering positive feelings by mentioning as many contributors as possible in the docs and websites.

@lrq3000 from https://github.com/tqdm/tqdm/issues/887#issuecomment-579637044:

- the list of mentions
> `* hadim made the initial [...]`

is great; it would be nice if you can post a script to help with auto-generating this.

- need to weigh up `github-changelog-generator` versus the current https://tqdm.github.io/releases/ (mirror https://github.com/tqdm/tqdm/wiki/Releases) and ideally combine best of both

from https://github.com/tqdm/tqdm/issues/887#issuecomment-581128956:

> the order and values in [this commit](https://github.com/tqdm/tqdm/commit/7580b90e718128222d881deecf88c62daf8de145) look a bit weird to me given what I saw during my code review of contributions, so I'm wondering if maybe LoC in this table is in fact the SLoC (surviving LoC)?

The full explanation of how the table is generated is given right above the table (`git fame -wMC` as of writing).

# 2) JOSS publication

Hoping to get feedback. Will propose a solution soon.
