### 2020-02-13 #1277 @danielskatz
If all other authors are ok with SC-B being listed as an author without writing or reviewing the paper, and SC-B is willing to have their name on something they didn't review, then from the point of view of JOSS, I think the answer is that SC-B should be listed as an author, as that would be a conclusion that would satisfy all authors ad give credit to all authors. (With the caveat that I can't advise that SC-B should be willing to have their name on something they didn't review.)

If other authors disagree, then I don't think JOSS can force agreement.
