### 2020-02-16 53a70ef
clarify etymology
- fixes #895
