### 2020-02-16 e5e7f8f
abstract tests, enable full coverage
