+++ tqdm/tests/tests_tqdm.py
@@ -1890,11 +1890,3 @@ def test_float_progress():
                         assert not w
                 assert w
                 assert "clamping frac" in str(w[-1].message)
-
-
-@with_setup(pretest, posttest)
-def test_notebook_disabled_description():
-    """Test that set_description works for disabled tqdm_notebook"""
-    from tqdm import tqdm_notebook
-    with tqdm_notebook(1, disable=True) as t:
-        t.set_description("description")
