### 2020-02-19 tqdm/tqdm#887 @casperdcl
Thanks all for the discussion.

Based on:

- JOSS' last clarification regarding its definition of authorship (https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-585848767);
- the apparent JOSS policy to have Zenodo meta data match the paper (https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-492158041);
- the fact that Zenodo's default is to objectively auto-generate author lists based on GitHub metadata (I have an email from them from last year stating this), and
- everyone's comments in this issue;

I think there is only one solution: think of the "paper authors list" as a "software contributor attributions list."

## Proposal

- Update paper author list to match either the Zenodo contributor list at the time of acceptance (https://doi.org/10.5281/zenodo.2792998) or the time of revision (currently https://doi.org/10.5281/zenodo.595120)
- Correct names (using e.g. [`.mailmap`](https://github.com/tqdm/tqdm/blob/master/.mailmap)) if necessary

## Notes

- I think affiliations don't apply as there was never any formal institution/org-related collaboration on this completely public-driven FOSS project. Formal authorisation and support was never sought from nor given by any organisation.
- I think seeking consent from "authors" is also not required as
  + nobody's had any issue with their name appearing on the Zenodo author list
  + nobody's had any issue with their name appearing on the GitHub contributor list
  + nobody's had any issue with their name appearing in the README
  + the "paper author list" is really a "software contributor attribution list"
    * software attribution consent is granted by the licence
