+++ README.rst
@@ -1123,14 +1123,14 @@ Name                 ID                                                       SL
 ==================== ======================================================== ==== ================================
 Casper da Costa-Luis `casperdcl <https://github.com/casperdcl>`__             ~75% primary maintainer |Gift-Casper|
 Stephen Larroque     `lrq3000 <https://github.com/lrq3000>`__                 ~15% team member
-Martin Zugnoni       ` <https://github.com/>`__                               ~3%
+Martin Zugnoni       `martinzugnoni <https://github.com/martinzugnoni>`__     ~3%
 Guangshuo Chen       `chengs <https://github.com/chengs>`__                   ~1%
 Hadrien Mary         `hadim <https://github.com/hadim>`__                     ~1%  team member
 Matthew Stevens      `mjstevens777 <https://github.com/mjstevens777>`__       ~1%
 Noam Yorav-Raphael   `noamraph <https://github.com/noamraph>`__               ~1%  original author
 Kyle Altendorf       `altendky <https://github.com/altendky>`__               ~1%
-Ivan Ivanov          ` <https://github.com/>`__                               ~1%
-James E. King III    ` <https://github.com/>`__                               ~1%
+Ivan Ivanov          `obiwanus <https://github.com/obiwanus>`__               ~1%
+James E. King III    `jeking3 <https://github.com/jeking3>`__                 ~1%
 Mikhail Korobov      `kmike <https://github.com/kmike>`__                     ~1%  team member
 ==================== ======================================================== ==== ================================
 
