### 2020-02-19 tqdm/tqdm#887 @lrq3000
Thank you @casperdcl for your in-depth look into this issue. I agree with your arguments and your proposition sounds perfect to me.

About whether to use the list at the first submission time, or at revision time (now), since JOSS will publish the new manuscript with an updated time, I think the safest is to use the current revision (and I think this has the advantage to include some late but significant contributors if I remember from the Gource video?).

The only practical issue may be that for some users, we may need to ask them their full name as they did not necessarily set it in their Github profile.
