### 2020-02-20 tqdm/tqdm#887 @arfon
> The only practical issue may be that for some users, we may need to ask them their full name as they did not necessarily set it in their Github profile.

We can also publish papers with just GitHub handles (see the authors at the end of this recent paper https://joss.theoj.org/papers/10.21105/joss.01832). It's not ideal, but it's possible.
