### 2020-02-23 tqdm/tqdm#887 @lrq3000
Dear @mjstevens777, @CrazyPython, @toddrme2178, @aplavin, @Socialery, @RedBug312, @FichteFoll, @zed, @songololo, @littlezz, @immerrr, @igorljubuncic, @elsonidoq, @deeenes, @darindf,

We would like to invite you to be co-authors of a paper about the tqdm library, to which you contributed in the software's code development. In order for us to properly credit you, could you please reply here with your full name?

Thank you very much in advance!
Best regards,
Stephen Karl Larroque
