### 2020-02-23 tqdm/tqdm#887 @songololo
@lrq3000 my contribution (4 lines) is quite minor so no need to add me, but thank you for checking.
