### 2020-02-23 tqdm/tqdm#887 @casperdcl
Need to clarify here... we're going to add everyone unless they explicitly ask for their name to be removed. "No need to add me" is a bit ambiguous. Please say "Please don't add me." if you want.

Also feel free to post ORCIDs if you have one.
