### 2020-02-23 tqdm/tqdm#887 @mjstevens777
Please don’t add me. I appreciate the gesture though.
