### 2020-02-23 tqdm/tqdm#887 @deeenes
Thanks for asking, please don't add me, I've done very little contribution
