### 2020-02-24 tqdm/tqdm#887 @FichteFoll
Thanks for asking. If Github handles are fine, then please do add me. Otherwise, please don't.
