### 2020-02-24 tqdm/tqdm#887 @immerrr
Thank you for asking, please, don't add me, my contribution was quite trivial.
