### 2020-02-24 tqdm/tqdm#887 @igorljubuncic
Hello,

Thank you for asking.

Our contribution was small, but if you want, you can credit: Snap Advocacy
team.

Regards,
Igor

On Mon, Feb 24, 2020 at 11:05 AM immerrr again <notifications@github.com>
wrote:

> Thank you for asking, please, don't add me, my contribution was quite
> trivial.
>
