### 2020-02-24 tqdm/tqdm#887 @elsonidoq
Same here, my contribution was just a tiny fix, don't add me to the paper.
Thanks!
