### 2020-03-04 33d5249
update paper auth list
- fixes #887
