### 2020-03-04 #1277 @casperdcl
PR here: https://github.com/tqdm/tqdm/pull/905
