### 2020-03-04 tqdm/tqdm#887 @lrq3000
Dear @altendky, @hadim, @chengs, @obiwanus, @kmike, @mbargull, @mihaild, @martinzugnoni, @charlienewey, @MPagel, @staffanm, @orivej, @maxnordlund, @hugovk, @danielpanteleit, @cancan101, @JackMC, @Berserker66, @yarikoptic, @wwwiiilll, @volans-, @uadnan, @to266, @tiagocoutinho, @tacaswell, @sizmailov, @shirish93, @robertzk, @rlmaers, @pjvandehaar, @minho42, @mikekutzma, @kcwu, @julienc91, @jnhansen, @jcea, @gsganden, @geekypandey, @fordhurley, @f0k, @dyno, @dwhswenson, @davidbau, @danking, @cheng10, @carlinmack, @arunpersaud, @arkottke, @anntzer, @andportnoy, @Separius, @RunOrVeith, @JoshKarpel, @JaviMerino, @EdwardBetts, @AbysmalBiscuit,

We would like to kindly invite you to be co-authors of a paper about the tqdm library, to which you contributed in the software's code development. If you agree to be part of the co-authors, could you please reply with a "yes, I agree to be a co-author" in the next 48h (before Saturday 7th March)? If you don't want to, you can either not reply or reply that you wish not be a co-author.

Thank you very much in advance!
Best regards,
Stephen Karl Larroque

PS: @casperdcl, @lrq3000, @noamraph, and other previously pinged users do not need to reply since they already stated whether they wished or not to be co-authors.
