### 2020-03-04 tqdm/tqdm#887 @arkottke
yes, I agree to be a co-author.

- Albert

On Wed, Mar 4, 2020, 12:39 PM Stephen Karl Larroque <
notifications@github.com> wrote:

> Dear @altendky <https://github.com/altendky>, @hadim
> <https://github.com/hadim>, @chengs <https://github.com/chengs>, @obiwanus
> <https://github.com/obiwanus>, @kmike <https://github.com/kmike>,
> @mbargull <https://github.com/mbargull>, @mihaild
> <https://github.com/mihaild>, @martinzugnoni
> <https://github.com/martinzugnoni>, @charlienewey
> <https://github.com/charlienewey>, @MPagel <https://github.com/MPagel>,
> @staffanm <https://github.com/staffanm>, @orivej
> <https://github.com/orivej>, @maxnordlund <https://github.com/maxnordlund>,
> @hugovk <https://github.com/hugovk>, @danielpanteleit
> <https://github.com/danielpanteleit>, @cancan101
> <https://github.com/cancan101>, @JackMc <https://github.com/JackMc>,
> @Berserker66 <https://github.com/Berserker66>, @yarikoptic
> <https://github.com/yarikoptic>, @wwwiiilll <https://github.com/wwwiiilll>,
> @volans- <https://github.com/volans->, @uadnan <https://github.com/uadnan>,
> @to266 <https://github.com/to266>, @tiagocoutinho
> <https://github.com/tiagocoutinho>, @tacaswell
> <https://github.com/tacaswell>, @sizmailov <https://github.com/sizmailov>,
> @shirish93 <https://github.com/shirish93>, @robertzk
> <https://github.com/robertzk>, @rlmaers <https://github.com/rlmaers>,
> @pjvandehaar <https://github.com/pjvandehaar>, @minho42
> <https://github.com/minho42>, @mikekutzma <https://github.com/mikekutzma>,
> @kcwu <https://github.com/kcwu>, @julienc91 <https://github.com/julienc91>,
> @jnhansen <https://github.com/jnhansen>, @jcea <https://github.com/jcea>,
> @gsganden <https://github.com/gsganden>, @geekypandey
> <https://github.com/geekypandey>, @fordhurley
> <https://github.com/fordhurley>, @f0k <https://github.com/f0k>, @dyno
> <https://github.com/dyno>, @dwhswenson <https://github.com/dwhswenson>,
> @davidbau <https://github.com/davidbau>, @danking
> <https://github.com/danking>, @cheng10 <https://github.com/cheng10>,
> @carlinmack <https://github.com/carlinmack>, @arunpersaud
> <https://github.com/arunpersaud>, @arkottke <https://github.com/arkottke>,
> @anntzer <https://github.com/anntzer>, @andportnoy
> <https://github.com/andportnoy>, @Separius, @RunOrVeith, @JoshKarpel,
> @JaviMerino, @EdwardBetts, @AbysmalBiscuit,
>
> We would like to kindly invite you to be co-authors of a paper about the
> tqdm library, to which you contributed in the software's code development.
> If you agree to be part of the co-authors, could you please reply with a
> "yes, I agree to be a co-author" in the next 48h (before Saturday 7th
> March)? If you don't want to, you can either not reply or reply that you
> wish not be a co-author.
>
> Thank you very much in advance!
> Best regards,
> Stephen Karl Larroque
>
> PS: @casperdcl, @lrq3000, @noamraph, and other previously pinged users do
> not need to reply since they already stated whether they wished or not to
> be co-authors.
>
