### 2020-03-04 tqdm/tqdm#887 @hadim
yes, I agree to be a co-author.
