### 2020-03-04 tqdm/tqdm#887 @andportnoy
Thank you for asking, please don't add me.

Initially I was inclined to say yes (out of vanity), but then I looked up my contribution and was really ashamed of myself.

Here it is for reference :)

![tqdm-contribution](https://user-images.githubusercontent.com/9360340/75921770-8f054480-5e16-11ea-839c-34d2ef067392.png)
