### 2020-03-04 tqdm/tqdm#887 @MPagel
Yes, I would like to be included as a co-author, provided I can at least see the paper before submission.

Alternatively, you may add my name to the acknowledgements section.

For relevant affiliations (if author), see my signature.

================
Matthew D. Pagel
Database Administrator: UCDavis Biotelemetry Lab
Data coordinator: multi-agency Enhanced Acoustic Telemetry group

1331 Academic Surge Building
Department of Wildlife, Fish and Conservation Biology (WFCB)
University of California, Davis
One Shields Ave
Davis, CA 95616

Office Ph: 530 752-5372
    Lab Ph: 530 752-1156
