### 2020-03-04 tqdm/tqdm#887 @cancan101
yes, I agree to be a co-author
