### 2020-03-04 tqdm/tqdm#887 @Berserker66
yes, I agree to be a co-author

if you want to list me. My contribution is a small fix for IronPython. I don't mind mind either way.
