### 2020-03-04 tqdm/tqdm#887 @jnhansen
yes, I agree to be a co-author
