### 2020-03-04 tqdm/tqdm#887 @mihaild
yes, I agree to be a co-author
if you find that fixing 1 bug with few lines of code is enough contribution for this.
