### 2020-03-04 tqdm/tqdm#887 @mbargull
Hi, thanks for asking!
Please don't add me as a co-author.
My only contributions are in https://github.com/tqdm/tqdm/pull/527/commits which was mostly just a trivial `time.sleep` -> `threading.Event.wait` change, i.e., a non-significant contribution/fix.
(Plus, as a co-author I would've wanted to review the work prior to submission, which I don't have the capacity for, currently.)
If you want to reference me in an acknowledgement section, I'm fine that (but I also don't demand it).
