### 2020-03-04 tqdm/tqdm#887 @julienc91
Thanks for asking, but my contribution is not worth it.
