### 2020-03-04 tqdm/tqdm#887 @altendky
yes, I agree to be a co-author
