### 2020-03-04 tqdm/tqdm#887 @hugovk
Yes, I agree to be a co-author, ORCID: 0000-0001-5715-8632. Thanks!
