### 2020-03-04 tqdm/tqdm#887 @hugovk
ping @Separius, @RunOrVeith, @JoshKarpel, @JaviMerino, @EdwardBetts, @AbysmalBiscuit:

Please see https://github.com/tqdm/tqdm/issues/887#issuecomment-594823015, looks like GitHub hit a limit with the long list, so you may not have got the mention. Cheers!
