### 2020-03-04 tqdm/tqdm#887 @yarikoptic
thank you for considering me!  But with occasional whining and a single commit (c49021da401ac5b68892df9b71d92dab0e85167a) I do not consider myself yet deserving the honor!  Thank you again and good luck!
