### 2020-03-04 tqdm/tqdm#887 @uadnan
Thanks for considering me. And yes, I agree to be a co-author.

- Adnan Umer
