### 2020-03-04 tqdm/tqdm#887 @EdwardBetts
Happy to be a co-author.
