### 2020-03-05 tqdm/tqdm#887 @anntzer
Thanks for the invite, but my contribution (so far) does not warrant authorship; please remove me from the list.
