### 2020-03-05 tqdm/tqdm#887 @MPagel
>     * I think affiliations don't apply as there was never any formal institution/org-related collaboration on this completely public-driven FOSS project. Formal authorisation and support was never sought from nor given by any organisation.

In my case, IP rules of my university that I probably signed when I was hired state something to the effect of: if any university resources were used, the contribution I make is not my own.  Even if I didn't explicitly sign a document to that effect, these rules are nonetheless in place and govern my ability to maintain employment - i.e. I can be fired if they "find out" that I'm using their resources but going around their back and claiming "ownership" over something.  Ostensibly, these rules even apply to the use of e-mail from off-campus outside of normal working hours for non-university related things, but practically speaking they probably only enforceable when I'm sitting at my desk or otherwise "clocked in."

In any case, I used tqdm as part of a script that I was writing at work, for work.  The contributions I made stemmed from changes I made to do my job.  Ergo, the university owns it.  Ergo, my affiliation applies.

This all being said, I'll be leaving the university next month and moving across the country, but there's no need to burn bridges and claim that the little bit of code I contributed is mine.
