### 2020-03-05 tqdm/tqdm#887 @cheng10
yes, I agree to be a co-author
