### 2020-03-05 tqdm/tqdm#887 @kcwu
yes, I agree to be a co-author.

Just in case if affiliations or email is needed, please use
Google Inc.
kcwu@google.com
