### 2020-03-05 tqdm/tqdm#887 @Separius
yes, I agree to be a co-author
