+++ .meta/.snapcraft.yml
@@ -3,7 +3,7 @@ summary: A fast, extensible CLI progress bar
 description: |
  https://tqdm.github.io
 
- `tqdm` means "progress" in Arabic (taqadum, تقدّم) and is an
+ `tqdm` means "progress" in Arabic (taqaddum, تقدّم) and is an
  abbreviation for "I love you so much" in Spanish (te quiero demasiado).
 
  Instantly make your loops show a smart progress meter and stats - just
