### 2020-03-05 tqdm/tqdm#887 @JaviMerino
Thank you for considering me :+1: . My contribution was very small. In my opinion I should not be included as co-author.
