### 2020-03-05 tqdm/tqdm#887 @to266
Thank you for considering, but my contribution does not warrant authorship - please remove me from the list.
