### 2020-03-05 tqdm/tqdm#887 @casperdcl
@MPagel

1. I've already added your affiliation, as requested
2. My comment earlier was more about not insisting on (rather than refusing) affiliations

However in your case legally the situation could be potentially far more complex. Given it's a minor contribution it's unlikely that any of this applies, but strictly speaking:

1. Being inspired in the workplace to develop something in your own time would fall into a grey area where neither party could straightforwardly assert automatic copyright.
2. In this instance an employer has not asserted copyright over your work. They've (probably) asserted the *right* to assert copyright. That is, on trust, they assume your work aligns with their goals and ergo may want the affiliation. On the other hand, they may decide (upon "finding out" about your contributions) that the work goes against their principles etc. and that they would want their name and yours removed and, in the extreme case, fire you and/or seek compensation for wasted time because you never sought their permission. In a third scenario, they could decide that your work deserved payment (i.e. you spent company time working on another org's project as a hired contractor), in which case they would seek payment from you rather than the org (since the org was always clear that contributions were FOSS). No matter the scenario, the responsibility unfortunately lies with the contributor.
3. It's unlikely that a rule about e.g. using a uni email off campus out of hours for unrelated things is legally valid/enforcable. Apart from if you were misrepresenting the uni (e.g. using your account to lend weight or credence to your personal emails) it's unlikely that they can do anything. Most "rules" are either illegal or superfluous (as it turns out the law is surprisingly complete).

Sorry for going off on a tangent but I think I'm as interested in legal theory as you :)
