### 2020-03-05 tqdm/tqdm#887 @arunpersaud
Hi

thanks for asking, but I only did a one line change (more like a one
word change), so no need to add me as an author.

Arun
