### 2020-03-05 tqdm/tqdm#887 @mikekutzma
Please don't include me, my contribution doesn't warrant co-authorship.
