### 2020-03-05 tqdm/tqdm#887 @wwwiiilll
Don't include me, I don't consider my contribution to be substantial enough to be a co-author.
