### 2020-03-06 tqdm/tqdm#887 @robertzk
yes, I agree to be a co-author
