### 2020-03-06 tqdm/tqdm#887 @maxnordlund
Thanks for asking, but I don't think a typo fix qualifies. Please don't add me to the paper.
