### 2020-03-06 29903b0
more auth updates
