### 2020-03-06 tqdm/tqdm#887 @MPagel
>
>
> @MPagel
...
>
> Sorry for going off on a tangent but I think I'm as interested in legal theory as you :)

Your take is probably correct.  In my particular case, there's 0% chance they would find my open source software participation problematic, as the unit I work with is all about promoting the 'sister' to OSS: open data.  They might have issue with the balance of time I spend on certain tasks, but they hired me for who I am.
