### 2020-03-06 tqdm/tqdm#887 @dwhswenson
Thanks, I'd be happy to be listed as an author. My contribution was small, but if you and the journal think it's enough, I'm pleased to be included.

Name: David W.H. Swenson
Affiliation: Univ Lyon, ENS de Lyon, Univ Claude Bernard, CNRS, Laboratoire de Physique and Centre Blaise Pascal

(That's really one affiliation, even though it looks like more. But it's what they tell me to put on papers.)

Personal aside: I'm actually quite proud of my (tiny) contribution to tqdm. It was exciting to contribute to a project with such wide impact. And the existing code was clear enough that it was as easy for me to fix the problem as it would have been to write an issue explaining it.
