+++ examples/paper.bib
@@ -74,13 +74,13 @@
  url="https://github.com/bstriner/keras-tqdm"
 }
 @misc{tqdm-results,
- year="2019",
+ year="2020",
  author="GitHub",
  title="{tqdm} Code Results",
  url="https://github.com/search?q=tqdm&type=Code"
 }
 @misc{tqdm-dependents,
- year="2019",
+ year="2020",
  author="GitHub",
  title="{tqdm} dependents",
  url="https://github.com/tqdm/tqdm/network/dependents"
@@ -92,28 +92,28 @@
  url="https://libraries.io/pypi/tqdm"
 }
 @misc{sourcerank,
- year="2019",
+ year="2020",
  author="Libraries.io",
  title="SourceRank Breakdown for {tqdm}",
  url="https://libraries.io/pypi/tqdm/sourcerank"
 }
 @misc{sourcerank-descending,
- year="2019",
+ year="2020",
  author="Libraries.io",
  title="Libraries - The Open Source Discovery Service",
  url="https://libraries.io/search?order=desc&platforms=PyPI&sort=rank"
 }
 @misc{stars,
- year="2019",
+ year="2020",
  author="GitHub",
  title="{tqdm} Stargazers",
  url="https://github.com/tqdm/tqdm/stargazers"
 }
 @misc{stars-hist,
- year="2019",
+ year="2020",
  author="{timqian}",
  title="Star history",
- url="https://timqian.com/star-history/#tqdm/tqdm"
+ url="https://star-history.t9t.io/#tqdm/tqdm"
 }
 @misc{trend-hist,
  year="2018",
@@ -124,7 +124,7 @@
  url="https://github.com/nihey/trending-history/blob/master/histories/Python.md"
 }
 @misc{hits,
- year="2019",
+ year="2020",
  title="{tqdm} hits",
  url="https://caspersci.uk.to/cgi-bin/hits.cgi?q=tqdm&a=plot",
  author="Casper O. {da Costa-Luis}"
