### 2020-03-08 tqdm/tqdm#887 @RunOrVeith
I contributed support for explicitly known infinite iterators.
If you think this warrants a contribution, I'd be happy to be included.

**Name**: Veith Röthlingshöfer
**ORCID**: 0000-0002-1824-3153

Is there a way to review/read the paper before publishing?
