### 2020-03-08 tqdm/tqdm#887 @danking
I decline co-authorship, but am grateful for this project and the thought you’ve put into this process. Godspeed!
