### 2020-03-09 tqdm/tqdm#887 @hugovk
> Is there a way to review/read the paper before publishing?

@RunOrVeith, the source is here:
* https://github.com/tqdm/tqdm/blob/master/examples/paper.md
* https://github.com/tqdm/tqdm/blob/master/examples/paper.bib

There's a proof here:
* https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-469883384
