### 2020-03-09 7798732
Update Arabic spelling
- closes #907
