+++ examples/paper.bib
@@ -61,7 +61,7 @@
  url="https://stackoverflow.com/questions/3857052/why-is-printing-to-stdout-so-slow-can-it-be-sped-up"
 }
 @misc{pypi-downloads,
- year="2019",
+ year="2020",
  author="{Python Packaging Authority ({PyPA})}",
  publisher="Python Software Foundation",
  title="Analyzing {PyPI} package downloads -- Python Packaging User Guide",
@@ -86,7 +86,7 @@
  url="https://github.com/tqdm/tqdm/network/dependents"
 }
 @misc{lib-io,
- year="2019",
+ year="2020",
  author="Libraries.io",
  title="{tqdm} on {PyPI}",
  url="https://libraries.io/pypi/tqdm"
