### 2020-03-09 1896501
GroupBy.progress_aggregator() bug fix
