### 2020-03-09 fdbf7f6
Fixed Issue #896: Improved standards
