+++ tqdm/std.py
@@ -750,9 +750,9 @@ class tqdm(Comparable):
                     # on the first column/row to decide whether it can
                     # take a fast or slow code path; so stop when t.total==t.n
                     t.update(n=1 if not t.total or t.n < t.total else 0)
-                    if not isinstance(func,dict) and not isinstance(func,list):
+                    if not isinstance(func, dict) and not isinstance(func, list):
                         return func(*args, **kwargs)
-                    else : 
+                    else :
                         return func
 
                 # Apply the provided function (in **kwargs)
