### 2020-03-09 4c353d1
Fixed Issue #896: Further improved standards
