+++ tqdm/std.py
@@ -750,14 +750,14 @@ class tqdm(Comparable):
                     # on the first column/row to decide whether it can
                     # take a fast or slow code path; so stop when t.total==t.n
                     t.update(n=1 if not t.total or t.n < t.total else 0)
-                    if not isinstance(func, dict) and not isinstance(func, list):
-                        return func(*args, **kwargs)
-                    else :
+                    if isinstance(func, dict) or isinstance(func, list):
                         return func
+                    else:
+                        return func(*args, **kwargs)
 
                 # Apply the provided function (in **kwargs)
                 # on the df using our wrapper (which provides bar updating)
-                if isBuiltinFunction :
+                if isBuiltinFunction:
                     result = getattr(df, df_function)(wrapper, **kwargs)
                 else:
                     result = getattr(df, df_function)(wrapper, **kwargs)
