### 2020-03-10 1f87d00
Added tests for progress_aggregate()
