### 2020-03-10 579eaf5
Fixed error for the new tests created
