### 2020-03-10 357cb62
Fixed error for the new tests created and imroved standards
