### 2020-03-10 0e05202
Fixed bug with test for progress_aggregate
