### 2020-03-10 f484324
Fixed bug with test for progress_aggregate
