### 2020-03-10 36ea84e
Fixed bug with test for progress_aggregate in group
