### 2020-03-10 bc55c75
Added test for dictionary object in aggregate
