+++ tqdm/tests/tests_pandas.py
@@ -161,6 +161,11 @@ def test_pandas_groupby_apply():
         res4 = df.groupby("a").aggregate([min, max, sum])
         assert res3.equals(res4)
 
+        # aggregate on list
+        res5 = df.groupby("a").progress_aggregate({'b': sum})
+        res6 = df.groupby("a").aggregate({'b': sum})
+        assert res5.equals(res6)
+
         our_file.seek(0)
 
         # don't expect final output since no `leave` and
