### 2020-03-21 4b9e898
Add environment variable TQDM_DYNAMIC_NCOLS
Mitigate https://github.com/tqdm/tqdm/issues/370
