### 2020-03-24 deb30a9
more auth updates
