### 2020-03-24 d7542e9
fix wrapped file comparison
- fixes #727
- closes #730
