### 2020-03-24 caf6dbb
performance/optimisation and slight tidy
