### 2020-03-25 23ce11d
minor docstrings
