### 2020-03-28 2918e32
initial `nrows`
- fixes #918
