### 2020-03-28 338aae0
re-add and deprecate `utils._environ_cols_wrapper`
