### 2020-03-28 2472e1b
actually use nrows
