### 2020-03-28 35538da
dynamically update nrows
- also fix potential use of dynamic ncols
