### 2020-03-28 d2fa326
fix `pos == nrows` and `clear()` logic
