### 2020-03-28 5aa9225
fix tests & make robust
