### 2020-03-29 9a6fbaf
exception safety
- fixes (partially) #548
