### 2020-03-29 0229cc6
fix import error => deprecation
- fixes #927
