### 2020-03-29 981c52e
document `reset()` and `refresh()` in notebooks
- reason for no auto-`close()` upon `Exception`
