### 2020-03-29 db8776f
update all documentation
- sync README.rst => DEMO.ipynb
