+++ .meta/.readme.rst
@@ -21,7 +21,7 @@ iterable with ``tqdm(iterable)``, and you're done!
     for i in tqdm(range(10000)):
         ...
 
-``76%|████████████████████████████         | 7568/10000 [00:33<00:10, 229.00it/s]``
+``76%|████████████████████████        | 7568/10000 [00:33<00:10, 229.00it/s]``
 
 ``trange(N)`` can be also used as a convenient shortcut for
 ``tqdm(xrange(N))``.
@@ -36,9 +36,10 @@ It can also be executed as a module with pipes:
     $ seq 9999999 | tqdm --bytes | wc -l
     75.2MB [00:00, 217MB/s]
     9999999
-    $ 7z a -bd -r backup.7z docs/ | grep Compressing | \
-        tqdm --total $(find docs/ -type f | wc -l) --unit files >> backup.log
-    100%|███████████████████████████████▉| 8014/8014 [01:37<00:00, 82.29files/s]
+
+    $ tar -zcf - docs/ | tqdm --bytes --total `du -sb docs/ | cut -f1` \
+        > backup.tgz
+     32%|██████████▍                      | 8.89G/27.9G [00:42<01:31, 223MB/s]
 
 Overhead is low -- about 60ns per iteration (80ns with ``tqdm.gui``), and is
 unit tested against performance regression.
@@ -154,19 +155,21 @@ Wrap ``tqdm()`` around any iterable:
 .. code:: python
 
     from tqdm import tqdm
-    import time
+    from time import sleep
 
     text = ""
     for char in tqdm(["a", "b", "c", "d"]):
-        time.sleep(0.25)
+        sleep(0.25)
         text = text + char
 
 ``trange(i)`` is a special optimised instance of ``tqdm(range(i))``:
 
 .. code:: python
 
+    from tqdm import trange
+
     for i in trange(100):
-        time.sleep(0.01)
+        sleep(0.01)
 
 Instantiation outside of the loop allows for manual control over ``tqdm()``:
 
@@ -174,19 +177,19 @@ Instantiation outside of the loop allows for manual control over ``tqdm()``:
 
     pbar = tqdm(["a", "b", "c", "d"])
     for char in pbar:
-        time.sleep(0.25)
+        sleep(0.25)
         pbar.set_description("Processing %s" % char)
 
 Manual
 ~~~~~~
 
-Manual control on ``tqdm()`` updates by using a ``with`` statement:
+Manual control of ``tqdm()`` updates using a ``with`` statement:
 
 .. code:: python
 
     with tqdm(total=100) as pbar:
         for i in range(10):
-            time.sleep(0.1)
+            sleep(0.1)
             pbar.update(10)
 
 If the optional variable ``total`` (or an iterable with ``len()``) is
@@ -199,7 +202,7 @@ but in this case don't forget to ``del`` or ``close()`` at the end:
 
     pbar = tqdm(total=100)
     for i in range(10):
-        time.sleep(0.1)
+        sleep(0.1)
         pbar.update(10)
     pbar.close()
 
@@ -236,16 +239,36 @@ Note that the usual arguments for ``tqdm`` can also be specified.
 
     $ find . -name '*.py' -type f -exec cat \{} \; |
         tqdm --unit loc --unit_scale --total 857366 >> /dev/null
-    100%|███████████████████████████████████| 857K/857K [00:04<00:00, 246Kloc/s]
+    100%|█████████████████████████████████| 857K/857K [00:04<00:00, 246Kloc/s]
 
 Backing up a large directory?
 
 .. code:: sh
 
-    $ 7z a -bd -r backup.7z docs/ | grep Compressing |
[…]
