### 2020-03-30 8ca3df8
minor optimisation
