### 2020-03-30 4665954
minor nested `clear()` bugfix
