### 2020-03-30 ef97656
slight nested skip `clear()` optimisation
