### 2020-03-30 0663419
fix `display()` override properly
