### 2020-03-30 bee52fa
rework `decr_instances()`
Reduces screen flicker/blank space at the cost of ordering.

- don't move displaying (non-overflow) bars upwards
- move the first unfixed (`pos >= 0`) overflow (`pos >= nrows`) bar
  to the current `pos`

TODO: test
