### 2020-03-30 086a573
fix and update screen_shape tests
