### 2020-03-31 ed9ebed
fix tests, optimise `close() when `leave=False`
