### 2020-04-01 3e98f9b
add pre-commit helper
