+++ Makefile
@@ -15,6 +15,7 @@
 	testtimer
 	distclean
 	coverclean
+	pre-commit
 	prebuildclean
 	clean
 	toxclean
@@ -111,6 +112,11 @@ distclean:
 	@+make coverclean
 	@+make prebuildclean
 	@+make clean
+pre-commit:
+	# quick sanity checks
+	@make testsetup
+	flake8 -j 8 --count --statistics tqdm/ examples/
+	nosetests tqdm --ignore-files="tests_(perf|keras)\.py" -e "pandas|monitoring" -d
 prebuildclean:
 	@+python -c "import shutil; shutil.rmtree('build', True)"
 	@+python -c "import shutil; shutil.rmtree('dist', True)"
