### 2020-04-15 5ea18cb
notebook: add text/plain representation from ascii
- fixes #937
- TODO: fix refresh
