### 2020-04-15 de956bd
Output wrapper
TODO: add __repr__
