### 2020-04-15 323d85b
Output wrapper still only prints __repr__ once
