### 2020-04-20 e6a3c21
Feature: shorted description when console width is too small
