### 2020-04-24 6fc41a8
completion.sh framework
- fixes #858
