### 2020-04-24 e773193
completion.sh framework
- fixes #858
