### 2020-04-25 27e9338
base functionality
- core CLI options + extra CLI options
- remove options with no CLI support
