### 2020-04-25 4196859
minor tidy
