### 2020-04-25 3e549bf
minor efficiency
