### 2020-04-26 895ee5f
test bash autocomplete args
