### 2020-04-26 3becbb6
documentation
- integrate with code
- add to readme/docs etc.
- remove file
