### 2020-04-26 153098e
add --comppath test
