+++ .meta/mkcompletion.py
@@ -9,43 +9,52 @@ from os import path
 import re
 
 RE_OPT = re.compile(r'(\w+)  :', flags=re.M)
+RE_OPT_INPUT = re.compile(
+    r'(\w+)  : (?:str|int|float|chr|dict|tuple)', flags=re.M)
 
 
-def doc2opt(doc):
+def doc2opt(doc, all=False):
     """
     doc  : str, document to parse
+    all  : bool, [default: False] for only options requiring user input
     """
-    return ('--' + i for i in RE_OPT.findall(doc))
+    RE = RE_OPT if all else RE_OPT_INPUT
+    return ('--' + i for i in RE.findall(doc))
 
 
 # CLI options
 options = {'-h', '--help', '-v', '--version'}
+options_input = set()
 for doc in (tqdm.tqdm.__init__.__doc__, tqdm.cli.CLI_EXTRA_DOC):
-    options.update(doc2opt(doc))
+    options.update(doc2opt(doc, all=True))
+    options_input.update(doc2opt(doc, all=False))
 options.difference_update(
     '--' + i for i in ('name',) + tqdm.cli.UNSUPPORTED_OPTS)
+options_input &= options
+options_input -= {"--log"}  # manually dealt with
 src_dir = path.abspath(path.dirname(__file__))
 completion = u"""\
 #!/usr/bin/env bash
+_tqdm(){{
+  local cur prv
+  cur="${{COMP_WORDS[COMP_CWORD]}}"
+  prv="${{COMP_WORDS[COMP_CWORD - 1]}}"
 
-_tqdm()
-{{
-    local cur prv
-    cur="${{COMP_WORDS[COMP_CWORD]}}"
-    prv="${{COMP_WORDS[COMP_CWORD - 1]}}"
-
-    case ${{prv}} in
-        "--log")
-            COMPREPLY=($(compgen -W 'CRITICAL FATAL ERROR WARN WARNING INFO DEBUG NOTSET' -- ${{cur}}))
-            ;;
-        *)
-            COMPREPLY=($(compgen -W '{0}' -- ${{cur}}))
-            ;;
-    esac
+  case ${{prv}} in
+  {opts_manual})
+    # await user input
+    ;;
+  "--log")
+    COMPREPLY=($(compgen -W \
+      'CRITICAL FATAL ERROR WARN WARNING INFO DEBUG NOTSET' -- ${{cur}}))
+    ;;
+  *)
+    COMPREPLY=($(compgen -W '{opts}' -- ${{cur}}))
+    ;;
+  esac
 }}
-
 complete -F _tqdm tqdm
-""".format(' '.join(options))
+""".format(opts=' '.join(options), opts_manual='|'.join(options_input))
 
 if __name__ == "__main__":
     fncompletion = path.join(path.dirname(src_dir), 'tqdm', 'completion.sh')
