### 2020-04-26 b466398
tests
- test auto complete script
- remove redundant test
