### 2020-04-26 b3d32b1
docs
- updated CONTRIBUTING.md
- removed sample docs
