### 2020-04-26 cb2feea
docs
- updated CONTRIBUTING.md
- removed sample docs
