+++ .meta/docs.md
@@ -1,15 +1,0 @@
-#### Integrating new arguments with Bash autocomplete for CLI
-
-Creating new argument
- * Available through tqdm CLI
-   - Mention in docstring under `tqdm.tqdm.__init__.__doc__` or if *extra cli option* add to `tqdm.cli.CLI_EXTRA_DOC`
-   - Make sure to follow docstring formatting `argument<space><space>:<space>string`
-   - Add new argument to tests `tests_completion.py -> all_opts`
- * Not available through tqdm CLI
-   - Mention in docstrings
-   - Add to unsupported arguments `tqdm.cli.UNSUPPORTED_OPTS`
-
-**Note:** If new CLI available argument contains available options to select from, add the list of options for the argument under `.meta/mkcompletion.py`
-
-**Warning:** Bash autocomplete script is generated by parsing the docstrings. Make sure the format is correct and the test pass. Currently an additional `name` argument is present that has been manually removed.
-
