### 2020-04-26 a2fb4ce
merge tests, fix flake8
