### 2020-04-26 6be1880
add completion.sh to setup
- also sort entries for consistent re-generation
