+++ .meta/mkdocs.py
@@ -1,12 +1,13 @@
 from __future__ import print_function
+from io import open as io_open
 from os import path
 import sys
+from textwrap import dedent
+
 sys.path.insert(0, path.dirname(path.dirname(__file__)))  # NOQA
 import tqdm
 import tqdm.cli
-from textwrap import dedent
-from io import open as io_open
-from os import path
+
 
 HEAD_ARGS = """
 Parameters
