### 2020-04-26 2a318d9
add missing docstrings
