### 2020-04-26 0b77a0c
set default tqdm.disable from os.environ
Allow setting the tqdm.disable option from the environment variables. The default value of `False` is unchanged, as is the expected behavior with `None`.
