### 2020-04-27 #1277 @arfon
:wave: @casperdcl @lrq3000 - I realize times are challenging for lots of people right now but I wanted to check in to see if you think this conversation about authorship has converged?
