### 2020-04-27 #1277 @casperdcl
dropped off my radar but I think I've responded to all the authorship/affiliation requests in https://github.com/tqdm/tqdm/issues/887 with :+1: and corresponding checkboxes in https://github.com/tqdm/tqdm/pull/905, don't think I've missed anything, haven't heard anything so perhaps could just merge?
