### 2020-04-27 #1277 @lrq3000
Hello guys, yes sorry it's my fault, i still need to take a look just to
see if we didn't forget (or add without consent) anyone. I am behind my
schedule with all that's happening right now. I'll tabe a look on
Wednesday, but Casper did an amazing job so if he says it's fine to merge
then I trust him and would agree to the merge if you guys want it done now.

Thanks a lot Arfon, Casper, and everyone who was involved for your great
help.

Take care and stay safe.

Le lun. 27 avr. 2020 à 23:28, Casper da Costa-Luis <notifications@github.com>
a écrit :

> dropped off my radar but I think I've responded to all the
> authorship/affiliation requests in tqdm/tqdm#887
> <https://github.com/tqdm/tqdm/issues/887> with 👍 and corresponding
> checkboxes in tqdm/tqdm#905 <https://github.com/tqdm/tqdm/pull/905>,
> don't think I've missed anything, haven't heard anything so perhaps could
> just merge?
>
