### 2020-04-27 #1277 @casperdcl
argh just realised there's also https://github.com/tqdm/tqdm/pull/910 and https://github.com/tqdm/tqdm/pull/911
