### 2020-04-28 #1277 @lrq3000
Ok I've done the review, I have just made one comment that needs your attention @casperdcl and after it's fine on that side. I'll take a look at those 2 other PRs if I can do something. Thank you again for your great work!
