### 2020-04-28 673cf3b
Paper: fix spelling Jupyter HTML notebooks -> Jupyter Notebook
Signed-off-by: Stephen L. <lrq3000@gmail.com>
