### 2020-04-28 d4ba166
contrib.telegram: set bar width 10, unicode, monospace
