### 2020-04-29 8f56179
mixed-class same-stream nested workaround
- related #509
