### 2020-04-29 9fee632
safer exception printing
