### 2020-04-29 ffc6b13
add flush(), fix codacy
