### 2020-04-29 af2ec21
minor lock resiliency
