### 2020-05-02 8e139fc
Don't insert spurious newlines in DummyTqdmFile.
Only call `tqdm.write` (which inserts a newline) if
`DummyTqdmFile.write` is itself called with a string containing a
newline; otherwise buffer the rest of the string.

See changes in test_tqdm.py for the patterns this allows.
