### 2020-05-02 cecc8e0
mention snap tab copletion in readme
