### 2020-05-02 e00501a
add requests examples
- fixes #242
