### 2020-05-03 ee651fd
fix docs silly requests usage
