+++ README.rst
@@ -24,7 +24,7 @@ iterable with ``tqdm(iterable)``, and you're done!
 ``76%|████████████████████████        | 7568/10000 [00:33<00:10, 229.00it/s]``
 
 ``trange(N)`` can be also used as a convenient shortcut for
-``tqdm(xrange(N))``.
+``tqdm(range(N))``.
 
 |Screenshot|
     |Video| |Slides|
