### 2020-05-09 214ff86
xarray: use pandas()
- fixes #653
