### 2020-05-22 571f0ed
minor linting, sync readme
