### 2020-05-23 8b5b4bb
add contrib.discord: first Discord Bot!
