### 2020-05-23 78f86fc
contrib.discord,telegram abstract & cleanup MonoWorker
