### 2020-05-24 8d721b2
add contrib.bells
All the bells & whistles!

- related #614
- realted #950
