### 2020-06-02 #1277 @arfon
> Ok I've done the review, I have just made one comment that needs your attention @casperdcl and after it's fine on that side. I'll take a look at those 2 other PRs if I can do something. Thank you again for your great work!

:wave: @lrq3000 @casperdcl - any chance we can wrap this up this week?
