### 2020-06-03 133b218
keras: fix tensorflow==2.2.0 compatibility
- fixes #979
