+++ tqdm/keras.py
@@ -91,3 +91,12 @@ class TqdmCallback(keras.callbacks.Callback):
         if self.verbose:
             self.batch_bar.close()
         self.epoch_bar.close()
+
+    def _implements_train_batch_hooks(self):
+        return True
+
+    def _implements_test_batch_hooks(self):
+        return True
+
+    def _implements_predict_batch_hooks(self):
+        return True
