### 2020-06-06 e95cc6a
catch-all tqdm developers author
Seems to be an approach for other JOSS papers
