### 2020-06-06 aae8770
remove deprecated link
- fixes #981
