### 2020-06-06 34fb7f4
update order
as per `git fame -wMC` using

- https://github.com/casperdcl/git-fame/commit/274df71ec0705568bbd46cda8bb7f540007c3e82
- https://github.com/tqdm/tqdm/commit/1e0bb74db632c10643ff5c3e8ae699bb99118c38
