### 2020-06-07 #1277 @casperdcl
Ok I think we've reached an impasse. All issues have been resolved except a question of ordering names. The relevant comments are https://github.com/tqdm/tqdm/pull/905#issuecomment-640124258 and https://github.com/tqdm/tqdm/pull/905#issuecomment-640164461.

Essentially (correct me if I'm wrong):

- the Journal refuses to define & enforce a definition of paper authorship
- the contributors of the library have refused to agree on a definition of paper authorship
