### 2020-06-07 #1277 @lrq3000
Just a suggestion about this:

> the contributors of the library have refused to agree on a definition of paper authorship

With paper authorship, we are not working with "human material", so a formal definition may not always be possible. But if we can reach a consensus, that would be good enough IMO.

I added a reply with some additional metrics, I hope this may resolve or at least help the issue progress.
