### 2020-06-10 d2f18a5
v1 of scikit-learn tqdm integration roadmap. Talks about How, why, and where to start implementing the tqdm integration.
