### 2020-06-10 eb54b18
Added sklearn function to sklearn.py
