+++ dev_tqdm/sklearn.py
@@ -1,0 +1,3 @@
+@classmethod
+def sklearn(tclass, *targs, **tkwargs):
+    pass
