### 2020-06-10 73995a3
Added tqdm sklearn function to tqdm class
