### 2020-06-12 1c21957
Implemented "Option 1" of roadmap for 'cross_val_predict', 'cross_val_score', and 'cross_validate'; Successfully tested with KNeighborsClassifier
