### 2020-06-12 01501c0
Added some minor comments to sklearn.py
