### 2020-07-18 #1277 @arfon
@casperdcl @lrq3000 thank you for all of the work you've done to date to try and resolve this issue. I've been trying to keep up to date with progress over the past few months (e.g. in https://github.com/tqdm/tqdm/pull/905) and I have to say I'm not exactly sure how close this is to coming to a resolution.

It it now nearly six months since the authorship issue was raised here, and [my comment](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-578402029) stating how JOSS planned to handle this.

I recognize that the last few months have been truly exceptional circumstances for many, and as a result we have given this issue significantly more time to resolve itself than originally planned. That said, we absolutely must reach a conclusion _eventually_ and I believe this time has come.

I would like to see a resolution to this within the next two weeks (Saturday 1st August 2020) with explicit agreement from both @casperdcl and @lrq3000 that a mutually agreeable outcome has been reached. As [we stated before](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-578402029), if no acceptable resolution can be found then we will retract this paper on the 2nd August 2020 and the JOSS website will reflect this retraction (see this paper as an example: https://joss.theoj.org/papers/10.21105/joss.01338).
