### 2020-07-18 #1277 @lrq3000
Thank you very much @arfon and the Joss team for your understanding and for allowing us so much time to resolve this issue.

Ok I agree with the deadline and the conditions. I will discuss with @casperdcl directly.
