### 2020-07-18 #1277 @casperdcl
Thanks all for your input/responses.

I have been careful to avoid volunteering any personal opinions as much as possible. However, in the interest of clarity and to hopefully help understand my reasoning, these are my working assumptions (aka what I believe to be irrefutable facts):

### Everyone

- **users who have copyright over the codebase**: complex, many and various; a good guide is https://github.com/tqdm/tqdm/graphs/contributors or https://doi.org/10.5281/zenodo.595120. However, specifics are not concerning to end-users due to the FOSS licence(s) associated with the code. The fact that this is not a concern is evident by the millions of weekly downloads
- **original copyright holder and sponsor/funder of the initial version of the codebase pre-open sourcing**: as yet unnamed company which @noamraph worked/works for
- **users who have copyright over the original publication**: @casperdcl
- **users who have made valuable contributions to the original publication, but do not assert copyright (aka reviewers)**: @GregaVrbancic, @Benjamin-Lee
- **users who have copyright over the proposed revised publication**: @casperdcl
- **users who have made valuable contributions to the proposed revised publication, but do not assert copyright (aka revision reviewers)**: @hugovk, @lrq3000

### JOSS

- unlike most (if not all) other journals, JOSS does not enforce a strict definition of authorship (c.f. [IEEE's definition](https://journals.ieeeauthorcenter.ieee.org/become-an-ieee-journal-author/publishing-ethics/definition-of-authorship/))
- JOSS does not have a prominent warning about this surprising difference to advise submitters and readers of this unexpected difference from their assumptions. If would be both prudent and ethical to add such a warning at the very least to its submission guidelines
- true to its principles, JOSS has stated it does not want to make a decision about the authorship list of this proposed revised publication

### Summary

Based on the following facts:

- JOSS has stated it would like, in general, authors to agree on the author list
- JOSS has stated it would like, in this case, @casperdcl and @lrq3000 to agree on the author list
- JOSS has stated it would like, in general, Zenodo author lists (which are automatically determined code contributors) to at least temporarily match publication author lists
- Everyone has stated that they would accept any final decision made by JOSS even if they don't feel happy about it

it would seem highly likely that JOSS would consider the following outcome acceptable:

- **authors of the revised publication**: @casperdcl, @lrq3000
- **contributors** (listed in the revised publication either at the top or in the acknowledgements): everyone I've made the effort to add currently; i.e. all codebase contributors minus the ones who have opted out, vis. https://github.com/tqdm/tqdm/pull/905. Order determined by Zenodo, GitHub, `git-fame` or even popular sentiment (e.g. "mention @noamraph and his company first")
- **reviewers of the revised publication**: @GregaVrbancic, @Benjamin-Lee, @hugovk, @lrq3000
