### 2020-07-18 #1277 @lrq3000
I did not know of this new proposition before and it is a total change from the previous proposition which we have worked on up to now. I will comment more later.
