### 2020-07-18 #1277 @casperdcl
Ah this is not a "new proposal". It's firstly not "new" as it was actually the first suggestion, I think. Secondly it's not a "propsal," it is only my private thoughts.

The "proposal" is still simply https://github.com/tqdm/tqdm/pull/905 as-is
