### 2020-07-19 #1277 @arfon
Thanks for the quick responses @casperdcl and @lrq3000. For clarity (and for future readers) I would like to provide some additional thoughts/responses to what has been said:

@casperdcl said:

> unlike most (if not all) other journals, JOSS does not enforce a strict definition of authorship (c.f. IEEE's definition)

While I agree IEEE does have a formal definition, many journals are much less formal about this _because authorship is such a complex topic with many nuances_.

For software journals, the issue of authorship has the added complexity of having to accommodate contributions to the software. For JOSS and its short papers, essentially the _majority_ of the effort behind a contribution (paper + software) resides in the software [by design](https://www.arfon.org/announcing-the-journal-of-open-source-software).

JOSS does talk about authorship however. In our ethics guidelines (https://joss.theoj.org/about#ethics) we say:

> - Author lists must be correct and complete. All listed authors must have made a contribution to the work, and all significant contributors should be included in the author list.

...and in our reviewer docs (https://joss.readthedocs.io/en/latest/review_criteria.html#authorship) we state:

> As discussed in the submission guidelines for authors, authorship is a complex topic with different practices in different communities. Ultimately, the authors themselves are responsible for deciding which contributions are sufficient for co-authorship, although JOSS policy is that purely financial contributions are not considered sufficient.

I should point out that over the past ~4 years, JOSS has published nearly 1000 papers and this is _the only one_ where authors have struggled to reach an amicable solution. This to me suggests the issue resides with this submission rather than some deficiency in our guidelines.

> JOSS has stated it would like, in this case, @casperdcl and @lrq3000 to agree on the author list.

I should clarify that I suggested this because my personal read of https://github.com/tqdm/tqdm/pull/905 was that @lrq3000 had been doing a good job of representing the group opinion over there. Specifically I believe the suggestion made in https://github.com/tqdm/tqdm/pull/905#issuecomment-599143376 by @lrq3000 to have the first three authors be: Casper da Costa-Luis, Stephen Karl Larroque, and Noam Yorav-Raphael seems very reasonable.

To remove any ambiguity here, given @noamraph's comment in https://github.com/tqdm/tqdm/pull/905#issuecomment-640164461, I revise my previous comment to be that we require approval of @casperdcl, @lrq3000, and @noamraph for any modified author list.

With a slight modification to @casperdcl's suggestion above I am providing the following suggested outcome:

- Authors of the revised publication: @casperdcl, @lrq3000, @noamraph
- Contributors (listed in the revised publication either at the top or in the acknowledgements): everyone @casperdcl made the effort to add currently; i.e. all codebase contributors minus the ones who have opted out, vis. tqdm/tqdm#905. Order determined by Zenodo, GitHub, git-fame or even popular sentiment.
- Reviewers of the revised publication: @GregaVrbancic, @Benjamin-Lee, @hugovk, @lrq3000

If you agree with this resolution please provide your revised list of authors in the form of a change to the `paper.md` by 2nd August 2020. As I stated before, we regret that if there isn’t an agreed resolution between @casperdcl, @lrq3000 & @noamraph by then, we will have to retract the paper.
