### 2020-07-20 #1277 @lrq3000
Dear @arfon,

Thank you very much for the clarification. I agree with the suggested outcome, and I think most contributors explicitly stated they would agree with this potential scenario with a reduced set of authors and others in contributors. We will continue working on https://github.com/tqdm/tqdm/pull/905 and reach a conclusion under the stated deadline.

I just have one question: you mention that contributors can be listed at the top, is there an example of this in another publication? I think it would be a very nice option for us.
