### 2020-07-20 #1277 @arfon
Great, thanks @lrq3000!

Yes, that draft paper looks good to me.
