### 2020-07-20 #1277 @lrq3000
@arfon Thank you, it looks good to me too!

About the list of contributors, you suggested "all codebase contributors minus the ones who have opted out". The contributors who have opted out did opt out of co-authorship, but not necessarily of acknowledgement (some who opted out even mentioned they would agree to/prefer an acknowledgement).

Do you think it would be OK if we (the 3 co-authors) decide to include all contributors in the end?
