### 2020-07-20 #1277 @arfon
> Do you think it would be OK if we (the 3 co-authors) decide to include all contributors in the end?

 Yes, I'm happy for you all to figure this out between yourselves.
