### 2020-08-02 #1277 @arfon
In May 2019, JOSS published this paper [tqdm: A Fast, Extensible Progress Meter for Python and CLI](https://joss.theoj.org/papers/10.21105/joss.01277).

While potential authorship issues were noted by [reviewers at the time](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-469208417) and [an attempt was made](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-469222023) by the submitting author to contact one of the other major contributors to the package, JOSS published the paper as a single-author paper.

Sometime later, the authorship (and JOSS handling) of this submission was [flagged as problematic](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-576505085) by a major contributor to the software and JOSS conducted an internal review, and [concluded](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-578402029) that there had been an ethical violation and offered the tqdm community one month to resolve the authorship issue.

Due to the challenges associated with the COVID-19 pandemic, JOSS did not hold the tqdm community to the original deadline of the 25th February 2020, instead setting a [new deadline on the 18th July 2020](https://github.com/openjournals/joss-reviews/issues/1277#issuecomment-660510272) for a resolution to this by 1st August 2020.

Unfortunately, even after [significant efforts from all parties](https://github.com/tqdm/tqdm/pull/905), the tqdm community have been unable to resolve this issue and come to an agreed set of authors for their paper.

As a result, with regret, JOSS is today retracting this submission due to a violation of its ethical standards by the submitting author of the original paper.

Arfon Smith (EiC JOSS) on behalf of the JOSS editorial team.
