### 2020-08-02 #1277 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
